% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{age2ageN}
\alias{age2ageN}
\title{repeat age lower bounds once for each single age}
\usage{
age2ageN(Age, AgeInt = NULL, OAG = FALSE)
}
\arguments{
\item{Age}{integer. Vector of lower age bounds.}

\item{AgeInt}{integer vector of age group widths.}

\item{OAG}{logical. Whether or not the final age group open. Default \code{FALSE}. See details}
}
\value{
integer vector of elements of \code{Age} repeated once for each implied single age.
}
\description{
This is a frequent grouping situation. For a given vector of lower age bounds, repeat each value N times, where N is the width of the corresponding age interval. Age intervals are in this case calculated from the original Age vector.
}
\details{
If \code{OAG = TRUE} then the last value is not repeated, otherwise the final age interval is assumed to be the same width as the penultimate. Here intervals do not need to be of uniform width.
}
\examples{
age1 <- seq(0,100,by=5)
(ageN1 <- age2ageN(age1, OAG = FALSE))
(ageN2 <- age2ageN(age1, OAG = TRUE))
}
