% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{age2int}
\alias{age2int}
\title{infer age class intervals from lower age bounds}
\usage{
age2int(Age, OAG = TRUE, OAvalue = NA)
}
\arguments{
\item{Age}{integer or numeric vector of age intervals}

\item{OAG}{logical. \code{TRUE} if the final age group is open.}

\item{OAvalue}{numeric or integer. The value to use for the final age interval if \code{OAG = TRUE}. Default \code{NA}.}
}
\value{
Age interval vector, of same length as \code{Age}.
}
\description{
Determine age calss intervals based on a vector of age class lower bounds.
}
\details{
If the final age group is open, it is given a value of \code{NA} by default, or else a user-determined value. 
If the final age group is closed, it is assumed to be equal to the next-lower interval. If the final age interval is 
known and not equal to the next lowest interval, specify \code{OAG = TRUE} and assign its value to \code{OAvalue}.
}
\examples{
# single age examples:
Age  <- 0:100
age2int(Age, OAG = TRUE, OAvalue = NA)
age2int(Age, OAG = TRUE, OAvalue = Inf)
age2int(Age, OAG = FALSE)

# and for abridged ages:
AgeA <- c(0,1,seq(5,85,by=5))
age2int(AgeA, OAG = TRUE, OAvalue = NA)
age2int(AgeA, OAG = TRUE, OAvalue = Inf)
age2int(AgeA, OAG = FALSE)
}
