% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AgefromAgeInt}
\alias{AgefromAgeInt}
\title{infer lower age class bounds from vector of age intervals}
\usage{
AgefromAgeInt(AgeInt, minA = 0)
}
\arguments{
\item{AgeInt}{vector of age intervals}

\item{minA}{what is the lowest age bound to start counting from (default 0)?}
}
\value{
integer vector of ages
}
\description{
a simple identity
}
\details{
It makes no difference whether the final age group is open or how that is coded,
as long as all lower age intervals are coercible to integer.
}
\examples{
int1 <- c(rep(1,100),NA)
int2 <- c(rep(1,100),"+")
int3 <- inferAgeIntAbr(vec = rep(1,20), OAG = TRUE)
AgefromAgeInt(int1)
AgefromAgeInt(int2) # character OK
AgefromAgeInt(int3)
}
