% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{AGEN}
\alias{AGEN}
\title{trim two age vectors to matching N-year age groups}
\usage{
AGEN(Age1, Age2, N = 5, consecutive = TRUE, ageMin = 0,
  ageMax = max(c(Age1, Age2)))
}
\arguments{
\item{Age1}{integer vector of first age groups (lower bounds)}

\item{Age2}{integer vector of second age groups (lower bounds)}

\item{N}{integer target age group interval (just one number)}

\item{consecutive}{logical default \code{TRUE}. Throw error if resulting age groups not consecutive?}

\item{ageMin}{integer optional lower age bound for output}

\item{ageMax}{integer optional upper age bound for output}
}
\value{
an integer vector of the N-year age groups present in both \code{Age1} and \code{Age2}
}
\description{
determining which N-year (e.g. 5-year) age groups two vectors have in common
is helpful for matching vector lengths, and questions of alignment. Used as a utility throughout.
}
\examples{
Age1 <- seq(0, 100, by = 5)
Age2 <- 0:80
AGEN(Age1, Age2, N = 5)
AGEN(Age1, Age2, N = 5, ageMax = 70)
}
