% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGESEX.R
\name{ageRatioScore}
\alias{ageRatioScore}
\title{Calculate the PAS age ratio score}
\usage{
ageRatioScore(
  Value,
  Age,
  ageMin = 0,
  ageMax = max(Age),
  method = "UN",
  OAG = TRUE
)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts in 5-year age groups.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{integer. The lowest age included in calculations. Default 0.}

\item{ageMax}{integer. The upper age bound used for calculations. Default \code{max(Age)}.}

\item{method}{character. Either \code{"UN"} (default), \code{"Zelnick"}, or \code{"Ramachandran"}}

\item{OAG}{logical. Whether or not the top age group is open. Default \code{TRUE}.}
}
\value{
The value of the index.
}
\description{
A single ratio is defined as the 100 times twice the size of an age group
to its neighboring two age groups. Under uniformity these would all be 100. The average
absolute deviation from 100 defines this index. This comes from the PAS spreadsheet called AGESEX
}
\details{
Age groups must be of equal intervals. Five year age groups are assumed.
It is also assumed that the final age group is open, unless \code{ageMax < max(Age)}.
Setting \code{OAG = FALSE} will override this and potentially include \code{max(Age)} in calculations.
The method argument determines the weighting of numerators and denominators, where the UN method puts
twice the numerator over the sum of the adjacent ages classes, Zelnich does thrice the
numerator over the sum of the whole range from the next lowest to the next highest age, and
Ramachandran does four times the numerator over the same sum, but with the central age
double-counted in the numerator.
}
\examples{
# data from PAS spreadsheet AGESEX.xlsx
Males   <- c(4677000,4135000,3825000,3647000,3247000,2802000,2409000,2212000,
		1786000,1505000,1390000,984000,745000,537000,346000,335000)
Females <- c(4544000,4042000,3735000,3647000,3309000,2793000,2353000,2112000,
		1691000,1409000,1241000,887000,697000,525000,348000,366000)
Age     <- seq(0, 75, by = 5)

ageRatioScore(Males, Age, ageMax = 75)    # 3.9, matches PAS
ageRatioScore(Females, Age, ageMax = 75)  # 3.66 matches PAS
ageRatioScore(Females, Age, ageMax = 75, method = "Ramachandran") # 1.8
ageRatioScore(Females, Age, ageMax = 75, method = "Zelnick")      # 2.4
}
\references{
\insertRef{accuracyun1952}{DemoTools}
\insertRef{ramachandran1967}{DemoTools}
\insertRef{PAS}{DemoTools}
}
