% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ages}
\alias{ages}
\title{Ages between 0 and 100 abridged in five year age groups}
\format{
A vector of length 21
}
\source{
Migration residual PAS spreadhseet
}
\usage{
ages
}
\description{
Ages between 0 and 100 abridged in five year age groups for unknown
country
}
\keyword{datasets}
