% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ages_asfr_single}
\alias{ages_asfr_single}
\title{Single ages between 15 and 49}
\format{
A vector of length 36
}
\source{
Migration residual PAS spreadsheet
}
\usage{
ages_asfr_single
}
\description{
Single ages between 15 and 49 for Sweden
}
\keyword{datasets}
