% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ages_fertility}
\alias{ages_fertility}
\title{Ages between 15 and 45 in five year age groups}
\format{
A vector of length 7
}
\source{
Migration residual PAS spreadhseet
}
\usage{
ages_fertility
}
\description{
Ages between 15 and 45 in five year age groups for unknown
country
}
\keyword{datasets}
