% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ages_single}
\alias{ages_single}
\title{Single ages between 0 and 100}
\format{
A vector of length 101
}
\source{
Migration residual PAS spreadsheet
}
\usage{
ages_single
}
\description{
Single ages between 0 and 100 for Sweden, 1999-2019.
}
\keyword{datasets}
