% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGESEX.R
\name{ageSexAccuracyDasGupta}
\alias{ageSexAccuracyDasGupta}
\title{Calculate Ajit Das Gupta's (1995) age sex accuracy index}
\usage{
ageSexAccuracyDasGupta(Males, Females, Age, ageMin = 0, ageMax = max(Age))
}
\arguments{
\item{Males}{numeric. A vector of population counts for males by age}

\item{Females}{numeric. A vector of population counts for females by age}

\item{Age}{integer. A vector of ages corresponding to the lower integer bound of the counts}

\item{ageMin}{integer. The lowest age included in calcs. Default 0}

\item{ageMax}{integer. The upper age bound used for calcs. Default \code{max(Age)}}
}
\description{
Given population counts in 5-year age groups for males and females, follow Ajit Das Gupta's steps
 to calculate a composite index of the quality of the age and sex structure for a given population.
}
\details{
It is assumed that the terminal age group is closed and not open. 
If the final element of \code{Males} and \code{Females} is the open age group,
then lower \code{ageMax} to remove it in calculations.
}
\examples{
# data from table for South West Africa (1946) given in reference
Males   <- c(2365, 2320, 1859, 1554, 1758, 1534, 1404, 1324,
              1118, 872, 795, 745, 743, 574)
Females <- c(2244, 2248, 1773, 1594, 1616, 1510, 1478, 1320,
              1085, 858, 768, 726, 533, 282)
Age     <- seq(0, 65, by = 5)
ageSexAccuracyDasGupta(Males, Females, Age)
# this method is not on the same scale as the others, so don't directly compare.
ageSexAccuracy(Males, Females, Age, method = "das gupta")
}
\references{
\insertRef{dasgupta1955}{DemoTools}
}
