% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGESEX.R
\name{AHI}
\alias{AHI}
\title{Kannisto's age heaping index}
\usage{
AHI(Value, Age, Agei = 90)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts or rates by single age.}

\item{Age}{integer. A vector of ages corresponding to the lower integer bound of the counts.}

\item{Agei}{integer. The age on which the index is centered.}
}
\value{
The value of the index.
}
\description{
This age heaping index is used for particular old ages, such as 90, 95, 100, 105, and so forth.
}
\details{
The index looks down two ages and up two ages, so the data must accommodate that range.
}
\examples{
Age <- 0:99
AHI(pop1m_pasex, Age, 90)
AHI(pop1m_pasex, Age, 95)
}
\references{
\insertRef{kannisto1999assessing}{DemoTools}
}
