% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asfr_mat}
\alias{asfr_mat}
\title{Age-specific fertility rates for age groups 15 to 45 between 1950
and 2045 for unknown country}
\format{
A matrix of dimensions 7 x 20
}
\source{
Migration residual PAS spreadhseet
}
\usage{
asfr_mat
}
\description{
Age-specific fertility rates for age groups 15 to 45 between 1950
and 2045 for unknown country
}
\keyword{datasets}
