% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asfr_mat_five}
\alias{asfr_mat_five}
\title{Age-specific fertility rates for age groups 15 to 45 between 1950 and 2045}
\format{
A matrix of dimensions 7 x 20
}
\source{
Migration residual PAS spreadsheet
}
\usage{
asfr_mat_five
}
\description{
Age-specific fertility rates for age groups 15 to 45 between 1950 and 2045
for unknown country
}
\keyword{datasets}
