% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asfr_mat_single}
\alias{asfr_mat_single}
\title{Age-specific fertility rates for single ages 15 to 49 between 1999 and 2018}
\format{
A matrix of dimensions 35 x 20
}
\source{
Migration residual PAS spreadhseet
}
\usage{
asfr_mat_single
}
\description{
Age-specific fertility rates for single ages 15 to 49 between 1999 and 2018
for Sweden
}
\keyword{datasets}
