% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{avg_adj}
\alias{avg_adj}
\title{calculate average of vector elements adjacent to and excluding the index element}
\usage{
avg_adj(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector the same length as \code{x}.
}
\description{
Calculate average of vector elements adjacent to and exlcuding the index element. 
For example, the second element of the result is the average of the first and third elements of the
input vector \code{x}. Used by \code{zigzag()}, and possibly useful elsewhere.
}
\details{
Tails are given a value of \code{NA}.
}
\examples{
x <- 1:10
all(avg_adj(x) == x, na.rm = TRUE)
}
