% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{ax.greville.mortpak}
\alias{ax.greville.mortpak}
\title{UN version of the Greville formula for a(x) from M(x)}
\usage{
ax.greville.mortpak(nMx, nqx, lx, IMR = NA, Sex = "m", region = "w",
  mod = TRUE, closeout = "mortpak")
}
\arguments{
\item{nMx}{numeric. Event exposure mortality rates.}

\item{nqx}{numeric.  Vector of age specific death probabilities in standard abridged age groups.}

\item{lx}{numeric.  Vector oflifetable survivorship in standard abridged age groups.}

\item{IMR}{numeric. Optional. {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}}, the death probability in first year of life, in case available separately.}

\item{Sex}{character. \code{"m"}, \code{"f"} or \code{"b"} for male, female, or both.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}

\item{mod}{logical. Whether or not to use Gerland's modification for ages 5-14. Default \code{TRUE}.}

\item{closeout}{character. Default \code{"mortpak"}.}
}
\value{
nax average contribution to exposure of those dying in the interval.
}
\description{
The UN a(x) formula uses Coale-Demeny for ages 0, and 1-4, values of 2.5 
for ages 5-9 and 10-14, and the Greville formula thereafter. In the original sources 
these are referred to as separation factors.
}
\details{
a(x) for age 0 and age group 1-4 are based on Coale-Demeny {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}}-based lookup tables.
An approximation to get from M(0) to {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}} for the sake of generating a(0) and 4a1 is used. 
By default the lifetbale is closed out using the Mortpak extrapolation. Otherwise the final a(x) value is closed out as if the final M(x) value were constant thereafter, 
which is a common lifetable closeout choice. Age groups must be standard abridged. No check on age groups is done.
}
\references{
\insertRef{greville1977short}{DemoTools}
\insertRef{un1982model}{DemoTools}
\insertRef{arriaga1994population}{DemoTools}
\insertRef{mortpak1988}{DemoTools}
}
