% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{axPAS}
\alias{axPAS}
\title{PAS a(x) rule of thumb}
\usage{
axPAS(nMx, AgeInt, IMR = NA, Sex = "M", region = "W", OAG = TRUE)
}
\arguments{
\item{IMR}{numeric. Optional. q0, the death probability in first year of life, in case available separately.}

\item{Sex}{character. \code{"m"}, \code{"f"} or \code{"b"} for male, female, or both.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}

\item{OAG}{logical (default \code{TRUE}). Is the last element of \code{nMx} the open age group?}

\item{M0}{numeric. Event exposure infant mortality rate}
}
\value{
nAx average contribution to exposure of those dying in the interval.
}
\description{
ax is calculated following the Coale-Demeny rules for ages 0 and 1-4, and assumes interval midpoints in higher ages. 
This is just a rule of thumb. This procedure is as found in the PAS spreadsheet \code{LTPOPDTH.XLS}.
}
\details{
If sex is given as both, \code{"b"}, then female 
values are taken for a0 and 4a1, per the PAS spreadsheet. If IMR is not given, the M0 is used in its 
stead for ages < 5. This function is not vectorized. ax closeout assumes constant mortality hazard in the open age group.
}
\references{
\insertRef{united1983manual}{DemoTools}
}
