% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGEHEAP.R
\name{Bachi}
\alias{Bachi}
\title{calculate Bachi's index of age heaping}
\usage{
Bachi(Value, Age, ageMin = 30, ageMax = 79, pasex = FALSE)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{integer. The lowest age included in calculations. Default 25.}

\item{ageMax}{integer. The upper age bound used for calculations. Default 65.}

\item{pasex}{logical. Whether or not reproduce the specific age weightings in the PASEX spreadsheet. Default \code{FALSE}.}
}
\value{
The value of the index.
}
\description{
Bachi's index involves applying the Whipple method repeatedly to determine the extent of preference for each final digit.
Similarly to Myers', it equals the sum of the positive deviations from 10 percent. It has a theoretical range from 0 to 90,
and 10 is the expected value for each digit.
Two Implementations: one following the PASEX spreadsheet SINGAGE (\code{pasex = TRUE}), with ages hard-coded, and another with flexible upper and lower age bounds, but that does not match the PASEX implementation.
}
\details{
\code{ageMax} is an inclusive upper bound, treated as interval. If you want ages
30 to 79, then give \code{ageMin = 30} and \code{ageMax = 79}, not 80. These are only heeded if \code{pasex = FALSE}.
}
\examples{
Age <- 0:99

Bachi(pop1m_pasex, Age, ageMin = 20, ageMax = 79, pasex = TRUE) # reproduces PASEX SINGAGE
Bachi(pop1m_pasex, Age, ageMin = 20, ageMax = 79) # default simpler
}
\references{
\insertRef{PAS}{DemoTools}
\insertRef{bachi1951tendency}{DemoTools}
\insertRef{shryock1973methods}{DemoTools}
}
