% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepop.R
\name{basepop_five}
\alias{basepop_five}
\title{BPA and BPE methods for adjusting age groups under 10}
\usage{
basepop_five(
  location = NULL,
  refDate = NULL,
  Age = NULL,
  Females_five = NULL,
  Males_five = NULL,
  nLxFemale = NULL,
  nLxMale = NULL,
  nLxDatesIn = NULL,
  AsfrMat = NULL,
  AsfrDatesIn = NULL,
  SRB = NULL,
  SRBDatesIn = NULL,
  radix = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{location}{UN Pop Division \code{LocName} or \code{LocID}}

\item{refDate}{The reference year for which the reported population pertain
(these are the population counts in \code{Females_five} and
\code{Males_five}). Can either be a decimal date, a \code{Date} class.
If \code{nLxDatesIn} or \code{AsfrDatesIn} are not supplied and the
corresponding \code{nLxFemale/Male}/\code{AsfrMat} is not supplied,
\code{refDate} must be at a minimum 1962.5. This is because we can only
fetch WPP data from 1955 onwards, and these minimum date is assumed to be
7.5 years before \code{refDate}, meaning 1955.}

\item{Age}{integer vector of lower bounds of abridged age groups given in \code{Females_five} and \code{Males_five}.}

\item{Females_five}{A named numeric vector with the population counts for
five-year abridged age groups for females in \code{refDate}. The names of the
vector should reflect the age groups. See the example section for some
examples.}

\item{Males_five}{A named numeric vector with the population counts for
five-year abridged age groups for males in \code{refDate}. The names of
the vector should reflect the age groups. See the example section for
some examples.}

\item{nLxFemale}{A numeric matrix. The female nLx function of two abridged life tables
with ages in the rows and time in columns. The earlier date should be at least
7.5 years before the reference date of the "reported" population. The later
date should be no earlier than one-half year before the reference date of
the "reported" population. If not provided, it's automatically downloaded if
\code{location}, \code{refDate} and the equivalent population counts
\verb{*_five} are provided.}

\item{nLxMale}{A numeric matrix. The male nLx function of two abridged life tables
with ages in the rows and time in columns. The dates which are represented
in the columns are assumed to be the same as \code{nLxDatesIn}. This
argument is only used when \code{female} is set to \code{FALSE} and
\code{Males_five} is provided. If \code{Males_five} is provided and
\code{female} set to \code{FALSE}, the \code{nLx} for males is
automatically downloaded for the dates in \code{nLxDatesIn}.}

\item{nLxDatesIn}{A vector of numeric years (for example, 1986). The dates
which pertain to the columns in \code{nLxFemale} and \code{nLxMale}. If not
provided, the function automatically determines two dates which are 8 years
before \code{refDate} and 0.5 years after \code{refDate}.}

\item{AsfrMat}{A numeric matrix. An age-period matrix of age specific
fertility rates with age in rows, time in columns. If not provided, the
function automatically downloads the ASFR matrix based on the dates in
\code{AsfrDatesIn}.}

\item{AsfrDatesIn}{A vector of numeric years (for example, 1986). These are
the dates which pertain to the columns in \code{AsfrMat}. If not provided,
the function automatically determines two dates which are 8 years before
\code{refDate} and 0.5 before \code{refDate}.}

\item{SRB}{A numeric. Sex ratio at birth (males / females). Default is set
to 1.046. Only a maximum of three values permitted.}

\item{SRBDatesIn}{A vector of numeric years (for example, 1986). Only a maximum
number of three dates allowed. These are
the dates which pertain to the values in \code{SRB}. If not provided,
the function automatically determines three dates which are 7.5 years,
2.5 and 0.5 years before \code{refDate}.}

\item{radix}{starting point to use in the adjustment of the three first age
groups. Default is NULL. If not provided, it is inferred based on the scale of age \verb{1L0}.}

\item{verbose}{when downloading new data, should the function print details
about the download at each step? Defaults to \code{TRUE}. We recommend the
user to set this to \code{TRUE} at all times because the function needs to
make decisions (such as picking the dates for the Asfr and nLx) that the user
should be aware of.}

\item{...}{Arguments passed to \verb{\link{interp}}. In particular, users
might be interested in changing the interpolation method for the \verb{nLx*}
matrices and the \code{Asfr} matrix. By default, it's linearly interpolated.}
}
\value{
\code{basepop_five} returns a list with the following elements:
*
\itemize{
\item \code{Females_adjusted} numeric vector of adjusted population counts for females. Age groups 0, 1-4, and 5-9 are adjusted, while ages 10 and higher are unchanged.
\item \code{Males_adjusted} numeric vector of adjusted population counts for males. Age groups 0, 1-4, and 5-9 are adjusted, while ages 10 and higher are unchanged.
\item \code{Females_five} numeric vector of female population counts given as input.
\item \code{Males_five} numeric vector of male population counts given as input.
\item \code{nLxf} numeric matrix of female \code{nLx}, abridged ages in rows and (potentially interpolated) time in columns. Potentially downloaded.
\item \code{nLxm} numeric matrix of male \code{nLx}, abridged ages in rows and (potentially interpolated) time in columns. Potentially downloaded.
\item \code{Asfr} numeric matrix of age specific fertility in 5-year age groups ages 15-19 until 45-49 in rows, and (potentially interpolated) time in columns. Potentially downloaded.
\item \code{Exposure_female} numeric matrix of approximated age-specific exposure in 5-year age groups ages 15-19 until 45-49 in rows, and (potentially interpolated) time in columns.
\item \code{Bt} births at three time points prior to census corresponding to the midpoints of the cohorts entering ages 0, 1-4, and 5-9.
\item \code{SRB} sex ratio at birth at three time points prior to census corresponding to the midpoints of the cohorts entering ages 0, 1-4, and 5-9. Potentially downloaded.
\item \code{Age} age groups of the input population counts.
}
}
\description{
Adjust population counts for the age groups 0 to 10
}
\details{
\code{basepop_five} and \code{basepop_single} can estimate both the BPA and
BPE methods. If the user specifies \code{SmoothedFemales}, both
\code{basepop_*} functions will return the BPA method.
If \code{SmoothedFemales} is left empty, both \code{basepop_*} functions will
adjust using the BPE method.

For \code{basepop_five}, adjusting the female population counts is the
default. For this, only the \code{location}, \code{refDate} and
\code{Females_five} are needed. All other arguments are downloaded
or set to sensible defaults. For adjusting the male population
counts, the user needs to specify the \code{Males_five} population
counts and set \code{female = FALSE}.

Currently, \code{basepop_five} works only with five year abridged age groups

The BPE method is used by default. To adjust the counts using
the BPA method, the user needs to provide the \code{SmoothedFemales}
argument. This is the female population counts passed through
a smoothing function such as \code{smooth_age_5}. See the examples
section for some examples.
}
\section{BPA}{


Description:

The method estimates a smoothed population ages 10 and over and adjusts
the population under age 10 using the smoothed population and estimates
of fertility and mortality.

Based on the smoothed female population counts, it rejuvenates the female
"reported" population 20 to 59 years of age for the two 5 year periods prior
to the census date to represent the female population in reproductive ages
5 and 10 years earlier.  Based on the rejuvenated population and fertility
and mortality levels, the method then estimates the male and female births
during the two 5 year periods prior to the census date. Next, it projects
the two 5-year birth cohorts to the census date. The projected figures
represent the adjusted population ages 0 to 4 years and 5 to 9 years
at the census date.

Advantages:

(1) The method adjusts under-10 population to be consistent with fertility
and mortality levels and adjusted adult female population.

Limitations:

(1) BPA assumes a linear change in fertility and mortality during the decade
prior to the reference year.

(2) The procedure ignores migration, which can lead to misleading results.
There are two issues. First, age groups 0-4 and 5-9 are subject to migration,
which would affect the comparability of estimated and reported populations
in the base year.  Second, the estimated size of age groups 0-4 and 5-9 are
calculated from numbers of women of reproductive age in the base year
rejuvenated to points in the past.  With migration, rejuvenated number of
women may exceed or be smaller than the number present, and giving
birth to children, in the decade prior to the base year.

(3) BPA’s smoothing calculations may mask unusual, but real, variations
in population age group size.  Smoothing irregularities in age structure
not attributable to age misreporting will distort estimated births and
survived children in the base year.

Assumptions:

(1) No significant international migration took place within the
reference periods for the population, mortality, and fertility input.

(2) The data input as the "reported" population is not affected by
underenumeration of persons in certain ages, nor by age misreporting.
}

\section{BPE}{


Description:

The method adjusts the population under age 10 using the reported population
ages 10 and above and estimates of fertility and mortality.

The method rejuvenates the reported female population 20 to 59 years of age
for the two 5 year periods prior to the census date to represent the female
population in reproductive ages 5 and 10 years earlier. Based on the
rejuvenated population and fertility and mortality levels, the method then
estimates the male and female births during the two 5 year periods prior to
the census date.  Next, it projects the two 5-year birth cohorts to the
census date.  The projected figures represent the adjusted population ages
0 to 4 years and 5 to 9 years at the census date.

Advantages:

(1) The method adjusts the under-10 population to be consistent with
fertility and mortality levels and adult female population.

Limitations:

(1) BPE assumes a linear change in fertility and mortality during the decade
prior to the reference year.

(2) The procedure ignores migration, which can lead to misleading results.
There are two issues.  First, age groups 0-4 and 5-9 are subject to
migration, which would affect the comparability of estimated and reported
populations in the base year.  Second, the estimated size of age groups
0-4 and 5-9 are calculated from numbers of women of reproductive age in
the base year rejuvenated to points in the past.  With migration, rejuvenated
number of women may exceed or be smaller than the number present, and
giving birth to children, in the decade prior to the base year.

(3) The method does not adjust for possible underenumeration and age
misreporting errors in the over-10 “reported” population. If the
reported population is subject to age-misreporting or age-sex-specific
underenumeration, the over-10 population should be smoothed or otherwise
corrected prior to use.

Assumptions:

(1) No significant international migration took place within the reference
periods for the population, mortality, and fertility input.

(2) The data input as the “reported” population is not affected by
underenumeration of persons in certain ages, nor by age misreporting.
}

\examples{

 \dontrun{
refDate  <- 1985
location <- "Brazil"
data("popF",  package = "wpp2024")
data("popM",  package = "wpp2024")

Age <- parse_number(unique(popF$age))

# so technically this function is for abridged data
# it does not work if first group is in 5 years
pop_female_counts <- popF \%>\%
 dplyr::filter(name == location) \%>\% 
 pull("1985")

pop_male_counts   <- popM \%>\%
 dplyr::filter(name == location) \%>\% 
 pull("1985")

names(pop_female_counts) <- names(pop_female_counts) <-  Age

# graduate
Females_five <- graduate_pclm(pop_female_counts, Age = Age) \%>\% 
 single2abridged()
Males_five   <- graduate_pclm(pop_male_counts, Age = Age) \%>\% 
 single2abridged()

Age <- names2age(Males_five)

# Automatically downloads the nLx, ASFR, and SRB data
bpe <- basepop_five(
 location     = location,
 refDate      = refDate,
 Females_five = Females_five,
 Males_five   = Males_five,
 Age          = Age,
 verbose      = TRUE
)

 }

}
\references{
\insertRef{arriaga1994population}{DemoTools}
\insertRef{PAS}{DemoTools}
}
