% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_AgeAbr_OAG}
\alias{calc_AgeAbr_OAG}
\title{produce abridged ages up to a given open age group}
\usage{
calc_AgeAbr_OAG(OAG = 80)
}
\arguments{
\item{OAG}{integer. Default 80}
}
\value{
integer vector of ages \code{c(0,1,5,10,15,...)}
}
\description{
Produce standard abrdiged age groups up to a specified open age group.
}
\details{
If the open age group is not evenly divisible by 5 then age classes only go up to its 5-year lower bound.
}
\examples{
calc_AgeAbr_OAG(80)
stopifnot(all(calc_AgeAbr_OAG(100) == calc_AgeAbr_OAG(102)))
}
