% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{calcAgeAbr}
\alias{calcAgeAbr}
\title{calculate which abridged age group single ages belong to}
\usage{
calcAgeAbr(Age)
}
\arguments{
\item{Age}{integer vector of single ages (lower bound)}
}
\value{
an integer vector of \code{length(Age)} indicating the abridged age group that each single age belongs to.
}
\description{
Assign single ages to 5-year abridged age groups. That means that age 0 is kept as a single age,
ages 1-4 are grouped together as abridged age 1, and thereafter 5-year age groups are used.
}
\details{
In the case that the single \code{Age} vector starts at some age higher than 4, 
this is just the same as \code{calcAgeN(Age,5,0)}.
}
\examples{
Age <- 0:70
calcAgeAbr(Age)
calcAgeN(Age,5,0)
}
