% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_heaping.R
\name{check_heaping_bachi}
\alias{check_heaping_bachi}
\title{calculate Bachi's index of age heaping}
\usage{
check_heaping_bachi(
  Value,
  Age,
  ageMin = 23,
  ageMax = NULL,
  method = "orig",
  details = FALSE,
  OAG = TRUE
)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{the minimum age used for estimation, default \code{23}}

\item{ageMax}{the maximum age used for estimation, default \code{77}}

\item{method}{either \code{"orig"} or \code{"pasex"}}

\item{details}{logical. Should a list of output be given}

\item{OAG}{logical. Is the highest age group open?}
}
\value{
The value of the index.
}
\description{
Bachi's index involves applying the Whipple method repeatedly to determine the extent of preference for each final digit. Similarly to Myers', it equals the sum of the positive deviations from 10 percent. It has a theoretical range from 0 to 90, and 10 is the expected value for each digit. There are two implementations.
}
\details{
\code{ageMax} is an inclusive upper bound, treated as interval. If you want ages
23 to 77, then give \code{ageMin = 23} and \code{ageMax = 77}. The \code{ageMin} is respected strictly, whereas \code{ageMax} is calculated flexibly- if you specify something too high then it is reduced and we warn accordingly, and if it's missing then we pick something reasonable. You can see the age ranges actually used by specifying \code{details = TRUE}.
}
\examples{
 check_heaping_bachi(pop1m_pasex, Age = 0:99, 
 ageMin = 23, ageMax = 77, method = "orig", OAG =FALSE)
 check_heaping_bachi(pop1m_ind, Age = 0:100, 
 ageMin = 23, ageMax = 77, method = "orig")
 # default simpler
 check_heaping_bachi(pop1m_pasex, Age = 0:99, 
 ageMin = 23, ageMax = 77, method = "pasex", OAG = FALSE)
# linear population, should give 0 for pasex
 check_heaping_bachi(seq(100000,1000,by=-1000),Age = 0:99, 
 ageMin = 23, ageMax = 77, method = "pasex", OAG = FALSE)
# fully concentrated, should give 90 
 pop_concetrated <- rep(c(100,rep(0,9)),10)
 check_heaping_bachi(pop_concetrated, Age = 0:99, 
 ageMin = 23, ageMax = 77, method = "pasex")
 check_heaping_bachi(pop_concetrated, Age = 0:99, 
 ageMin = 23, ageMax = 77, method = "orig")
}
\references{
\insertRef{PAS}{DemoTools}
\insertRef{bachi1951tendency}{DemoTools}
\insertRef{shryock1973methods}{DemoTools}
}
