% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_heaping.R
\name{check_heaping_myers}
\alias{check_heaping_myers}
\title{Calculate Myer's blended index of age heaping}
\usage{
check_heaping_myers(Value, Age, ageMin = 23, ageMax = 82, details = FALSE)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{integer. The lowest age included in calculations. Default 25.}

\item{ageMax}{integer. The upper age bound used for calculations. Default 82.}

\item{details}{logical. Default \code{FALSE}. If \code{TRUE}, then a list is returned with some relevant parameters.}
}
\value{
The value of the index.
}
\description{
Implementation following the PASEX spreadsheet SINGAGE. Myers' measures preferences for each of the ten possible digits as a blended index. It is based on the principle that in the absence of age heaping, the aggregate population of each age ending in one of the digits 0 to 9 should represent 10 percent of the total population.
}
\details{
\code{ageMax} is an inclusive upper bound, treated as interval. If you want ages
23 to 82, then give \code{ageMin = 23} and \code{ageMax = 82}, not 83 \code{ageMax} may be
internally rounded down if necessary so that \code{ageMax - ageMin + 1} is evenly divisible by 10. If in doubt, specify \code{details = TRUE}, and you can check which \code{ageMax} is actually used internally.
}
\examples{
Age <- 0:99
check_heaping_myers(pop1m_pasex, Age, 10, 89)

check_heaping_myers(pop1m_pasex, Age, 10, 90) * 2 #47.46, replicates SINGAGE males
}
\references{
\insertRef{myers1954accuracy}{DemoTools}
\insertRef{spoorenberg2007quality}{DemoTools}
\insertRef{PAS}{DemoTools}
}
