% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_heaping.R
\name{check_heaping_noumbissi}
\alias{check_heaping_noumbissi}
\title{calculate Noumbissi's digit heaping index}
\usage{
check_heaping_noumbissi(
  Value,
  Age,
  ageMin = 20 + digit,
  ageMax = ageMin + 30,
  digit = 0
)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{lower age used for estimation (inclusive). Default 20}

\item{ageMax}{upper bound used for estimation (inclusive). Default 50}

\item{digit}{integer. Any digit between 0 and 9. Default \code{c(0,5)}. Otherwise it needs to be a single digit.}
}
\value{
The value of the index.
}
\description{
Noumbissi's method improves on Whipple's method by extending its basic principle to all ten digits. It compares single terminal digit numerators to denominators consisting in 5-year age groups centered on the terminal digit of age in question. For example, 9, 19, 29, etc are all of digit 9.
}
\details{
\code{ageMin} and \code{ageMax} are applied to numerator ages, not denominators.
Denominators are always 5-year age groups centered on the digit in question,
and these therefore stretch into ages a bit higher or lower than the numerator ages. \code{ageMax} is an inclusive upper bound, treated as interval.
If you want ages 20 to 64, then give \code{ageMin = 20} and \code{ageMax = 64}, not 65.
}
\examples{
Age <- 0:99
check_heaping_noumbissi(pop1m_pasex, Age, digit = 0) # 2.32
check_heaping_noumbissi(pop1m_pasex, Age, digit = 1) # 0.55
check_heaping_noumbissi(pop1m_pasex, Age, digit = 2) # 0.73
check_heaping_noumbissi(pop1m_pasex, Age, digit = 3) # 0.76
check_heaping_noumbissi(pop1m_pasex, Age, digit = 4) # 0.49
check_heaping_noumbissi(pop1m_pasex, Age, digit = 5) # 2.08
check_heaping_noumbissi(pop1m_pasex, Age, digit = 6) # 0.66
check_heaping_noumbissi(pop1m_pasex, Age, digit = 7) # 1.08  7 looks good!
check_heaping_noumbissi(pop1m_pasex, Age, digit = 8) # 0.57
check_heaping_noumbissi(pop1m_pasex, Age, digit = 9) # 0.59
}
\references{
\insertRef{noumbissi1992indice}{DemoTools}
\insertRef{spoorenberg2007quality}{DemoTools}
}
