% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_heaping.R
\name{check_heaping_whipple}
\alias{check_heaping_whipple}
\alias{Whipple}
\title{Calculate Whipple's index of age heaping}
\usage{
check_heaping_whipple(Value, Age, ageMin = 25, ageMax = 65, digit = c(0, 5))

Whipple(Value, Age, ageMin = 25, ageMax = 65, digit = c(0, 5))
}
\arguments{
\item{Value}{numeric. A vector of demographic counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{integer. The lowest age included in calculations. Default 25.}

\item{ageMax}{integer. The upper age bound used for calculations. Default 65.}

\item{digit}{integer. Any digit between 0 and 9. Default \code{c(0,5)}. Otherwise it needs to be a single digit.}
}
\value{
The value of the index.
}
\description{
Implementation following the PASEX spreadsheet SINGAGE, with some extra options for more digit checking. Whipple's index is a summary measure of age heaping on ages ending usually in the digits 0 or 5 used to determine variability in the quality of age reporting between regions or countries and its evolution over time. It assumes a linear distribution of ages in each five-year age range. Digits refer to terminal age digits. For example, 9, 19, 29, etc are all of digit 9.
}
\details{
\code{ageMin} and \code{ageMax} refer to the bounds of the digits evaluated, which end up in the numerator. \code{ageMax} is inclusive. The denominator looks 7 ages lower and 2 ages higher than this range, so these ages must be available. You can get arbitrary W(i) indices by specifying other digits. Note you can only do pairs of digits they are 0 and 5. Otherwise just one digit at a time.
}
\examples{
 Age <- 0:99
# 2.34,replicates SINGAGE males
 (w05 <- check_heaping_whipple(pop1m_pasex, Age, 25, 60, digit = c(0,5))) 

 # implements formula from Roger et al. (1981, p. 148)
 (w0 <- check_heaping_whipple(pop1m_pasex, Age, 25, 60, digit = 0))
 (w5 <- check_heaping_whipple(pop1m_pasex, Age, 25, 60, digit = 5))

 # Whipple types
 check_heaping_whipple(pop1m_pasex, Age, 25, 60, digit = 3)
}
\references{
\insertRef{GDA1981IREDA}{DemoTools}
\insertRef{PAS}{DemoTools}
}
