% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_mortality.R
\name{coef.extra_mortality}
\alias{coef.extra_mortality}
\alias{coefficients.extra_mortality}
\title{coef function for extra_mortality method}
\usage{
\method{coef}{extra_mortality}(object, ...)
}
\arguments{
\item{object}{An object of the class \code{"extra_mortality"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
coef function for extra_mortality method
}
\keyword{internal}
