% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{convertSingleTo5Year}
\alias{convertSingleTo5Year}
\title{aggregates single year age groups into 5 year age groups}
\usage{
convertSingleTo5Year(Value)
}
\arguments{
\item{Value}{numeric vector of single year age groups.}
}
\description{
Creates five year age groups from single year ages.
}
\details{
Sums five year age intervals
}
\examples{
MalePop <- seq(1,100)
convertSingleTo5Year(MalePop)
}
