% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{DemoTools-renamed}
\alias{DemoTools-renamed}
\title{Renamed functions in DemoTools}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{monoCloseout}}: This function is now called \code{\link{graduate_mono_closeout}}.

\item \code{\link{splitMono}}: This function is now called \code{\link{graduate_mono}}.

\item \code{\link{grabill}}: This function is now called \code{\link{graduate_grabill}}.

\item \code{\link{sprague}}: This function is now called \code{\link{graduate_sprague}}.

\item \code{\link{beers}}: This function is now called \code{\link{graduate_beers}}.
 
 \item \code{\link{splitUniform}}: This function is now called \code{\link{graduate_uniform}}.
 
 \item \code{\link{aomegaMortalityLaws}}: This function is now called \code{\link{lt_ax_closeout}}.
}
}
