% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{detect_age_grouping}
\alias{detect_age_grouping}
\title{detect if age groups are single, abridged, five-year or other}
\usage{
detect_age_grouping(Age)
}
\arguments{
\item{Age}{integer vector of age class lower bounds}
}
\description{
This is can be used to standardize age group checking, used in flow control, method decision-making, input-output matching, and so on. Used firstly in ODAPbackend package, but generally useful and general enough to make it into DemoTools.
}
\examples{
detect_age_grouping(0:100)
detect_age_grouping(seq(0,100,by=5))
detect_age_grouping(c(0,1,seq(5,100,by=5)))
detect_age_grouping(c(0,1,2,3,4,seq(5,100,by=5)))
}
\seealso{
is_single, is_abridged
}
