% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_downloads.R
\name{downloadAsfr}
\alias{downloadAsfr}
\title{Extract ASFR estimates from WPP2019. Mainly an util function for other ones.}
\usage{
downloadAsfr(Asfrmat, location = NULL, AsfrDatesIn, method = "linear")
}
\arguments{
\item{Asfrmat}{numeric.}

\item{location}{vector. UN Pop Div \code{LocName} or \code{LocID}}

\item{AsfrDatesIn}{numeric. Vector of decimal dates.}

\item{method}{character. Could be \code{"linear"}, \code{"exponential"}, or \code{"power"}}
}
\value{
numeric matrix interpolated asfr
}
\description{
We extract \code{ASFRx} from \code{wpp2019}, interpolated to exact dates. Different methods availables.
A vector of countries can handle, but with an unique sex. Row names are not indicative of countries.
}
\examples{
# Total fertility ratio calculated from ASFRx downloaded from WPP19. 
# See `downloadnLx` for analogous examples on multiple countries or using codes instead of names. 
ASFR_Arg <- downloadAsfr(Asfrmat = NULL, location = "Argentina", AsfrDatesIn = 1950:2025)
\dontrun{
plot(1950:2025, as.numeric(colSums(ASFR_Arg))*5, xlab = "Year", ylab="TFR", ylim=c(1.5,4), t="l")
}
}
