% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dth5_zigzag}
\alias{dth5_zigzag}
\title{Deaths by 5 year age groups}
\format{
A numeric vector of length 20
}
\source{
The data comes from
  \url{http://demographer.com/white-papers/2013-removing-zigzag-from-age-data/}
}
\usage{
dth5_zigzag
}
\description{
Deaths by 5 year age groups in South Africa 1997, from Feeney Zigzag 2013
}
\keyword{datasets}
