% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e0_swe}
\alias{e0_swe}
\title{Swedish life expectancy at birth}
\format{
A data frame with:
\describe{
\item{Date}{Reference time.}
\item{Sex}{Male \code{m} and female \code{m}.}
\item{e0}{Life expectancy at birth.}
}
}
\source{
Human Mortality Database. Retrieved 2021-20-01, from \url{https://mortality.org}
}
\usage{
e0_swe
}
\description{
Life expectancy at birth by sex in tidy format for dates from 1960-07-01 to 2015-07-01 by 5 calendar years.
}
\keyword{datasets}
