% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fitted_logquad_m}
\alias{fitted_logquad_m}
\title{LogQuad model for MALES fitted for all HMD life tables}
\format{An object of class \code{wilmoth} of length 6.}
\source{
Human Mortality Database [Portal]. Retrieved 2019-11-28, from 
  \url{https://mortality.org}
}
\usage{
fitted_logquad_m
}
\description{
LogQuad model fitted using \code{\link[MortalityEstimate]{wilmoth}} for all MALE period life tables present in Human Mortality Database (\url{https://mortality.org}) in December 2019 (968 life tables). Object of class \code{wilmoth}.
}
\keyword{datasets}
