% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{geta0CD}
\alias{geta0CD}
\title{Coale-Demeny a0 as function of m0, region, and sex}
\usage{
geta0CD(M0, IMR = NA, Sex = "M", region = "W")
}
\arguments{
\item{M0}{numeric. Event exposure infant mortality rate}

\item{IMR}{numeric. Optional. q0, the death probability in first year of life, in case available separately.}

\item{Sex}{character. \code{"m"}, \code{"f"} or \code{"b"} for male, female, or both.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}
}
\value{
a0 the average age at death in the first year of life.
}
\description{
Coale-Demeny a0 from Manual X Table 164. This is just a rule of thumb. 
In this and some other older texts, a(0) is known as a 'separation factor'.
}
\details{
If sex is given as both, \code{"b"}, then female 
values are taken, per the PAS spreadsheet. This function is not vectorized. 
Formulas for North, South, and West are identical- only East is different. If \code{IMR} is not given,
then \code{M0} is converted to q0 using the following approximation:
\itemize{
\item{find \eqn{\alpha,\Beta} }{ Look up the appropriate slope and intercept for the given sex and region.}
\item{calc \eqn{a} as: }{ \deqn{a = M_0 * \Beta} }
\item{calc \eqn{b} as: }{ \deqn{b =  1 + M_0 * (1 - \alpha)} }
\item{approx \eqn{q_0} as: }{ \deqn{q_0 = \frac{ b - sqrt(b^2 - 4 * a * M_0) }{ 2 * a } } }
}
q0 is then taken as IMR, and applied directly to the Coale-Demeny piecewise linear formula.
}
\examples{
m0 <- seq(.001,.2,by=.001)
\dontrun{
plot(m0, sapply(m0, geta0CD, Sex = "m", region = "e"), ylab = "a0", 
		type = 'l', ylim = c(0,.36), lty = 2, col = "blue")
lines(m0,sapply(m0, geta0CD, Sex = "m", region = "w"), col = "blue")
lines(m0,sapply(m0, geta0CD, Sex = "f", region = "e"), lty = 2, col = "red")
lines(m0,sapply(m0, geta0CD, Sex = "f", region = "w"), col = "red")
text(.15, geta0CD(.15,Sex = "m", region = "e"),"males E",font=2)
text(.15, geta0CD(.15,Sex = "m", region = "w"),"males N,W,S",font=2)
text(.15, geta0CD(.15,Sex = "f", region = "e"),"females E",font=2)
text(.15, geta0CD(.15,Sex = "f", region = "w"),"females N,W,S",font=2)

# compare with the Preston approximation
# constants identical after m0 = .107
m0 <- seq(.001,.107,by=.001)
a0CDm0 <- sapply(m0, geta0CD, Sex = "m", region = "w")
a0CDpr <- 0.045 + 2.684 * m0
plot(m0, a0CDm0, type = 'l', lty = 2, col = "red")
lines(m0, a0CDpr)
plot(m0, (a0CDm0 - a0CDpr) * 365, main = "difference (days)", ylab = "days")
}
}
\references{
\insertRef{united1983manual}{DemoTools}
}
