% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{geta1_4CD}
\alias{geta1_4CD}
\title{Coale-Demeny 4a1 as function of M(0), region, and sex.}
\usage{
geta1_4CD(M0, IMR = NA, Sex = "m", region = "w")
}
\arguments{
\item{M0}{numeric. Event exposure infant mortality rate.}

\item{IMR}{numeric. Optional. {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}}, the death probability in first year of life, in case available separately.}

\item{Sex}{character. \code{"m"}, \code{"f"} or \code{"b"} for male, female, or both.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}
}
\value{
The average age at death between ages 1-4, 4a1.
}
\description{
Coale-Demeny 4a1. This is a rule of thumb. In this and some other older texts, 4a1 is known as a 'separation factor'. These coefficients were pulled from the PAS spreadsheets \code{LTPOPDTH.XLS} and not located in the original Manual X.
}
\details{
If sex is given as both, \code{"b"}, then female values are taken, per the PAS spreadsheet. This function is not vectorized. If \code{IMR} is not given, then \code{M0} is used in its stead.
}
\examples{
m0 <- seq(.001,.2,by =.001)
\dontrun{
plot(m0, sapply(m0, geta1_4CD, Sex = "m", region = "e"), ylab = "4a1",
		type = 'l', ylim = c(1,2), lty = 2, col = "blue")
lines(m0,sapply(m0, geta1_4CD, Sex = "m", region = "w"), col = "blue")
lines(m0,sapply(m0, geta1_4CD, Sex = "m", region = "n"), col = "blue", lty = "8383",lwd=2)
lines(m0,sapply(m0, geta1_4CD, Sex = "m", region = "s"), col = "blue", lty = "6464",lwd=2)
lines(m0,sapply(m0, geta1_4CD, Sex = "f", region = "e"), lty = 2, col = "red")
lines(m0,sapply(m0, geta1_4CD, Sex = "f", region = "w"), col = "red")
lines(m0,sapply(m0, geta1_4CD, Sex = "f", region = "n"), col = "red", lty = "8383",lwd=2)
lines(m0,sapply(m0, geta1_4CD, Sex = "f", region = "s"), col = "red", lty = "6464",lwd=2)

text(.05, geta1_4CD(.05,Sex = "m", region = "e"),"males E",font=2,pos=4)
text(.05, geta1_4CD(.05,Sex = "m", region = "w"),"males W",font=2,pos=4)
text(.05, geta1_4CD(.05,Sex = "m", region = "s"),"males S",font=2,pos=4)
text(.05, geta1_4CD(.05,Sex = "m", region = "n"),"males N",font=2,pos=4)

text(0, geta1_4CD(.01,Sex = "f", region = "e"),"females E",font=2,pos=4)
text(0, geta1_4CD(.01,Sex = "f", region = "w"),"females W",font=2,pos=4)
text(0, geta1_4CD(.01,Sex = "f", region = "s"),"females S",font=2,pos=4)
text(0, geta1_4CD(.01,Sex = "f", region = "n"),"females N",font=2,pos=4)

}
}
\references{
\insertRef{united1983manual}{DemoTools}
\insertRef{PAS}{DemoTools}
}
