% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getModelLifeTable}
\alias{getModelLifeTable}
\title{Wrapper to provide a single location to reference all model life tables.}
\usage{
getModelLifeTable(ModelName, Sex)
}
\arguments{
\item{ModelName}{string naming the life table to return. Can be "coale-demeny west".}

\item{Sex}{string indicating which sex should be returned. Can be either "M" or "F".}

\item{ModelName}{string naming the life table to return. Can be "coale-demeny west".}

\item{Sex}{string indicating which sex should be returned. Can be either "M" or "F".}
}
\value{
list of life tables

list of life tables
}
\description{
Still in the works.

Still in the works.
}
\details{
More model families can easily be added.

More model families can easily be added.
}
