% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduate.R
\name{graduate_beers_expand}
\alias{graduate_beers_expand}
\title{Create the Beers ordinary or modified coefficient matrix}
\usage{
graduate_beers_expand(Value, OAG = FALSE, method = "Mod")
}
\arguments{
\item{Value}{numeric vector, presumably counts in grouped ages}

\item{OAG}{logical, default = \code{TRUE} is the final age group open?}

\item{method}{character. Valid values are \code{"mod"} or \code{"ord"}. Default \code{"mod"}.}
}
\description{
The resulting coefficient matrix is based on the number of rows in \code{Value}
which must be in 5-year age groups (not abridged). The final row may be an open
or closed age group, as indicated by the \code{OAG} argument.
}
\details{
The \code{Value} vector is a placeholder in this case. This function is
a utility called by the Beers family of functions, where it is most convenient to just pass
in the same matrix being used in those calculations to determine the layout of the coefficient matrix.
}
\examples{
coefsOA     <- graduate_beers_expand(pop5_mat, OAG = TRUE, method = "mod")
coefsclosed <- graduate_beers_expand(pop5_mat, OAG = FALSE, method = "mod")
dim(graduate_beers_expand(pop5_mat, TRUE))
dim(graduate_beers_expand(pop5_mat, FALSE))
coefso      <- graduate_beers_expand(pop5_mat, OAG = TRUE, method = "ord")

# how to use (under the hood in beers()

}
\references{
\insertRef{beers1945modified}{DemoTools}
\insertRef{siegel2004methods}{DemoTools}
}
