% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grabill.R
\name{graduate_grabill_expand}
\alias{graduate_grabill_expand}
\title{Create the Grabill coefficient matrix.}
\usage{
graduate_grabill_expand(Value, Age, OAG = TRUE)
}
\arguments{
\item{Value}{numeric vector, presumably counts in grouped ages}

\item{Age}{integer vector, lower bounds of age groups}

\item{OAG}{logical, default = \code{TRUE} is the final age group open?}
}
\description{
The resulting coefficient matrix is based on the number of rows in \code{popmat} where we assume that each row of data is a 5-year age group and the final row is an open age group to be preserved as such.
}
\details{
The \code{Value} vector is really just a placeholder in this case. This function is a utility called by the Grabill family of functions, where it is most convenient to just pass in the same matrix being used in those calculations to determine the layout of the coefficient matrix. Note that these coefficients do not constrain population counts to their year totals. This function is called by \code{grabill()}, which ensures matching marginals by 1) blending boundary ages into the Sprague estimated population, and 2) a second constraint on the middle age groups to enforce matching sums.
}
\examples{
a5 <- as.integer(rownames(pop5_mat))
graduate_grabill_expand(pop5_mat[,1], Age = a5, OAG = TRUE)
graduate_grabill_expand(pop5_mat[,1], Age = a5, OAG = FALSE)
}
\references{
\insertRef{shryock1973methods}{DemoTools}
}
