% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduate.R
\name{graduate_pclm}
\alias{graduate_pclm}
\title{wrapper for \code{ungroup::pclm} method of splitting binned counts}
\usage{
graduate_pclm(Value, Age, OAnew = max(Age), ...)
}
\arguments{
\item{Value}{numeric vector, presumably counts in grouped ages}

\item{Age}{integer vector, lower bounds of age groups}

\item{OAnew}{integer, optional new open age, higher than \code{max(Age)}. See details.}

\item{...}{further arguments passed to \code{ungroup::pclm()}}
}
\description{
This is exactly the function \code{pclm()} from the \code{ungroup} package, except with arguments using standard \code{DemoTools} argument names.
}
\details{
The PCLM method can also be used to graduate rates using an offset if both numerators and denominators are available. In this case \code{Value} is the event count and \code{offset} is person years of exposure. The denominator must match the length of \code{Value} or else the length of the final single age result \code{length(min(Age):OAnew)}.  This method can be used to redistribute counts in the open age group if \code{OAnew} gives sufficient space. Likewise, it can give a rate extrapolation beyond the open age.
}
\examples{
a5  <- seq(0,100,by=5)
p5  <- pop5_mat[, 1]
p1  <- graduate_pclm(p5, Age = a5)
p1s <- sprague(p5,a5)
plot(a5, p5/5, type = "s",xlim=c(40,60),ylim=c(2000,4000))
lines(0:100, p1, lwd = 2, col = "red")
lines(0:100, p1s, lwd = 1, col = "blue",lty="8282")
}
\references{
\insertRef{pascariu2018ungroup}{DemoTools}
\insertRef{rizzi2015efficient}{DemoTools}
}
\seealso{
\code{\link[ungroup]{pclm}}
}
