% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduate.R
\name{graduate_sprague_expand}
\alias{graduate_sprague_expand}
\title{Create the Sprague coefficient matrix.}
\usage{
graduate_sprague_expand(Value, Age, OAG = TRUE)
}
\arguments{
\item{Value}{numeric vector, presumably counts in grouped ages}

\item{Age}{integer vector, lower bounds of age groups}

\item{OAG}{logical, default = \code{TRUE} is the final age group open?}
}
\description{
The resulting coefficient matrix is based on the number of rows in \code{popmat} where is assumed that each row of data is a 5-year age group. The final row may be an open or closed age group, as indicated by the \code{OAG} argument.
}
\details{
The \code{popmat} matrix is really just a placeholder in this case. This function is a utility called by the Sprague family of functions, where it is most convenient to just pass in the same matrix being used in those calculations to determine the layout of the coefficient matrix.
}
\examples{
a5           <- as.integer(rownames(pop5_mat))
coefsOA      <- graduate_sprague_expand(pop5_mat[,1], Age = a5, OAG = TRUE)
coefsclosed  <- graduate_sprague_expand(pop5_mat[,1], Age = a5, OAG = FALSE)
dim(coefsOA)
dim(coefsclosed)
}
\references{
\insertRef{sprague1880explanation}{DemoTools}
\insertRef{shryock1973methods}{DemoTools}
\insertRef{siegel2004methods}{DemoTools}
}
