% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupOAG}
\alias{groupOAG}
\title{Group down to a new open age class}
\usage{
groupOAG(Value, Age, OAnew)
}
\arguments{
\item{Value}{numeric vector of counts}

\item{Age}{integer vector of age classes}

\item{OAnew}{the desired open age group}
}
\value{
Value vector potentially of reduced length up to OAG
}
\description{
This simple utility lowers the open age group. It only returns the input value vector, not the age vector.
}
