% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{int2age}
\alias{int2age}
\title{Infer lower age bounds from age class intervals.}
\usage{
int2age(AgeInt, minA = 0)
}
\arguments{
\item{AgeInt}{integer or numeric. Vector of age intervals.}

\item{minA}{integer. The lowest age, default 0.}
}
\value{
Age vector of same length as \code{AgeInt}.
}
\description{
Determine lower bounds of age classes based on a vector of age intervals and a starting age.
}
\examples{
AgeInt <- c(1,4,rep(5,17))
int2age(AgeInt)
}
