% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{int2ageN}
\alias{int2ageN}
\title{repeat age lower bounds once for each single age}
\usage{
int2ageN(AgeInt, OAG)
}
\arguments{
\item{AgeInt}{integer or numeric. Vector of age intervals.}

\item{OAG}{logical. Whether or not the final age group open. Default \code{FALSE}. See details}
}
\value{
integer vector of elements of \code{Age} repeated once for each implied single age.
}
\description{
This is a frequent grouping situation. For a given vector of lower age bounds, as implied by \code{AgeInt}, repeat each value N times, where N is the width of the corresponding age interval. Age intervals are in this case given, and age lower bound values are calculated from \code{AgeInt} vector.
}
\details{
If {OAG = TRUE} then the last value is given just once, irrespective of the final value of \code{AgeInt}, otherwise the final age interval is repeated \code{AgeInt[length(AgeInt)]} times. Here intervals do not need to be of uniform width.
}
\examples{
int5 <- rep(5,21)
(ageN1 <- int2ageN(int5, OAG = FALSE))
(ageN2 <- int2ageN(int5, OAG = TRUE))
\dontshow{
stopifnot(length(ageN1) == 105)
stopifnot(length(ageN2) == 101)
}
}
