% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_coh.R
\name{interp_coh_bare}
\alias{interp_coh_bare}
\title{component-free intercensalcohort interpolation}
\usage{
interp_coh_bare(c1, c2, date1, date2, age1, age2, ...)
}
\arguments{
\item{c1}{numeric vector. The second (right) census in single age groups}

\item{date1}{reference date of c1`. Either a Date class object or an unambiguous character string in the format "YYYY-MM-DD".}

\item{date2}{reference date of c2`. Either a Date class object or an unambiguous character string in the format "YYYY-MM-DD".}

\item{age1}{integer vector. single ages of \code{c1}}

\item{age2}{integer vector. single ages of \code{c2}}
}
\description{
Cohorts between two censuses are interpolated flexibly using linear, exponential, or power rules. The lower and upper intercensal triangles are filled using within-age interpolation. This function is experimental and still in development.
}
\seealso{
interp
}
