% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_lc_lim.R
\name{interp_lc_lim}
\alias{interp_lc_lim}
\title{Lee-Carter method with limited data.}
\usage{
interp_lc_lim(
  Males = NULL,
  Females = NULL,
  type = "m",
  dates_in = as.numeric(colnames(Males)),
  Age = as.numeric(rownames(Males)),
  dates_out = dates_in,
  dates_e0 = NULL,
  e0_Males = NULL,
  e0_Females = NULL,
  prev_divergence = FALSE,
  OAnew = max(Age),
  OAG = TRUE,
  extrapLaw = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{Males}{numeric. Matrix of abridged age classes by year.}

\item{Females}{numeric. Matrix of abridged age classes by year.}

\item{type}{character. Either rates \code{"m"}, conditionated probabilities \code{"q"} or survival function \code{"l"} included in matrices}

\item{dates_in}{numeric. Vector of observed decimal years for input matrix. Same number of columns as input matrices.}

\item{Age}{numeric. Vector with inferior limit of abridged age classes with data. Same number of rows than input matrix.}

\item{dates_out}{numeric. Vector of decimal years to interpolate or extrapolate.}

\item{dates_e0}{numeric. Vector of decimal years where \code{"e_0"} should be fitted when apply method.}

\item{e0_Males}{numeric. Vector of life expectancy by year to be fitted. Same length than \code{"dates_e0"}.}

\item{e0_Females}{numeric. Vector of life expectancy by year to be fitted. Same length than \code{"dates_e0"}.}

\item{prev_divergence}{logical. Whether or not prevent divergence and sex crossover. Default \code{FALSE.}}

\item{OAG}{logical. Whether or not the last element of \code{nMx} (or \code{nqx} or \code{lx}) is an open age group. Default \code{TRUE.}}

\item{extrapLaw}{character. If extrapolating, which parametric mortality law should be invoked? Options include  \code{"Coale-Guo"} or other functions in \code{"lt_abridged"} function.}

\item{...}{Arguments passed to \verb{\link{lt_abridged}}.}
}
\value{
Lifetable in a data.frame with columns
\itemize{
\item{Year}{integer. Years included in dates_out},
\item{Age}{integer. Lower bound of abridged age class},
\item{AgeInt}{integer. Age class widths.}
\item{nMx}{numeric. Age-specific central death rates.}
\item{nAx}{numeric. Average time spent in interval by those deceased in interval. }
\item{nqx}{numeric. Age-specific conditional death probabilities.}
\item{lx}{numeric. Lifetable survivorship}
\item{ndx}{numeric. Lifetable deaths distribution.}
\item{nLx}{numeric. Lifetable exposure.}
\item{Sx}{numeric. Survivor ratios in uniform 5-year age groups.}
\item{Tx}{numeric. Lifetable total years left to live above age x.}
\item{ex}{numeric. Age-specific remaining life expectancy.}
}
}
\description{
Given a matrix of age (rows) by year (cols) for each sex, this function interpolate/extrapolate life tables
using the method for limited data suggested by  Li et. al (2004) (at least three observed years).
}
\details{
Based on spreedsheet "Li_2018_Limited_Lee-Carter-v4.xlsm" from UN. Only useful for abridged ages.
The main options are the use of non-divergent method for sex coherency (Li & Lee, 2005),
and the possibility of fitting \code{"k"} to replicate \code{"e_0"} at some given years.
Coale-Guo method allows to smooth rates at older ages.
}
\note{
Draft Version
}
\references{
\insertRef{Li2005}{DemoTools}
\insertRef{Li2004}{DemoTools}
}
