% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_lc_lim.R
\name{interp_lc_lim_abk_m}
\alias{interp_lc_lim_abk_m}
\title{wrapper fun for \code{"interp_lc_lim_estimate"} function}
\usage{
interp_lc_lim_abk_m(k, ax, bx)
}
\arguments{
\item{k}{numeric. k parameter from LC model.}

\item{ax}{numeric. Vector (same length of age) of parameters from LC model.}

\item{bx}{numeric. Vector (same length of age) of parameters from LC model.}
}
\description{
wrapper fun to estimate rates from LC parameters
}
