% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_lc_lim_group.R
\name{interp_lc_lim_group}
\alias{interp_lc_lim_group}
\title{Lee-Carter method with limited data for groups.}
\usage{
interp_lc_lim_group(
  input = NULL,
  dates_out = NULL,
  Single = FALSE,
  input_e0 = NULL,
  prev_divergence = FALSE,
  weights = NULL,
  OAG = TRUE,
  verbose = TRUE,
  SVD = FALSE,
  ...
)
}
\arguments{
\item{input}{data.frame. Columns: id, Date, Sex, Age, nMx (opt), nqx (opt), lx (opt).
The first column (id) cn be a numeric index or charcter vector identifying each group.}

\item{dates_out}{numeric. Vector of decimal years to interpolate or extrapolate.}

\item{Single}{logical. Wheter or not the lifetable output is by single ages.}

\item{input_e0}{data.frame with cols: id, Date, Sex and \code{"e_0"}. This should be fitted when apply method.}

\item{prev_divergence}{logical. Whether or not prevent divergence and sex crossover between groups. Default \code{FALSE.}}

\item{weights}{list. For \code{prev_divergence} option. A double for each element of a list with names as \code{id} columns. Should sum up to 1. Default: same weight for each group.}

\item{OAG}{logical. Whether or not the last element of \code{nMx} (or \code{nqx} or \code{lx}) is an open age group. Default \code{TRUE.}}

\item{verbose}{logical. Default \code{FALSE}.}

\item{SVD}{logical. Use Singular Value Decomposition for estimate b and k or Maximum Likelihood Estimation. Default \code{FALSE} for Maximum Likelihood Estimation.}

\item{...}{Other arguments to be passed on to the \code{\link[DemoTools]{lt_abridged}} function.}
}
\value{
List with:
\itemize{
\item Lifetable in a data.frame with columns:
\itemize{
\item \code{Date} numeric. Dates included in dates_out,
\item \code{Sex} character. Male \code{"m"} or female \code{"f"},
\item \code{Age} integer. Lower bound of abridged age class,
\item `AgeInt`` integer. Age class widths.
\item \code{nMx} numeric. Age-specific central death rates.
\item \code{nAx} numeric. Average time spent in interval by those deceased in interval.
\item \code{nqx} numeric. Age-specific conditional death probabilities.
\item \code{lx} numeric. Lifetable survivorship
\item \code{ndx} numeric. Lifetable deaths distribution.
\item \code{nLx} numeric. Lifetable exposure.
\item \code{Sx} numeric. Survivor ratios in uniform 5-year age groups.
\item \code{Tx} numeric. Lifetable total years left to live above age x.
\item \code{ex} numeric. Age-specific remaining life expectancy.
\item List with parameters estimated for each group:
\item \code{kt} numeric time vector. Time trend in mortality level.
\item \code{ax} numeric age vector. Average time of \verb{log(m_\{x,t\})}.
\item \code{bx} numeric age vector. Pattern of change in response to \code{kt}.
}
}
}
\description{
Given a data frame with groups (country, region, sex), dates, sex and mortality data by age (rates, conditionated probabilities of death
or survival function), this function interpolate/extrapolate life tables
using the method for limited data suggested by  Li et. al (2004) (at least three observed years).
}
\details{
Based on spreedsheet "Li_2018_Limited_Lee-Carter-v4.xlsm" from UN.
Useful for abridged or single ages, and allows output in both formats also.
One option is the use of non-divergent method for sex coherency (Li & Lee, 2005).
The other is the possibility of fitting \code{"k"} to replicate \code{"e_0"} at some given dates.
\code{id} column in \code{input} argument works for separate between groups. In case only one population/sex is given,
is recommended to give some group name to \code{id}, if not the function will try to infer the case.
}
\note{
Draft Version
}
\examples{
# mortality rates from Sweden, for specific dates. Each sex a group.
data("mA_swe")
mA_swe$id = c(rep("A",nrow(mA_swe)/2),
             rep("B",nrow(mA_swe)/2))

# needs mortality rates in this dates: 
dates_out <- as.Date(paste0(seq(1948,2018,5),"-07-01"))

# apply LC with limited data to extrap/interpolate
lc_lim_data <- interp_lc_lim_group(input = mA_swe, dates_out = dates_out)

\dontrun{
lc_lim_data[["lt_hat"]] \%>\% ggplot(aes(Age,nMx,col=factor(round(Date,1)))) +
  geom_step() + scale_color_viridis_d() + 
  scale_y_log10() + theme_classic() + facet_wrap(~Sex)
}

# avoid cross-over between groups
lc_lim_data <- interp_lc_lim_group(input = mA_swe, dates_out = dates_out,
                                    prev_divergence = TRUE, weights=list(A=.4,B=.6))

\dontrun{
lc_lim_data[["lt_hat"]] \%>\% ggplot(aes(Age,nMx,col=factor(round(Date,1)))) +
  geom_step() + scale_color_viridis_d() + 
  scale_y_log10() + theme_classic() + facet_wrap(~id)
}
}
\references{
\insertRef{Li2005}{DemoTools}
\insertRef{Li2004}{DemoTools}
}
\seealso{
\code{\link[DemoTools]{lt_abridged}}
}
