% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_lc_lim.R
\name{interp_lc_lim_kt_min}
\alias{interp_lc_lim_kt_min}
\title{Optimize k}
\usage{
interp_lc_lim_kt_min(k, ax, bx, age, sex, e0_target, ...)
}
\arguments{
\item{k}{numeric. k parameter from LC model.}

\item{ax}{numeric. Vector (same length of age) of parameters from LC model.}

\item{bx}{numeric. Vector (same length of age) of parameters from LC model.}

\item{age}{numeric.}

\item{sex}{numeric.}

\item{e0_target}{numeric.}
}
\description{
Optimize estimated k from LC with limited data model,
for fitting given e_0 at same dates
}
\details{
Given LC parameters at some date, change a bit k for replicate already know e_0 values.
This is useful to give some sort of flexibility, and not follow strictly linear model implied in LC model,
but taking advantage of estimated structure (ax) and change by age (bx) for some trustable period.
}
