% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRDID.R
\name{IRD}
\alias{IRD}
\title{index of relative difference}
\usage{
IRD(pop1, pop2)
}
\arguments{
\item{pop1}{numeric vector of counts from population 1}

\item{pop2}{numeric vector of counts from population 2}
}
\value{
an index value ranging from 0 to infinity.
}
\description{
Calculate the relative percent difference between two population structures. A returned 
value of zero means that the two population have identical structure. A value of 100 means
that the populations have no overlap at all (not likely for populations structured only by age).
}
\details{
We assume that the populations are ordered in the same way and 
that said ordering happens before this function is called. We only check here that the vectors
are of the same length. The input arguments could indeed be popualtions structured on multiple
variables (more than just age), as long as they are ordered in the same way. It is advised to lower 
the open age group for this method because each age has the same weight.
}
\examples{
pop1 <- c(7.38,14.16,14.79,17.36,15.11,10.14,8.50,7.28,5.28)
pop2 <- c(6.48,12.27,15.25,15.10,14.66,10.80,8.95,9.28,7.21)
IRD(pop1, pop2)     # 6.7 reproduces table 7.20 of Siegel & Swanson (2004)
IRD(pop1,pop1)      # identical pops  = 0
IRD(pop1, pop2 * 2) # only structure matters
pop3      <- pop1
pop3[1:7] <- 0
IRD(pop1, pop3)     # theoretical max > 100 
}
