% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{is.abridged}
\alias{is.abridged}
\title{detect if a vector of lower age bounds is plausibly of abridged ages}
\usage{
is.abridged(Age)
}
\arguments{
\item{Age}{integer vector of lower age bounds}
}
\value{
logical \code{TRUE} if abridged, \code{FALSE} otherwise.
}
\description{
a logical utility that checks if a given vector is of the lower bounds of abridged age groups or not.
}
\examples{
# as expected, TRUE
is.abridged(c(0,1,5,10,15,20,25))
# standard 5, not abridged, FALSE
is.abridged(c(0,5,10,15,20,25))
# plausible, TRUE
is.abridged(c(1,5,10,15,20,25))
# plausible, TRUE
is.abridged(c(5,10,15,20,25))
# 10 year age group not abridged, FALSE
is.abridged(c(0,1,5,10,15,25))
}
