% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.single}
\alias{is.single}
\title{logical checking of whether age classes appear single}
\usage{
is.single(Age)
}
\arguments{
\item{Age}{integer vector of age classes}
}
\value{
logical \code{TRUE} if detected as single ages, \code{FALSE} otherwise
}
\description{
check whether a vector of ages consists in single ages. This
makes sense sometimes when age intervals are not explicitly given.
}
\details{
In cases where ages are indeed single, but repeated, this will still return \code{FALSE}. 
Therefore make sure that the age vector given refers to a single year of a single population, etc.
}
\examples{
Age <- 0:99
Age2 <- c(0:10,0:10)
Age3 <- seq(0,80,by=5)
Age4 <- seq(0,10,by=.5)
is.single(Age)  # TRUE
is.single(Age2) # FALSE repeated, can't tell.
is.single(Age3) # FALSE not single ages
is.single(Age4) # FALSE not single ages
}
