% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{is_age_sequential}
\alias{is_age_sequential}
\title{is \code{Age} sorted sequentially?}
\usage{
is_age_sequential(Age)
}
\arguments{
\item{Age}{integer vector of single ages (lower bound)}
}
\value{
logical. Is the \code{Age} vector sorted sequentially?
}
\description{
Check if \code{Age} is sorted sequentially.
}
\details{
This does not check for redundancy.
}
