% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isLeapYear}
\alias{isLeapYear}
\title{determine whether a year is a leap year.}
\usage{
isLeapYear(Year)
}
\arguments{
\item{Year}{integer of year to query}
}
\value{
logical is the Year a leap year or not
}
\description{
In order to remove \code{lubridate} dependency, we self-detect leap years and adjust February accordingly. Code inherited from HMD.
}
\author{
Carl Boe
}
