% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGIVE.R
\name{loess_smth1}
\alias{loess_smth1}
\title{wrapper to loess}
\usage{
loess_smth1(Value, Age, OAG = TRUE, ...)
}
\arguments{
\item{Value}{numeric vector of counts in single, abridged, or 5-year age groups.}

\item{Age}{integer vector of ages corresponding to the lower integer bound of the counts.}

\item{OAG}{logical. Whether or not the top age group is open. Default \code{TRUE}.}

\item{...}{optional arguments passed to \code{stats::loess()}}
}
\description{
smooth data over age, preserving the open age group if necessary. This is a simple wrapper to \code{stats::loess()} but using standard demographic arguments.
}
\details{
The total sum of \code{Value} is preserved in the output. One can control smoothness using the \code{spar} argument of \code{stats::loess()}. See \code{\link[stats]{loess}} for more details.
}
\examples{
\dontrun{
Age <- 0:99
plot(Age,pop1m_pasex)
lines(Age, loess_smth1(pop1m_pasex, Age, FALSE))
}
}
\seealso{
\code{\link[stats]{loess}}
}
