% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_quad_augm.R
\name{logquad_augmented}
\alias{logquad_augmented}
\title{Augmented logquad}
\usage{
logquad_augmented(coeffs, k, q0_5, q60_15, Sex = "b", Age, ...)
}
\arguments{
\item{coeffs}{data.frame. Columns \code{a(x)}, \code{b(x)}, \code{c(x)} and \code{v(x)} from fitted logquad model. See \code{fitted_logquad_b}.}

\item{k}{numeric. Adult mortality related value from log-quad estimatation based on one or two input parameters. See \code{lt_model_lq}.}

\item{q0_5}{numeric. Probability of death from born to age 5.}

\item{q60_15}{numeric. Probability of death from age 60 to age 75.}

\item{Sex}{character. Either male \code{"m"}, female \code{"f"}, or both \code{"b"}.}

\item{Age}{integer. Abridged lower bound ages. Same length than rows in \code{coeffs}.}

\item{...}{Other arguments to be passed on to the \code{lt_abridged} function.}
}
\value{
life table as in \code{lt_abridged} function.
}
\description{
Adjust rates in oldest ages that comes from a HMD model, using an external estimate of 15q60 (Li, 2014). As an example see\code{\link[DemoTools]{HMD_old_logquad}}.
}
\details{
Parameter \code{a(x)} is augmented based on an external estimate of 15q60.
}
\references{
See \href{https://www.un.org/development/desa/pd/content/estimating-life-tables-developing-countries}{Li (2014)}.
}
