% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{lt_a_pas}
\alias{lt_a_pas}
\alias{axPAS}
\title{PAS a(x) rule of thumb.}
\usage{
lt_a_pas(nMx, AgeInt, IMR = NA, Sex = "m", region = "w",
  OAG = TRUE)

axPAS(nMx, AgeInt, IMR = NA, Sex = "m", region = "w", OAG = TRUE)
}
\arguments{
\item{nMx}{numeric. Event exposure mortality rates.}

\item{AgeInt}{integer. Vector of age interval widths.}

\item{IMR}{numeric. Optional. {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}}, the death probability in first year of life, in case available separately.}

\item{Sex}{character. \code{"m"}, \code{"f"} or \code{"b"} for male, female, or both.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}

\item{OAG}{logical. Whether or not the last element of \code{nMx} is the open age group Default \code{TRUE}.}
}
\value{
nax average contribution to exposure of those dying in the interval.
}
\description{
a(x) is calculated following the Coale-Demeny rules for ages 0 and 1-4, and assumes interval midpoints in higher ages.
This is just a rule of thumb. This procedure is as found in the PAS spreadsheet \code{LTPOPDTH.XLS}.
}
\details{
If sex is given as both, \code{"b"}, then female values are taken for a(0) and 4a1, per the PAS spreadsheet. If IMR is not given, the M(0) is used to estimate a(x) for ages < 5. This function is not vectorized. a(x) closeout assumes constant mortality hazard in the open age group. One safeguard is different from PAS: If assuming the interval midpoint implies a qx greater than 1, then we derive a(x) for the interval by assuming midpoint a(x) for each single age within the interval along with a constant death rate.
}
\examples{
Exposures <- c(100958,466275,624134,559559,446736,370653,301862,249409,
		247473,223014,172260,149338,127242,105715,79614,53660,
		31021,16805,8000,4000,2000,1000)

Deaths <- c(8674,1592,618,411,755,1098,1100,1357,
		1335,3257,2200,4023,2167,4578,2956,4212,
		2887,2351,1500,900,500,300)
# lower age bounds
Age    <- c(0, 1, seq(5, 100, by = 5))
AgeInt <- c(diff(Age), NA)
nMx <- Deaths/Exposures
lt_a_pas(nMx = nMx,AgeInt = AgeInt,Sex = 'm',region = 'n',OAG = TRUE)
}
\references{
\insertRef{united1983manual}{DemoTools}
\insertRef{PAS}{DemoTools}
}
