% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_regroup_age.R
\name{lt_ambiguous}
\alias{lt_ambiguous}
\title{calculate an abidged or single age lifetable from abridged or sinlge age data}
\usage{
lt_ambiguous(
  nMx_or_nqx_or_lx = NULL,
  type = "m",
  Age = NULL,
  Sex = NULL,
  Single = FALSE,
  ...
)
}
\arguments{
\item{nMx_or_nqx_or_lx}{numeric vector of either \code{nMx}, \code{nqx}, or \code{lx}}

\item{type}{character, which variable is \code{x}?, either \code{"m"}, \code{"q"}, or \code{"l"}. Default \code{"m"}}

\item{Age}{integer vector of the lower age bounds of \code{x}}

\item{Sex}{character, \code{"m"}, \code{"f"}, or \code{"b"}.}

\item{Single}{logical, do we want output in single ages?}

\item{...}{optional arguments passed to \code{lt_abridged()} or \verb{lt_single*()}}
}
\description{
This is a wrapper around the other lifetable utilities. We start with either \code{nMx}, \code{nqx}, or \code{lx} in single or abridged ages, and returns a full lifetable in either single or abridged ages. All optional arguments of \code{lt_abridged()} or \verb{lt_single*()} can be passed in, for instance the \code{nax} assumptions or the extrapolation arguments.
}
