% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{lt_ax_closeout}
\alias{lt_ax_closeout}
\alias{aomegaMortalityLaws}
\title{Life expectancy in the open age group.}
\usage{
lt_ax_closeout(mx, Age, law = c("kannisto", "kannisto_makeham",
  "gompertz", "ggompertz", "makeham", "beard", "beard_makeham",
  "quadratic")[1], extrapFrom = max(Age), extrapFit = Age[Age >= 40],
  ...)

aomegaMortalityLaws(mx, Age, law = c("kannisto", "kannisto_makeham",
  "gompertz", "ggompertz", "makeham", "beard", "beard_makeham",
  "quadratic")[1], extrapFrom = max(Age), extrapFit = Age[Age >= 40],
  ...)
}
\arguments{
\item{mx}{Vector or matrix of age specific death-rates.}

\item{Age}{integer. A vector of ages of the lower integer bound of the age classes.}

\item{law}{The name of the mortality law/model to be used.
The following options are available: \itemize{
  \item{\code{"kannisto"}} -- The Kannisto model;
  \item{\code{"kannisto_makeham"}} -- The Kannisto-Makeham model;
  \item{\code{"gompertz"}} -- The Gompertz model;
  \item{\code{"ggompertz"}} -- The Gamma-Gompertz model;
  \item{\code{"makeham"}} -- The Makeham model;
  \item{\code{"beard"}} -- The Beard model;
  \item{\code{"beard_makeham"}} -- The Beard-Makeham model;
  \item{\code{"quadratic"}} -- The Quadratic model.
  }}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality.}

\item{extrapFit}{integer vector. Ages to include in model fitting. Defaults to all ages \code{>          =60}.}

\item{...}{Other arguments to be passed on to the
\code{\link[MortalityLaws]{MortalityLaw}} function.}
}
\value{
life expectancy in the open age group
}
\description{
Get an estimate of life expectancy in the open age group.
}
\details{
This method estimates life expectancy in the open age group by fitting one of several potential old-age parametric mortality models, extrapolating rates to age 130, then backing out the implied remaining life expectancy in the open age group. This function replaces \code{aomegaMORTPAK()}.
}
\examples{
nMx <- c(0.12846,0.02477,0.00603,0.0034,
 		0.00417,0.00513,0.00581,0.00645,0.00725,
 		0.00813,0.00913,0.01199,0.01647,
 		0.0256,0.04047,0.06624,0.10638,0.19611)
Age <- c(0,1,seq(5,80,by =5))


lt_ax_closeout(nMx,Age,"Kannisto")
lt_ax_closeout(nMx,Age,"Kannisto_Makeham")
lt_ax_closeout(nMx,Age,"Makeham")
lt_ax_closeout(nMx,Age,"Gompertz")
lt_ax_closeout(nMx,Age,"GGompertz")
lt_ax_closeout(nMx,Age,extrapLaw ="Beard")
lt_ax_closeout(nMx,Age,"Beard_Makeham")
lt_ax_closeout(nMx,Age,"Quadratic")
}
\seealso{
\code{\link[DemoTools]{extra_mortality}}
}
