% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_id.R
\name{lt_id_d_l}
\alias{lt_id_d_l}
\title{Derive survivorship from lifetable deaths}
\usage{
lt_id_d_l(ndx, radix = sum(ndx))
}
\arguments{
\item{ndx}{numeric.  Vector of age-specific lifetable deaths.}

\item{radix}{numeric.}
}
\value{
lx vector of lifetable survivorship
}
\description{
This lifetable identity is the same no matter what kind of lifetable is required. You can find it in any demography textbook.
}
\details{
The vector returned is the same length as \code{dx} and it sums to the lifetable radix. If the radix is one then this is the discrete deaths distribution.
}
\references{
\insertRef{preston2000demography}{DemoTools}
}
