% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_id.R
\name{lt_id_d_q}
\alias{lt_id_d_q}
\title{Derive death probabilities from lifetable deaths}
\usage{
lt_id_d_q(ndx)
}
\arguments{
\item{ndx}{numeric.  Vector of age-specific lifetable survivorship.}
}
\value{
nqx vector of lifetable death probabilities.
}
\description{
This lifetable identity is the same no matter what kind of lifetable is required.  You can find it in any demography textbook.
}
\details{
The vector returned is the same length as \code{dx}.
}
\references{
\insertRef{preston2000demography}{DemoTools}
}
