% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_id.R
\name{lt_id_Ll_S}
\alias{lt_id_Ll_S}
\title{Calculate survivor ratios}
\usage{
lt_id_Ll_S(nLx, lx, Age, AgeInt, N = c(5, 1))
}
\arguments{
\item{nLx}{numeric vector of lifetable exposure.}

\item{lx}{numeric. Vector of lifetable survivorship at abridged ages.}

\item{Age}{integer. A vector of ages of the lower integer bound of the age classes.}

\item{AgeInt}{integer. Vector of age class widths. Default \code{inferAgeIntAbr(Age = Age)}.}

\item{N}{integer, the age width for survivor ratios, either 5 or 1. Default 5.}
}
\description{
An extra lifetable column for use in projections, which require uniform time steps both both age and period. Intervals are either single age (\code{N=1}) or five-year ages (\code{N=5}). Input vectors are assumed to come from either single or standard abridged ages.
}
\details{
This function does not account for \code{nLx} having been pre-binned into uniform 5-year age widths, which will throw an error. Just leave them in abridged ages instead. Note that in the case of abridged ages, the interpretation for the first and second value don't follow the original abridged age intervals: the first value in the probability of surviving from birth into ages 0-4 in the first five years, and the second value is the probability of surviving from 0-4 to 5-9. This represents a slight misalignment with the rest of the lifetable, user beware.
}
