% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{lt_id_morq_a}
\alias{lt_id_morq_a}
\title{wrapper to invoke PAS or UN ax methods given qx or mx}
\usage{
lt_id_morq_a(
  nMx,
  nqx,
  axmethod = "pas",
  Age,
  AgeInt,
  a0rule = "ak",
  IMR = NA,
  Sex = "m",
  region,
  OAG = TRUE,
  mod = TRUE,
  SRB = SRB,
  extrapLaw = "kannisto",
  extrapFrom = max(Age),
  extrapFit = Age[Age >= 60],
  ...
)
}
\arguments{
\item{nMx}{numeric. Event exposure mortality rates.}

\item{nqx}{numeric.  Vector of age specific death probabilities in standard abridged age groups.}

\item{axmethod}{character. Either \code{"pas"} or \code{"un"}}

\item{Age}{integer. A vector of ages of the lower integer bound of the age classes.}

\item{AgeInt}{integer. Vector of age interval widths.}

\item{a0rule}{character. Either \code{"ak"} (default) or \code{"cd"}.}

\item{IMR}{numeric. Optional. {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}}, the death probability in first year of life, in case available separately.}

\item{Sex}{character. \code{"m"}, \code{"f"} or \code{"b"} for male, female, or both.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}

\item{OAG}{logical. Whether or not the last element of \code{nMx} is the open age group Default \code{TRUE}.}

\item{mod}{logical.  Whether or not to use Gerland's modification for ages 5-14. Default \code{TRUE}.}

\item{SRB}{numeric. The sex ratio at birth (boys/girls), default 1.05.}

\item{extrapLaw}{character. If extrapolating, which parametric mortality law should be invoked? Options include  \code{"Kannisto", "Kannisto_Makeham", "Makeham","Gompertz", "GGompertz", "Beard",    "Beard_Makeham", "Quadratic"}. Default \code{"Kannisto"}. See details.}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality.}

\item{extrapFit}{integer vector. Ages to include in model fitting. Defaults to all ages \code{>          =60}.}

\item{...}{Other arguments to be passed on to the
\code{\link[MortalityLaws]{MortalityLaw}} function.}
}
\value{
nax average contribution to exposure of those dying in the interval.
}
\description{
Given either mx or qx, call either the \code{lt_a_un()} or \code{lt_a_pas()} functions.
}
\references{
\insertRef{greville1977short}{DemoTools}
\insertRef{un1982model}{DemoTools}
\insertRef{arriaga1994population}{DemoTools}
\insertRef{mortpak1988}{DemoTools}
\insertRef{united1983manual}{DemoTools}
\insertRef{PAS}{DemoTools}
}
