% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{lt_id_morq_a_greville}
\alias{lt_id_morq_a_greville}
\alias{ax.greville.mortpak}
\title{UN version of the Greville formula for a(x) from M(x)}
\usage{
lt_id_morq_a_greville(
  nMx,
  nqx,
  lx,
  Age,
  AgeInt = age2int(Age, OAvalue = 5),
  IMR = NA,
  Sex = "m",
  region = "w",
  mod = TRUE,
  closeout = TRUE,
  law = c("kannisto", "kannisto_makeham", "makeham", "gompertz", "ggompertz", "beard",
    "beard_makeham", "quadratic"),
  extrapFrom = max(Age),
  extrapFit = Age[Age >= 60],
  ...
)

ax.greville.mortpak(
  nMx,
  nqx,
  lx,
  Age,
  AgeInt = age2int(Age, OAvalue = 5),
  IMR = NA,
  Sex = "m",
  region = "w",
  mod = TRUE,
  closeout = TRUE,
  law = c("kannisto", "kannisto_makeham", "makeham", "gompertz", "ggompertz", "beard",
    "beard_makeham", "quadratic"),
  extrapFrom = max(Age),
  extrapFit = Age[Age >= 60],
  ...
)
}
\arguments{
\item{nMx}{numeric. Event exposure mortality rates.}

\item{nqx}{numeric.  Vector of age specific death probabilities in standard abridged age groups.}

\item{lx}{numeric.  Vector of lifetable survivorship in standard abridged age groups.}

\item{Age}{integer. Vector of lower bounds of abridged age groups.}

\item{AgeInt}{integer. Vector of age group intervals.}

\item{IMR}{numeric. Optional. {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}}, the death probability in first year of life, in case available separately.}

\item{Sex}{character. \code{"m"}, \code{"f"} or \code{"b"} for male, female, or both.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}

\item{mod}{logical. Whether or not to use Gerland's modification for ages 5-14. Default \code{TRUE}.}

\item{closeout}{logical. Whether or not to estimate open age a(x) via extrapolation. Default \code{TRUE}.}

\item{law}{The name of the mortality law/model to be used.
The following options are available: \itemize{
  \item{\code{"kannisto"}} -- The Kannisto model;
  \item{\code{"kannisto_makeham"}} -- The Kannisto-Makeham model;
  \item{\code{"gompertz"}} -- The Gompertz model;
  \item{\code{"ggompertz"}} -- The Gamma-Gompertz model;
  \item{\code{"makeham"}} -- The Makeham model;
  \item{\code{"beard"}} -- The Beard model;
  \item{\code{"beard_makeham"}} -- The Beard-Makeham model;
  \item{\code{"quadratic"}} -- The Quadratic model.
  }}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality.}

\item{extrapFit}{integer vector. Ages to include in model fitting. Defaults to all ages \code{>          =60}.}

\item{...}{Other arguments to be passed on to the
\code{\link[MortalityLaws]{MortalityLaw}} function.}
}
\value{
nax average contribution to exposure of those dying in the interval.
}
\description{
The UN a(x) formula uses Coale-Demeny for ages 0, and 1-4, values of 2.5 for ages 5-9 and 10-14, and the Greville formula thereafter. In the original sources these are referred to as separation factors.
}
\details{
a(x) for age 0 and age group 1-4 are based on Coale-Demeny {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}}-based lookup tables. An approximation to get from M(0) to {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}} for the sake of generating a(0) and 4a1 is used. The final a(x) value is closed out using the \code{lt_a_closeout()} method (reciprocal and Mortpak methods are deprecated). Age groups must be standard abridged. No check on age groups is done.

There are different vectors one can specify for this method: ultimately it's either \code{nMx} or \code{nqx}, and the \code{nax} results will differ potentially quite a lot depending which you have on hand.
}
\examples{
#Example witn Mexican data from UN
nMx <- c(0.11621,0.02268,0.00409,0.00212,0.00295,0.00418,0.00509,0.00609,
0.00714,0.00808,0.00971,0.0125,0.0175,0.02551,0.03809,0.05595,0.08098,
0.15353,0.2557)

nqx <- c(0.10793,0.08554,0.02025,0.01053,0.01463,0.02071,0.02515,0.02999,
0.03507, 0.03958,0.04742,0.0606,0.08381,0.11992,0.17391,0.2454,0.33672,
0.54723,NA)

lx  <- c(100000,89207,81577,79924,79083,77925,76312,74393,72162,69631,66875,
63704,59843,54828,48253,39861,30079,19951,9033)

Age <- c(0,1,seq(5,85,by = 5))
AgeInt <- age2int(Age, OAvalue = 5)
# two quite different results depending whether you start with mx or qx
lt_id_morq_a_greville(nMx = nMx, Age = Age, AgeInt = AgeInt, Sex = 'f',region = 'w')
lt_id_morq_a_greville(nqx = nqx, Age = Age, AgeInt = AgeInt, Sex = 'f',region = 'w')
# same, qx comes from lx
lt_id_morq_a_greville(lx = lx, Age = Age, AgeInt = AgeInt, Sex = 'f',region = 'w')
# both qx and lx given, but lx not used for anything = same
lt_id_morq_a_greville(nqx = nqx, lx = lx, Age = Age, AgeInt = AgeInt, Sex = 'f',region = 'w')

# if both qx and mx given, then same as lt_id_qm_a identity,
# except young ages follow Coale-Demeny, and greville uses
# MortalityLaws closeout.
lt_id_morq_a_greville(nMx = nMx, nqx = nqx, Sex = 'f', Age = Age, AgeInt = AgeInt, region = 'w')-
		lt_id_qm_a(nqx,nMx,age2int(Age,TRUE,5))
# same (qx comes from lx)
lt_id_morq_a_greville(nMx = nMx, lx = lx, Sex = 'f', Age = Age, AgeInt = AgeInt, region = 'w')
}
\references{
\insertRef{greville1977short}{DemoTools}
\insertRef{un1982model}{DemoTools}
\insertRef{arriaga1994population}{DemoTools}
\insertRef{mortpak1988}{DemoTools}
}
\seealso{
\code{\link[DemoTools]{lt_a_closeout}}
}
