% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_id.R
\name{lt_id_qa_m}
\alias{lt_id_qa_m}
\alias{qxax2mx}
\title{Derive nMx from nqx and nax.}
\usage{
lt_id_qa_m(nqx, nax, AgeInt)

qxax2mx(nqx, nax, AgeInt)
}
\arguments{
\item{nqx}{numeric. Vector of age specific death probabilities.}

\item{nax}{numeric. Vector of average time spent in interval by those dying in interval.}

\item{AgeInt}{integer. Vector of age class widths.}
}
\value{
nMx vector of age specific death rates derived via identity.
}
\description{
This is the standard identity to derive nMx from nax and nqx.
}
\references{
\insertRef{preston2000demography}{DemoTools}
}
