% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_model_lq.R
\name{lt_model_lq}
\alias{lt_model_lq}
\title{Estimate Wilmoth Model Life Table}
\usage{
lt_model_lq(
  Sex,
  fitted_logquad = NULL,
  q0_5 = NULL,
  q0_1 = NULL,
  q15_45 = NULL,
  q15_35 = NULL,
  e0 = NULL,
  radix = 1e+05,
  tol = 1e-09,
  maxit = 200,
  axmethod = "pas",
  a0rule = "ak",
  IMR = NA,
  region = "w",
  mod = TRUE,
  SRB = 1.05
)
}
\arguments{
\item{Sex}{Choose the sex of the population. This choice defines the use
of a corresponding Log-Quadratic (\code{wilmoth})
model fitted for the whole Human Mortality Database (as of Dec 2019,
there are 968 life tables for each sex).
The following options are available: \itemize{
\item{\code{"b"}} -- Both sex;
\item{\code{"f"}} -- Females;
\item{\code{"m"}} -- Males.
}}

\item{fitted_logquad}{Optional, defaults to \code{NULL}. An object of class
\code{wilmoth}. If full HMD is not enough, one
can fit a Log-Quadratic (\url{https://github.com/mpascariu/MortalityEstimate}) model
based on any other collection  of life tables;}

\item{q0_5}{5q0. The probability that a new-born will die during the
subsequent 5 years;}

\item{q0_1}{1q0. The probability that a life aged 0 will die during the
following year;}

\item{q15_45}{45q15. The probability that a life aged 15 will die during
the subsequent 45 years;}

\item{q15_35}{35q15. The probability that a life aged 15 will die during
the subsequent 35 years;}

\item{e0}{Life expectancy at birth;}

\item{radix}{Life table radix. Default: 10^5;}

\item{tol}{Tolerance level for convergence. The tolerance level, is relevant
for case 7 and 8 (e0 and 45q15 or 35q15 are known);}

\item{maxit}{Maximum number of iterations allowed. Default: 100;}

\item{axmethod}{character. Either \code{"pas"} or \code{"un"}.}

\item{a0rule}{character. Either \code{"ak"} (default) or \code{"cd"}.}

\item{IMR}{numeric. Infant mortality rate \ifelse{html}{\out{q<sub>0}}{\eqn{q_0}}, in case available and \code{nqx} is not specified. Default \code{NA}.}

\item{region}{character. North, East, South, or West: \code{"n"}, \code{"e"}, \code{"s"}, \code{"w"}. Default \code{"w"}.}

\item{mod}{logical. If \code{"un"} specified for \code{axmethod}, whether or not to use Nan Li's modification for ages 5-14. Default \code{TRUE}.}

\item{SRB}{the sex ratio at birth (boys / girls), default 1.05}
}
\value{
The output is of class \code{lt_model_lq} with the components:
\item{lt}{ Life table matching given inputs}
\item{values}{ Associated values of \code{q0_5, q0_1, q15_45, q15_35}
and \code{e0}.}
}
\description{
Construct model life tables based on the Log-Quadratic (Wilmoth) estimates
with various choices of 2 input parameters:
\code{q0_5, q0_1, q15_45, q15_35} and \code{e0}. There are 8 possible
combinations (see examples below).
}
\details{
Due to limitations of the R language the notation for probability
of dying \code{nqx} is written \code{qx_n}, where \code{x} and \code{n} are
integers. For example \code{45q15} is represented as \code{q45_15}.
}
\note{
This function is ported from \code{MortalityEstimate::wilmothLT} experimental package by Marius Pascariu. The package is no longer maintained. The latest version can be found here: \url{https://github.com/mpascariu/MortalityEstimate}
}
\examples{

# Build life tables with various choices of 2 input parameters
# case 1: Using 5q0 and e0
L1 <- lt_model_lq(Sex = "b", q0_5 = 0.05, e0 = 65)
L1
ls(L1)

L1f <- lt_model_lq(Sex = "f", q0_5 = 0.05, e0 = 65)
L1m <- lt_model_lq(Sex = "m", q0_5 = 0.05, e0 = 65)

# case 2: Using 5q0 and 45q15
L2 <- lt_model_lq(Sex = "b", q0_5 = 0.05, q15_45 = 0.2)

# case 3: Using 5q0 and 35q15
L3 <- lt_model_lq(Sex = "b", q0_5 = 0.05, q15_35 = 0.125)

# case 4: Using 1q0 and e0
L4 <- lt_model_lq(Sex = "b", q0_1 = 0.01, e0 = 65)

# case 5: Using 1q0 and 45q15
L5 <- lt_model_lq(Sex = "b", q0_1 = 0.05, q15_45 = 0.2)

# case 6: Using 1q0 and 35q15
L6 <- lt_model_lq(Sex = "b", q0_1 = 0.05, q15_35 = 0.125)

# case 7: Using 45q15 and e0
L7 <- lt_model_lq(Sex = "b", q15_45 = 0.125, e0 = 65)

# case 8: Using 35q15 and e0
L8 <- lt_model_lq(Sex = "b", q15_35 = 0.15, e0 = 65)

}
