% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_rule.R
\name{lt_rule_1a0}
\alias{lt_rule_1a0}
\title{calculate a0 in different ways}
\usage{
lt_rule_1a0(
  rule = "ak",
  M0 = NULL,
  q0 = NULL,
  Sex = "m",
  IMR = NA,
  region = "w",
  SRB = 1.05
)
}
\arguments{
\item{rule}{character. Either \code{"ak"} (Andreev-Kingkade) or \code{"cd"} (Coale-Demeny).}

\item{M0}{numeric. Event exposure infant mortality rate.}

\item{q0}{a value or vector of values of m0, the death risk in the first year of life.}

\item{Sex}{character, either \code{"m"}, \code{"f"}, or \code{"b"}}

\item{IMR}{numeric. Optional. {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}}, the death probability in first year of life, in case available separately.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}

\item{SRB}{the sex ratio at birth (boys / girls), default 1.05}
}
\value{
a0, the estimated average age at death of those dying in the first year of life, either a single value or a vector of values.
}
\description{
This function wraps the Coale-Demeny and Andreev-Kingkade approximations for \code{a0}, which can come from \code{M0}, \code{q0}, or \code{IMR.}
}
\details{
If sex is given as both, \code{"b"}, then we calculate the male and female results separately, then weight them together using SRB. This is bad in theory, but the leverage is trivial, and it's better than using male or female coefs for the total population.

Neither Coale-Demeny nor Andreev-Kingkade have explicit a0 rules for both-sexes combined. There's not a good way to arrive at a both-sex a0 estimate without increasing data requirements (you'd need data from each sex, which are not always available). It's more convenient to blend sex-specific a0 estimates based on something. Here we use SRB to do this, for no other reason than it has an easy well-known default value. This is bad because it assumes no sex differences in infant mortality, but this choice has a trivial impact on results.
}
