% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{lt_rule_1a0_cd}
\alias{lt_rule_1a0_cd}
\title{Coale-Demeny a(0) as function of m(0), region, and sex.}
\usage{
lt_rule_1a0_cd(M0, IMR = NA, Sex = "m", region = "w")
}
\arguments{
\item{M0}{numeric. Event exposure infant mortality rate.}

\item{IMR}{numeric. Optional. {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}}, the death probability in first year of life, in case available separately.}

\item{Sex}{character. \code{"m"}, \code{"f"} or \code{"b"} for male, female, or both.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}
}
\value{
The average age at death in the first year of life a(0).
}
\description{
Coale-Demeny a(0) from Manual X Table 164. This is a rule of thumb.
In this and some other older texts, a(0) is known as a 'separation factor'.
}
\details{
If sex is given as both, \code{"b"}, then female values are taken, per the PAS spreadsheet. This function is not vectorized. Formulas for North, South, and West are identical- only East is different. If \code{IMR} is not given, then \code{M0} is converted to q(0) using the following approximation:
\enumerate{
\item{Find \eqn{\alpha , \beta}.}{ Look up the appropriate slope and intercept for the given sex and region.}
\item{calculate \eqn{a} as: }{\ifelse{html}{\out{a = M<sub>0</sub> * &beta;}}{\eqn{a = M_0 * \beta}}}
\item{calculate \eqn{b} as: }{\ifelse{html}{\out{b = 1 + M<sub>0</sub> *(1- &alpha;)}}{\eqn{b =  1 + M_0 * (1 - \alpha)}}}
\item{approximate {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}} as:}{ \ifelse{html}{\out{q<sub>0</sub> = (b<sup>2</sup>- &radic; [b -4*a*M<sub>0</sub>]) / (2*a)}}{\eqn{q_0 = \frac{ b - sqrt(b^2 - 4 * a * M_0) }{ 2 * a } }}}
\item{use {\ifelse{html}{\out{q<sub>0</sub>}}{\eqn{q_0}}} as}{ IMR, and applied directly to the Coale-Demeny piecewise linear formula.}
}
If \code{IMR} is given, then \code{M0} is disregarded, and transitivity is therefore not guaranteed. In this case, one has the option to use \code{lt_id_qm_a()} to derive \code{a(0)}, however discrepancies between these two parameters could force implausible results in \code{a(0)}, whereas the CD rule always gives something plausible.
}
\examples{
m0 <- seq(.001, .2, by = .001)
\dontrun{
plot(m0, sapply(m0, lt_rule_1a0_cd, Sex = "m", region = "e"), ylab = "a0",
		type = 'l', ylim = c(0,.36), lty = 2, col = "blue")
lines(m0,sapply(m0, lt_rule_1a0_cd, Sex = "m", region = "w"), col = "blue")
lines(m0,sapply(m0, lt_rule_1a0_cd, Sex = "f", region = "e"), lty = 2, col = "red")
lines(m0,sapply(m0, lt_rule_1a0_cd, Sex = "f", region = "w"), col = "red")
text(.15, lt_rule_1a0_cd(.15,Sex = "m", region = "e"),"males E",font=2)
text(.15, lt_rule_1a0_cd(.15,Sex = "m", region = "w"),"males N,W,S",font=2)
text(.15, lt_rule_1a0_cd(.15,Sex = "f", region = "e"),"females E",font=2)
text(.15, lt_rule_1a0_cd(.15,Sex = "f", region = "w"),"females N,W,S",font=2)

# compare with the Preston approximation
# constants identical after m0 = .107
m0 <- seq(.001,.107,by =.001)
a0CDm0 <- sapply(m0, lt_rule_1a0_cd, Sex = "m", region = "w")
a0CDpr <- 0.045 + 2.684 * m0
plot(m0, a0CDm0, type = 'l', lty = 2, col = "red")
lines(m0, a0CDpr)
plot(m0, (a0CDm0 - a0CDpr) * 365, main = "difference (days)", ylab = "days")
}
}
\references{
\insertRef{united1983manual}{DemoTools}
\insertRef{PAS}{DemoTools}
}
