% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_rule.R
\name{lt_rule_ak}
\alias{lt_rule_ak}
\title{Andreev-Kingkade approximation for a0}
\usage{
lt_rule_ak(m0 = NULL, q0 = NULL, Sex)
}
\arguments{
\item{m0}{a value or vector of values of m0, the death probability in the first year of life.}

\item{q0}{a value or vector of values of m0, the death risk in the first year of life.}

\item{Sex}{either "m" or "f"}
}
\value{
a0, the estimated average age at death of those dying in the first year of life, either a single value or a vector of values.
}
\description{
This function wraps the two approximations for a0 based on either q0 (IMR) or m0.
}
