% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_rule.R
\name{lt_rule_ak_m0_a0}
\alias{lt_rule_ak_m0_a0}
\title{estimates a0 using the Andreev-Kingkade rule of thumb starting with an event exposure rate}
\usage{
lt_rule_ak_m0_a0(m0, Sex)
}
\arguments{
\item{m0}{a value or vector of values of m0, the death risk in the first year of life.}

\item{Sex}{either "m" or "f"}
}
\value{
a0, the estimated average age at death of those dying in the first year of life, either a single value or a vector of values.
}
\description{
These formulas and cutpoints are based on a supplementary analysis from Andreev & Kingkade. The original formulation was in terms of IMR. There is also an analytic path to convert m0 to q0 and then use the original q0 cutpoints. Code ported from HMD code base.
}
