% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_rule.R
\name{lt_rule_ak_q0_a0}
\alias{lt_rule_ak_q0_a0}
\title{estimates a0 using the Andreev-Kingkade rule of thumb starting with IMR}
\usage{
lt_rule_ak_q0_a0(q0, Sex)
}
\arguments{
\item{q0}{a value or vector of values of q0, the death probability in the first year of life.}

\item{Sex}{either "m" or "f"}
}
\value{
a0, the estimated average age at death of those dying in the first year of life, either a single value or a vector of a_0 values.
}
\description{
\code{AKq02a0} Andreev Kingkade a0 method. This version has a 3-part segmented linear model, based on cut points in q0. Code ported from HMDLifeTables.
}
