% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_rule.R
\name{lt_rule_l0_1L0}
\alias{lt_rule_l0_1L0}
\title{Infer Life Table radix from 1L0
Infers the radix (\code{l0}) of a life table based on the value of \code{1L0}.
The function tests whether \code{1L0} is already a power of 10.
If it is not, the next higher power of 10 is used as the radix.
If \code{1L0 <= 1}, the radix is set to 1.}
\usage{
lt_rule_l0_1L0(L0)
}
\arguments{
\item{L0}{Numeric. Numeric value of \code{1L0} from which to infer the radix.}
}
\value{
A numeric scalar giving the inferred radix.
}
\description{
This helper is useful when reconstructing life tables from partial information,
such as when only \code{1L0} is available and the scaling factor (radix)
must be inferred.
}
\examples{
lt_rule_l0_1L0(10000)
lt_rule_l0_1L0(20)
lt_rule_l0_1L0(1)

}
