% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_regroup_age.R
\name{lt_single2abridged}
\alias{lt_single2abridged}
\title{calculate an abridged life table that is consistent with a life table by single year of age}
\usage{
lt_single2abridged(lx, nLx, ex, Age = 1:length(lx) - 1, ...)
}
\arguments{
\item{lx}{numeric. Vector of lifetable survivorship at single ages.}

\item{nLx}{numeric. Vector of lifetable exposure at single ages.}

\item{ex}{numeric. Vector of Age-specific remaining life expectancy at single ages.}

\item{Age}{integer. Lower bounds of single ages.}

\item{...}{optional args, not currently used.}
}
\value{
Abridged lifetable in \code{data.frame} with columns
\code{Age} integer. Lower bound of abridged age class
\code{AgeInt} integer. Age class widths.
\code{nMx} numeric. Age-specific central death rates.
\code{nAx} numeric. Average time spent in interval by those deceased in interval.
\code{nqx} numeric. Age-specific conditional death probabilities.
\code{lx} numeric. Lifetable survivorship
\code{ndx} numeric. Lifetable deaths distribution.
\code{nLx} numeric. Lifetable exposure.
\code{Sx} numeric. Survivor ratios in uniform 5-year age groups.
\code{Tx} numeric. Lifetable total years left to live above age x.
\code{ex} numeric. Age-specific remaining life expectancy.
}
\description{
Computes abridged life table columns based on the lx, nLx , and ex values from
a single year life table, in accordance with step 2.2 of the Human Life Table Protocol
https://www.lifetable.de/methodology.pdf. Output abridged life table has same open age group
as input single age life table
}
\details{
Similar to \code{lt_abridged()} details, forthcoming
}
