% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_single.R
\name{lt_single_mx}
\alias{lt_single_mx}
\title{calculate a single age lifetable}
\usage{
lt_single_mx(
  nMx,
  Age = 1:length(nMx) - 1,
  radix = 1e+05,
  a0rule = "ak",
  Sex = "m",
  region = "w",
  IMR = NA,
  mod = TRUE,
  SRB = 1.05,
  OAG = TRUE,
  OAnew = max(Age),
  extrapLaw = NULL,
  extrapFrom = max(Age),
  extrapFit = NULL,
  ...
)
}
\arguments{
\item{nMx}{numeric. Vector of mortality rates in abridged age classes.}

\item{Age}{integer. A vector of ages of the lower integer bound of the age classes.}

\item{radix}{numeric. Lifetable radix, \ifelse{html}{\out{l<sub>0}}{\eqn{l_0}}. Default 100000.}

\item{a0rule}{character. Either \code{"ak"} (default) or \code{"cd"}.}

\item{Sex}{character. Either male \code{"m"}, female \code{"f"}, or both \code{"b"}.}

\item{region}{character. North, East, South, or West: code{"n"}, code{"e"}, code{"s"}, code{"w"}. Default code{"w"}.}

\item{IMR}{numeric. Infant mortality rate \ifelse{html}{\out{q<sub>0}}{\eqn{q_0}}, in case available and \code{nqx} is not specified. Default \code{NA}.}

\item{mod}{logical. If \code{"un"} specified for \code{axmethod}, whether or not to use Nan Li's modification for ages 5-14. Default \code{TRUE}.}

\item{SRB}{the sex ratio at birth (boys / girls), detault 1.05}

\item{OAG}{logical. Whether or not the last element of \code{nMx} (or \code{nqx} or \code{lx}) is an open age group. Default \code{TRUE}.}

\item{OAnew}{integer. Desired open age group (5-year ages only). Default \code{max(Age)}. If higher then rates are extrapolated.}

\item{extrapLaw}{character. If extrapolating, which parametric mortality law should be invoked? Options include
\code{"Kannisto", "Kannisto_Makeham", "Makeham", "Gompertz", "GGompertz", "Beard",    "Beard_Makeham", "Quadratic"}. Default \code{"Kannisto"} if the highest age is at least 90, otherwise \code{"makeham"}. See details.}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality.}

\item{extrapFit}{integer vector. Ages to include in model fitting. Defaults to all ages \code{>          =60}.}

\item{...}{Other arguments to be passed on to the
\code{\link[MortalityLaws]{MortalityLaw}} function.}
}
\value{
Lifetable in data.frame with columns
\itemize{
\item{Age}{integer. Lower bound of abridged age class},
\item{AgeInt}{integer. Age class widths.}
\item{nMx}{numeric. Age-specific central death rates.}
\item{nAx}{numeric. Average time spent in interval by those deceased in interval. }
\item{nqx}{numeric. Age-specific conditional death probabilities.}
\item{lx}{numeric. Lifetable survivorship}
\item{ndx}{numeric. Lifetable deaths distribution.}
\item{nLx}{numeric. Lifetable exposure.}
\item{Sx}{numeric. Survivor ratios in uniform 5-year age groups.}
\item{Tx}{numeric. Lifetable total years left to live above age x.}
\item{ex}{numeric. Age-specific remaining life expectancy.}
}
}
\description{
Fuller description forthcoming
}
\details{
Similar to \code{lt_abridged()} details, forthcoming
}
