% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_lc_lim.R
\name{lt_smooth_ambiguous}
\alias{lt_smooth_ambiguous}
\title{Smooth and apply lt_ambiguous}
\usage{
lt_smooth_ambiguous(input, ExtraArgs, ...)
}
\arguments{
\item{input}{data.frame. with cols: Date, Sex, Age, nMx (opt), nqx (opt), lx (opt)}

\item{ExtraArgs}{list. Parameters produced in \code{interp_lc_lim} environment function.}

\item{...}{Other arguments to be passed on to the \code{\link[DemoTools]{lt_abridged}} function.}
}
\description{
Considering different mortality input for each sex/year data,
smooth olders with gompertz or kannisto in case no law was specified,
and return a data.frame with standard LT.
}
\details{
Gompertz is chosen if last age is less than 90. Else Kannisto.
}
