% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{lx2dx}
\alias{lx2dx}
\title{derive lifetable deaths from survivorship}
\usage{
lx2dx(lx)
}
\arguments{
\item{lx}{numeric vector. Age-specific lifetable survivorship}
}
\value{
ndx numeric vector of lifetable deaths.
}
\description{
This lifetable identity is the same no matter what kind of lifetable
we're making. You can find it in any demography textbook.
}
\details{
The vector returned is the same length as \code{lx} and it sums to the lifetable radix. 
If the radix is one then this is the discrete deaths distribution.
}
