% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mA_swe}
\alias{mA_swe}
\title{Swedish abridged mortality rates}
\format{
A data frame with:
\describe{
\item{Date}{Reference time for the rates estimate.}
\item{Age}{Inferior age for abridged groups. Carefull: last age 100 is not an OAG}
\item{Sex}{Male \code{m} and female \code{m}.}
\item{nMx}{Mortality rates.}
}
}
\source{
Human Mortality Database. Retrieved 2021-20-01, from \url{https://mortality.org}
}
\usage{
mA_swe
}
\description{
Mortality rates in tidy format for each sex in dates 1990-07-01, 2000-07-01, 2010-07-01
}
\keyword{datasets}
