% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGESMTH.R
\name{mav_smth}
\alias{mav_smth}
\title{Smooth in 5-year age groups using a moving average}
\usage{
mav_smth(Value, Age, OAG = TRUE, n = 3)
}
\arguments{
\item{Value}{numeric vector of (presumably) counts in 5-year age groups.}

\item{Age}{integer vector of age group lower bounds.}

\item{OAG}{logical. Whether or not the top age group is open. Default \code{TRUE}.}

\item{n}{integer. The width of the moving average. Default 3 intervals (x-5 to x+9).}
}
\value{
numeric vector of smoothed counts in 5-year age groups.
}
\description{
Smooth data in 5-year age groups.
}
\details{
This function calls \code{zigzag()}, but prepares data in a way consistent with other methods called by \code{agesmth()}. It is probably preferable to call \code{zigzag()} from the top level, or else call this method from \code{agesmth()} for more control over tail imputations.

This function calls \code{mav()}, which itself relies on the more general \code{ma()}. We lose the lowest and highest ages with this method, unless \code{n=1}, in which case data is returned in the original 5-year age groups.
}
\examples{
Value <- c(13331,4151,1746,1585,3859,8354,11146,12076,
		12216,12016,12473,11513,12899,11413,12710,11516,11408,6733,4031,2069)
Age <- c(0,1,seq(5,90,by=5))
# defaults
ns   <- sapply(1:5,mav_smth,Value=Value,Age=Age,OAG=TRUE)
cols <- paste0(gray(seq(.8,0,length=5)),"A0")
lwds <- seq(4,1,length=5)
\dontrun{
plot(Age, Value,pch=16)
matplot(as.integer(rownames(ns)),ns,type='l',
		col = cols,
		lty = 1,
		add = TRUE,
		lwd = lwds)
legend("topright", col = cols, lty = 1, lwd = lwds, legend = paste("n =",1:5))
}
}
