% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{maxA2abridged}
\alias{maxA2abridged}
\title{Determine abridged ages up to a given maximum age group.}
\usage{
maxA2abridged(ageMax = 80)
}
\arguments{
\item{ageMax}{integer. Default 80.}
}
\value{
integer. Vector of ages, e.g. \code{c(0,1,5,10,15,...)}.
}
\description{
Produce standard abridged age groups (lower bounds) up to a specified maximum age group.
}
\details{
If the highest age group is not evenly divisible by 5 then age classes only go up to its 5-year lower bound.
}
\examples{
maxA2abridged(80)
all(maxA2abridged(100) == maxA2abridged(102))
}
