% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig_estimate_rc.R
\name{mig_estimate_rc}
\alias{mig_estimate_rc}
\title{Estimate Rogers-Castro migration age schedule}
\usage{
mig_estimate_rc(ages, mx, num_pars, ...)
}
\arguments{
\item{ages}{numeric. A vector of ages.}

\item{mx}{numeric. A vector of observed age-specific migration rates.}

\item{num_pars}{integer. Number of parameters to be estimated in the model.}

\item{...}{additional inputs to stan, see ?rstan::stan for details.}
}
\description{
Given a set of ages and observed age-specific migration rates, estimate the parameters of a Rogers-Castro model migration schedule. 
Choose between a 7,9,11 or 13 parameter model.
}
