% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig_resid.R
\name{mig_resid_stock}
\alias{mig_resid_stock}
\alias{mig_resid_cohort}
\alias{mig_resid_time}
\title{Estimate net migration using residual methods: stock change,
time even flow and cohort even flow}
\usage{
mig_resid_stock(
  pop_m_mat,
  pop_f_mat,
  sr_m_mat,
  sr_f_mat,
  asfr_mat,
  srb_vec,
  ages = NULL,
  ages_fertility = NULL,
  years_pop = NULL,
  years_sr = NULL,
  years_asfr = NULL,
  years_srb = NULL
)

mig_resid_cohort(
  pop_m_mat,
  pop_f_mat,
  sr_m_mat,
  sr_f_mat,
  asfr_mat,
  srb_vec,
  ages = NULL,
  ages_fertility = NULL,
  years_pop = NULL,
  years_sr = NULL,
  years_asfr = NULL,
  years_srb = NULL
)

mig_resid_time(
  pop_m_mat,
  pop_f_mat,
  sr_m_mat,
  sr_f_mat,
  asfr_mat,
  srb_vec,
  ages = NULL,
  ages_fertility = NULL,
  years_pop = NULL,
  years_sr = NULL,
  years_asfr = NULL,
  years_srb = NULL
)
}
\arguments{
\item{pop_m_mat}{A \code{numeric} matrix with population counts. Rows should
be ages and columns should be years. Only five year age groups are supported.
See examples.}

\item{pop_f_mat}{A \code{numeric} matrix with population counts. Rows should
be ages and columns should be years. Only five year age groups are supported.
See examples.}

\item{sr_m_mat}{A \code{numeric} matrix with survival rates for males. Rows
should be ages and columns should be years. ** This matrix should have
one column less than \code{pop_m_mat} and \code{pop_f_mat}. For example,
if the last year in these matrices is 2050, then the last year in
\code{sr_m_mat} should be 2045. **}

\item{sr_f_mat}{A \code{numeric} matrix with survival rates for females. Rows
should be ages and columns should be years. ** This matrix should have
one column less than \code{pop_m_mat} and \code{pop_f_mat}. For example,
if the last year in these matrices is 2050, then the last year in
\code{sr_f_mat} should be 2045. **.}

\item{asfr_mat}{A \code{numeric} matrix with age specific fertility rates.
Rows should be ages and columns should be years. ** This matrix should have
one column less than \code{pop_m_mat} and \code{pop_f_mat}. For example,
if the last year in these matrices is 2050, then the last year in
\code{asfr_mat} should be 2045**. This row will usually have fewer age groups
(rows) than in the population matrices or survival matrices, so the user
needs to supply the specific ages in the \code{ages_fertility} argument.}

\item{srb_vec}{A \code{numeric} vector of sex ratios at birth for every year.
The years should be the same as the years in \code{sr_m_mat},
\code{sr_f_mat}, and \code{asfr_mat}.}

\item{ages}{A \code{numeric} vector of ages used in the rows in
\code{pop_m_mat}, \code{pop_f_mat}, \code{sr_m_mat}, \code{sr_f_mat}.}

\item{ages_fertility}{A \code{numeric} vector of ages used in the rows in
\code{asfr_mat}.}

\item{years_pop}{Years used in the column names of population. If
\code{pop_m_mat} or \code{pop_f_mat} doesn't have column names, these
names are used. Otherwise ignored.}

\item{years_sr}{Years used in the column names of survival rates. If
\code{sr_r_mat} doesn't have column names, these names are used. Otherwise
ignored.}

\item{years_asfr}{Years used in the column names of age-specific fertility
rate. If code{asfr_r_mat} doesn't have column names, these names are used.
Otherwise ignored.}

\item{years_srb}{Years used in the column names of sex-ratio at birth. If
\code{srb_r_mat} is not named, these names are used. Otherwise ignored.}
}
\value{
A list with two matrices. One is for males (called \code{mig_m}) and the
other for females (called \code{mig_f}). Both matrices contain net migration
estimates by age/period using one of the three methods.
}
\description{
Estimate net migration using residual methods: stock change,
time even flow and cohort even flow
}
\details{
\enumerate{
\item The stock method (\code{mig_resid_stock}) is the difference in stocks that
survive between t and t+5, and the first age group is based on the difference
with the surviving births by sex.  It provides net migrants by lexis cohort
parallelograms, and basically such info gets used as end-period migration
since the migrants don't get exposed to mortality within the period.
\item The time even flow (\code{mig_resid_time}) method uses the result from
the first option, but splits it back into lexis period squares and assumes
that half of the net migrants get exposed to the mortality risk during this
period. Such info can get used as evenly distributed migration by period,
but the assumptions lead to zig-zag age patterns that are highly implausible.
\item The cohort even flow (\code{mig_resid_cohort}) method provides the most
meaningful pattern of net migration by age consistent by cohort and assumes
an evenly distribution within the 5-year period, and half of the migrants
get exposed both fertility and mortality within this period.
}
}
\examples{

# The data is loaded with DemoTools

################ Stock change method #####################

mig_res <-
 mig_resid_stock(
   pop_m_mat = pop_m_mat,
   pop_f_mat = pop_f_mat,
   sr_m_mat = sr_m_mat,
   sr_f_mat = sr_f_mat,
   asfr_mat = asfr_mat,
   srb_vec = srb_vec,
   ages = ages,
   ages_fertility = ages_fertility
 )

# Net migration for males using stock change method
mig_res$mig_m

# Net migration for females using stock change method
mig_res$mig_f


################ cohort even flow method  #####################

# We reuse the same data from before

mig_res <-
  mig_resid_cohort(
    pop_m_mat = pop_m_mat,
    pop_f_mat = pop_f_mat,
    sr_m_mat = sr_m_mat,
    sr_f_mat = sr_f_mat,
    asfr_mat = asfr_mat,
    srb_vec = srb_vec,
    ages = ages,
    ages_fertility = ages_fertility
  )

# Net migration for males using the cohort even flow method
mig_res$mig_m

# Net migration for females using the cohort even flow method
mig_res$mig_f

################ time even flow method  #####################

# We reuse the same data from before

mig_res <-
  mig_resid_time(
    pop_m_mat = pop_m_mat,
    pop_f_mat = pop_f_mat,
    sr_m_mat = sr_m_mat,
    sr_f_mat = sr_f_mat,
    asfr_mat = asfr_mat,
    srb_vec = srb_vec,
    ages = ages,
    ages_fertility = ages_fertility
  )

# Net migration for males using the time even flow method
mig_res$mig_m

# Net migration for females using the time even flow method
mig_res$mig_f

}
