% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mig_un_families}
\alias{mig_un_families}
\title{Proportion of net migrants by age and sex for considered migration profiles}
\format{
A data frame with:
\describe{
\item{family}{Types Family, Male Labor or Female Labor.}
\item{sex}{Male and Female.}
\item{mig_sign}{Inmigration or Emigration.}
\item{age}{Simple ages from 0 to 80 (OAG).}
\item{prop}{Proportion of net migrants due to that sex and age.}
}
}
\source{
UN spreadsheet "UNPD_Migration Age Profiles.xlsx"
}
\usage{
mig_un_families
}
\description{
Roger-Castro estimated proportion of total net migrants using parameters from \code{mig_un_params} data.
}
\keyword{datasets}
