% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{mx2qx}
\alias{mx2qx}
\title{derive nqx from nMx and nax}
\usage{
mx2qx(nMx, nax, AgeInt = inferAgeIntAbr(vec = nMx))
}
\arguments{
\item{nMx}{numeric vector. Age specific death rates.}

\item{nax}{numeric vector. Average time spent in interval by those dying in interval}

\item{AgeInt}{integer width of each age interval}
}
\value{
nqx vector of age specific death probabilities derived via identity.
}
\description{
This is the common ax-qx-mx identity as found in any demography text.
}
