% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{mxax2qx}
\alias{mxax2qx}
\title{derive nqx from nMx and nax}
\usage{
mxax2qx(nMx, nax, AgeInt, closeout = TRUE, IMR)
}
\arguments{
\item{nMx}{numeric vector. Age specific death rates.}

\item{nax}{numeric vector. Average time spent in interval by those dying in interval}

\item{AgeInt}{integer width of each age interval}

\item{closeout}{logical (default \code{TRUE}). Set to 1 if TRUE. Identity otherwise}

\item{IMR}{numeric. Optional q0 to impute, in case available separately.}
}
\value{
nqx vector of age specific death probabilities derived via identity.
}
\description{
This is the common ax-qx-mx identity as found in any demography text. 
This is a more full-service wrapper of \code{mx2qx()}, with closeoput options and optional age 0 
treatment.
}
