% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{mxax2qx_Backstop}
\alias{mxax2qx_Backstop}
\title{adjust ax estimate so that qx less than or equal to 1}
\usage{
mxax2qx_Backstop(nMx, nax, AgeInt)
}
\arguments{
\item{nMx}{numeric. Vector of age-specific death rates.}

\item{nax}{numeric. Vector of average time spent in interval by those dying in interval.}

\item{AgeInt}{integer. Vector of age class widths.}
}
\description{
Sometimes a given age interval, death rate, and a(x) imply a death probability that is greater than 1. In this case either the interval needs to be extended or a(x) decreased. This especially arises with mid interval a(x) has been assumed in five-year age groups. This backstop reduces a(x) by assuming a constant death rate over the single ages within the interval, assuming mid interval a(x) for each single age, producing nq(x) by identity from the (5) single ages.
}
\details{
nMx equal to 2 will imply nqx of 1 by this formula. Implied nqx greater than 1 after this procedure are returned as 1. This is not vectorized!
}
\examples{
# implies a qx > 1
mx     <- .65
AgeInt <- 5
ax     <- 2.5
# this is a problematic case
(mx2qx(mx, ax, AgeInt))
# here a workable value
(mxax2qx_Backstop(mx, ax, AgeInt))
#' # still less than 1
mxax2qx_Backstop(1.99, ax, AgeInt) < 1
# unity
(mxax2qx_Backstop(2, ax, AgeInt))
# qx imputed as 1
(mxax2qx_Backstop(2.1, ax, AgeInt))
}
