% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{mxorqx2ax}
\alias{mxorqx2ax}
\title{wrapper to invoke PAS or UN ax methods given qx or mx}
\usage{
mxorqx2ax(nMx, nqx, axmethod = c("pas", "un")[1], AgeInt, IMR = NA,
  Sex, region, OAG = TRUE, mod = TRUE)
}
\arguments{
\item{nMx}{numeric. Vector of mortality rates in abridged age classes.}

\item{nqx}{numeric. Vector of conditional death probabilities in abridged age classes.}

\item{axmethod}{character. Either \code{"pas"} or \code{"un"}.}

\item{AgeInt}{integer. Vector of age class widths. Ddefault \code{inferAgeIntAbr(Age = Age)}.}

\item{IMR}{numeric. Infant mortality rate \ifelse{html}{\out{q<sub>0}}{\eqn{q_0}}, in case available and \code{nqx} is not specified. Default \code{NA}.}

\item{Sex}{character. Either male \code{"m"}, female \code{"f"}, or both \code{"b"}.}

\item{region}{character. North, East, South, or West: code{"n"}, code{"e"}, code{"s"}, code{"w"}. Default code{"w"}.}

\item{OAG}{logical. Whether or not the last element of \code{nMx} is the open age group Default \code{TRUE}.}

\item{mod}{logical. If \code{"un"} specified for \code{axmethod}, whether or not to use Patrick Gerland's modification for ages 5-14. Default \code{TRUE}.}
}
\value{
nax average contribution to exposure of those dying in the interval.
}
\description{
Given either mx or qx, call either the \code{axUN()} or \code{axPAS()} functions.
}
\references{
\insertRef{greville1977short}{DemoTools}
\insertRef{un1982model}{DemoTools}
\insertRef{arriaga1994population}{DemoTools}
\insertRef{mortpak1988}{DemoTools}
\insertRef{united1983manual}{DemoTools}
\insertRef{PAS}{DemoTools}
}
