% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGEHEAP.R
\name{Myers}
\alias{Myers}
\title{Calculate Myer's blended index of age heaping}
\usage{
Myers(Value, Age, ageMin = 10, ageMax = 90)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{integer. The lowest age included in calculations. Default 10.}

\item{ageMax}{integer. The upper age bound used for calculations. Default 90.}
}
\value{
the index value
}
\description{
Implementation following the PASEX spreadsheet SINGAGE. Myers' measures preferences for each of the ten possible digits
as a blended index. It is based on the principle that in the absence of age heaping, the aggregate population of each age ending in one of the digits
 0 to 9 should represent 10 percent of the total population.
}
\details{
\code{ageMax} is the hard upper bound, treated as interval. If you want ages
20 to 89, then give \code{ageMin = 20} and \code{ageMax = 90}, not 89.
}
\examples{
Value <- c(80626,95823,104315,115813,100796,105086,97266,116328,
		75984,89982,95525,56973,78767,65672,53438,85014,
		47600,64363,42195,42262,73221,30080,34391,29072,
		20531,66171,24029,44227,24128,23599,82088,16454,
		22628,17108,12531,57325,17220,28425,16206,17532,
		65976,11593,15828,13541,8133,44696,11165,18543,
		12614,12041,55798,9324,10772,10453,6773,28358,
		9916,13348,8039,7583,42470,5288,5317,6582,
		3361,17949,3650,5873,3279,3336,27368,1965,
		2495,2319,1335,12022,1401,1668,1360,1185,
		9167,424,568,462,282,6206,343,409,333,291,4137,133,169,157,89,2068,68,81,66,57)
Age <- 0:99
Myers(Value, Age, 10, 90) 

Myers(Value, Age, 10, 90) * 2 #47.46, replicates SINGAGE males
}
\references{
\insertRef{myers1954accuracy}{DemoTools}
\insertRef{spoorenberg2007quality}{DemoTools}
}
