% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noumbissi.R
\name{Noumbissi}
\alias{Noumbissi}
\title{calculate Noubmissi's digit heaping index}
\usage{
Noumbissi(Value, Age, ageMin = 20, ageMax = 65, digit = 0)
}
\arguments{
\item{Value}{a vector of demographic counts by single age}

\item{Age}{a vector of ages corresponding to the lower integer bound of the counts}

\item{ageMin}{the lowest age included in calcs. Default 20}

\item{ageMax}{the upper age bound used for calcs. Default 65}

\item{digit}{which terminal digit do we calculate for?}
}
\value{
the index value
}
\description{
this method compares single terminal digit numerators to denominators 
consisting in 5-year age groups centered on the digit in question. As seen in Spoorenberg (2007)
}
