% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop1m_ind}
\alias{pop1m_ind}
\title{Indian male population 1991}
\format{
A numeric vector of length 101
}
\source{
REGISTRAR GENERAL & CENSUS COMMISSIONER, INDIA, 1976,
  Census of India 1971, Series 1, India, Social and Cultural Tables,
  Part II-C(ii), New Delhi
}
\usage{
pop1m_ind
}
\description{
Indian male population 1991
}
\keyword{datasets}
