% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop1m_pasex}
\alias{pop1m_pasex}
\title{Male population by 1 year age groups}
\format{
A numeric vector of length 100
}
\source{
The data comes from
  \url{https://URL}
}
\usage{
pop1m_pasex
}
\description{
Male population by 1 year age groups from PASEX SINGAGE
}
\keyword{datasets}
