% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop1m_rus2002}
\alias{pop1m_rus2002}
\title{Russian census 2002 male population by 1 year age groups}
\format{
A numeric vector of length 101
}
\source{
The data comes from
\url{http://www.demoscope.ru/weekly/ssp/rus2002_01.php}
}
\usage{
pop1m_rus2002
}
\description{
Male population by 1 year age groups from Russian census help on 2002-10-16
}
\keyword{datasets}
