% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop5_mat}
\alias{pop5_mat}
\title{Matrix of population over 5 years}
\format{
A matrix of 21 rows and 5 columns
}
\source{
The data comes from
\url{https://URL}
}
\usage{
pop5_mat
}
\description{
Matrix of population over 5 years, 5 year age groups, some population 1950-1954
}
\keyword{datasets}
