% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop5m_pasex}
\alias{pop5m_pasex}
\title{Male population by 5 year age groups}
\format{
A numeric vector of length 17
}
\source{
The data comes from
  \url{https://URL}
}
\usage{
pop5m_pasex
}
\description{
Male population by 5 year age groups from PASEX AGESMTH
}
\keyword{datasets}
