% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_f_mat}
\alias{pop_f_mat}
\title{Population matrix for females five year age groups between 1950 and 2050 for
unknown country}
\format{
A matrix of dimensions 21 x 21
}
\source{
Migration residual PAS spreadhseet
}
\usage{
pop_f_mat
}
\description{
Population matrix for females five year age groups between 1950 and 2050 for
unknown country
}
\keyword{datasets}
