% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_m_mat_five}
\alias{pop_m_mat_five}
\title{Population matrix for males five year age groups between 1950 and 2050}
\format{
A matrix of dimensions 21 x 21
}
\source{
Migration residual PAS spreadsheet
}
\usage{
pop_m_mat_five
}
\description{
Population matrix for males five year age groups between 1950 and 2050 for
unknown country
}
\keyword{datasets}
