% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_m_mat_single}
\alias{pop_m_mat_single}
\title{Population matrix for males single ages between 1999 and 2019}
\format{
A matrix of dimensions 101 x 21
}
\source{
Migration residual PAS spreadsheet
}
\usage{
pop_m_mat_single
}
\description{
Population matrix for males single ages between 1999 and 2019 for
Sweden
}
\keyword{datasets}
