% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popA_earlier}
\alias{popA_earlier}
\title{Abridged population from PAS AGEINT -- earlier}
\format{
A numeric vector of length 18
}
\source{
The data comes from
  \url{http://}
}
\usage{
popA_earlier
}
\description{
Abridged population from PAS AGEINT -- earlier, 1980
}
\keyword{datasets}
