% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popA_later}
\alias{popA_later}
\title{Abridged population from PAS AGEINT -- later}
\format{
A numeric vector of length 18
}
\source{
The data comes from
\url{http://}
}
\usage{
popA_later
}
\description{
Abridged population from PAS AGEINT -- later, 1990
}
\keyword{datasets}
