% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_mortality.R
\name{print.extra_mortality}
\alias{print.extra_mortality}
\title{Print function for extra_mortality method}
\usage{
\method{print}{extra_mortality}(x, ...)
}
\arguments{
\item{x}{An object of the class \code{"extra_mortality"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print function for extra_mortality method
}
\keyword{internal}
