% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{qxax2mx}
\alias{qxax2mx}
\title{derive nMx from nqx and nax}
\usage{
qxax2mx(nqx, nax, AgeInt = inferAgeIntAbr(vec = nqx))
}
\arguments{
\item{nqx}{numeric vector. Age specific death probabilities.}

\item{AgeInt}{integer width of each age interval}

\item{ax}{numeric vector. Average time spent in interval by those dying in interval}
}
\value{
nMx vector of age specific death rates derived via identity.
}
\description{
This is the common ax-qx-mx identity as found in any demography text.
}
