% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abacus.R
\name{REGRES}
\alias{REGRES}
\title{Calculate simple linear model by hand.}
\usage{
REGRES(X, Y, N)
}
\arguments{
\item{X}{x}

\item{Y}{y}

\item{N}{length of X and Y.}
}
\value{
A list containing, A, B, R, SUMSQ, and potentially an error code.
}
\description{
A legacy function internal to the mortpak lifetable. It calculates intercept, slope, correlation coef, and the residual sum of squares. Just like \code{lm()}? Unchecked.
}
\details{
Used internally by \code{AbacusLIFTB()}.
}
\references{
\insertRef{mortpak1988}{DemoTools}
}
