% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale.vector}
\alias{rescale.vector}
\title{rescale a vector proportionally to a new sum}
\usage{
rescale.vector(x, scale = 1)
}
\arguments{
\item{x}{numeric vector}

\item{scale}{what you want the vector to sum to. Default 1.}
}
\value{
the vector, rescaled
}
\description{
THis is a frequently needed operation, so it's being added as a general utility.
}
\details{
For a distribution, use \code{scale = 1}. For percentages, use \code{scale = 100}, etc.
}
\examples{
x <- runif(10)
sum(x)
xx <- rescale.vector(x,100)
sum(xx)
}
