% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rlog}
\alias{rlog}
\title{logging that doesn't cause jams}
\usage{
rlog(x, base = exp(1))
}
\arguments{
\item{x}{a numeric or complex vector.}

\item{a}{positive or complex number: the base with respect to which
logarithms are computed.  Defaults to \code{e=exp(1)}.}
}
\description{
Zeros happen. Logging zeros leads to errors in downstream code.
For the sake of robustness, we replace 0s with a small positive number to be 
able to continue with calcs.
}
\details{
Use judiciously, since errors are good too.
}
