% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGESEX.R
\name{sexRatioScore}
\alias{sexRatioScore}
\title{calculate the PASEX sex ratio score}
\usage{
sexRatioScore(Males, Females, Age, ageMin = 0, ageMax = max(Age))
}
\arguments{
\item{Males}{numeric. A vector of population counts for males by age}

\item{Females}{numeric. A vector of population counts for females by age}

\item{Age}{integer. A vector of ages corresponding to the lower integer bound of the counts}

\item{ageMin}{integer. The lowest age included in calcs. Default 0}

\item{ageMax}{integer. The upper age bound used for calcs. Default \code{max(Age)}}
}
\value{
the value of the index
}
\description{
A single ratio is defined as males per 100 females.
Taking the first difference of the ratios over age would give 0s
if the ratio were constant. The average absolute difference over
age defines the index. This comes from the PASEX spreadsheet called AGESEX.
}
\details{
Age groups must be of equal intervals. Five year age groups are assumed.
 We also assume that the final age group is open, unless \code{ageMax < max(Age)}
}
\examples{
Males   <- c(4677000,4135000,3825000,3647000,3247000,2802000,2409000,2212000,
		1786000,1505000,1390000,984000,745000,537000,346000,335000)
Females <- c(4544000,4042000,3735000,3647000,3309000,2793000,2353000,2112000,
		1691000,1409000,1241000,887000,697000,525000,348000,366000)
Age     <- seq(0, 75, by = 5)
sexRatioScore(Males, Females, Age)  # 2.2, matches pasex
}
