% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify.text}
\alias{simplify.text}
\title{convert strings to concatenation of lower case alphabet}
\usage{
simplify.text(string)
}
\arguments{
\item{string}{character string}
}
\value{
character string
}
\description{
all characters are converteusd to lower case, and all non-alphabet characters are removed. 
This is usefull for reducing checks on names of user-specified strings, like \code{method} arguments.
For example, \code{"Carrier-Farrag"}, \code{"Carrier Farrag"}, or \code{"carrier_farrag"} all get 
converted to \code{"carrierfarrag"}.
}
\examples{
simplify.text(c("carrier_farrag","CarrierFarrag","Carrier-Farrag"))
}
