% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMono.R
\name{splitMono}
\alias{splitMono}
\title{Split age groups using a monotonic spline.}
\usage{
splitMono(Value, AgeInt, Age, OAG = FALSE)
}
\arguments{
\item{Value}{numeric. Vector of (presumably) counts in grouped ages.}

\item{AgeInt}{integer or numeric. Vector of age intervals.}

\item{Age}{numeric. Vector of ages corresponding to the lower integer bound of the age range.}

\item{OAG}{boolean. Argument that determines whether the final age group (assumed open ended) is kept as it is or has the same length as the rest of the age groups. Default is FALSE, i.e. use the same length for the final age group.}
}
\value{
Numeric. vector of single smoothed age counts.
}
\description{
Take the cumulative sum of \code{Value} and then run a monotonic spline through it. The first differences split back single-age estimates of \code{Value}. Optionally keep the open age group untouched.
}
\details{
The \code{"monoH.FC"} method of \code{stats::splinefun()} is used to fit the spline because 1) it passes exactly through the points, 2) it is monotonic and therefore guarantees positive counts, and 3) it seems to be a bit less wiggly (lower average first differences of split counts). Single-age data is returned as-is. If you want to use this function as a smoother you first need to group to non-single ages.
}
\examples{
Value                <- structure(c(88623, 90842, 93439, 96325, 99281, 102051, 104351,
			 106555, 109170, 112188, 113582, 112614, 108904, 102622, 95867,
			 80874, 60196, 37523, 17927, 5642, 1110), .Names = c("0", "5",
			 "10", "15", "20", "25", "30", "35", "40", "45", "50", "55", "60",
			 "65", "70", "75", "80", "85", "90", "95", "100"))

# overwrite open age group with a single age estimate for that age
# (doesn't extrapolate)
splitMono(Value)
# or respect open age group
splitMono(Value, OAG = TRUE)

# Also accepts single ages:
Value                <- structure(pop1m_ind, .Names = 0:100)

		\dontrun{
	ages                <- seq(0,100,5)
 plot(splitMono(Value),xlab = 'Age', ylab = 'Counts', type = 'l',main = 'Ungrouped counts')
		}
}
\references{
\insertRef{fritsch1980monotone}{DemoTools}
}
