% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitToSingleAges}
\alias{splitToSingleAges}
\title{Convert arbitrary age groupings into single years of age}
\usage{
splitToSingleAges(Value, Age, OAG = FALSE)
}
\arguments{
\item{Value}{numeric vector of age group counts}

\item{Age}{numeric vector of ages corresponding to the lower integer bound of the age range.}

\item{OAG}{boolean argument that determines whether the final age group (assumed open ended) is kept as it is or has the same length as the rest of the age groups. Default is FALSE, i.e. use the same length for the final age group.}
}
\value{
numeric vector of counts for single year age groups.
}
\description{
Splits aggregate counts for a vector of age groups of a common width into single year age groups.
}
\details{
Assumes that all age groups are equal width. (Default is 5-year age groups.) Also, assumes that the population is uniformly distributed across each age interval. If there is a split under 5 age group (0 and 1-4 age groups), use \code{groupAges()} to consolidate before using this function. The default setting is to assume that the final age group is the same width as the other age groups.
}
\examples{
MalePop <- c(9544406,7471790,11590109,11881844,11872503,12968350,            
11993151,10033918,14312222,8111523,15311047,6861510,13305117,7454575,
             9015381,10325432,9055588,5519173)
Ages <- seq(0,85, by=5)
splitToSingleAges(MalePop, Ages)
splitToSingleAges(MalePop, Ages, OAG = TRUE)
}
