% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spoorenberg.R
\name{Spoorenberg}
\alias{Spoorenberg}
\title{Spoorenbergs total modified Whipple index}
\usage{
Spoorenberg(Value, Age, ageMin = 20, ageMax = 65)
}
\arguments{
\item{Value}{a vector of demographic counts by single age}

\item{Age}{a vector of ages corresponding to the lower integer bound of the counts}

\item{ageMin}{the lowest age included in calcs. Default 20}

\item{ageMax}{the upper age bound used for calcs. Default 65}
}
\value{
the index value
}
\description{
This index consists in calculating the Noumbissi index, Wi, for each
terminal digit, then taking the sum of the absolute deviations of these from unity.
}
\details{
\code{ageMin} and \code{ageMax} are applied to numerator ages, not denominators.
 Denominators are always 5-year age groups centered on the digit in question,
and these therefore stretch into ages a bit higher or lower than the numerator ages.
}
