% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprague.R
\name{spragueCloseout}
\alias{spragueCloseout}
\title{blend the Sprague upper boundary age estimates into monotonic spline estimates}
\usage{
spragueCloseout(popmat, pops, pivotAge = 90)
}
\arguments{
\item{popmat}{a numeric matrix of population counts in 5-year age groups, with integer-labeled 
margins (age in rows and year in columns).}

\item{pops}{optional numeric matrix of single age population counts derived from \code{popmat}.}

\item{pivotAge}{integer (default 90). Age at which to switch to spline-based estimates.}
}
\value{
numeric matrix of age by year estimates of single-age counts.
}
\description{
A simple monotonic spline on the cumulative sum of population counts
may return more convincing single age count estimates than the Sprague splitting method.
This function blends the Sprague estimates starting at \code{pivotAge}.
}
\details{
The \code{pivotAge} must be at least 10 years below the maximum age detected from 
\code{rownames(popmat)}, but not lower than 80. In the exact \code{pivotAge}, we may either take
the Sprague estimates or the spline estimates, depending on which is larger, then the single-age estimates
for this 5-year age group are rescaled to sum to the original total in \code{popmat}. Higher ages are taken from
the spline-based age splits. The spline results are derive from the \code{"monoH.FC"} method of \code{splinefun()} 
on the cumulative sum of the original age grouped data. One could use this function to perform the same
closeout to Grabill estimates, if these are given via the \code{pops} argument. See examples. Note
that the Grabill split method mixed with this closeout will not necessarily preserve the annual totals,
and this function performs to rescaling. The open age group is preserved (and must be included in \code{popmat}).
}
\examples{
 popmat <- structure(c(54170.08, 44774.6, 42141.587, 38463.515, 34405.607, 
30386.087, 26933.215, 23481.447, 20601.952, 16488.665, 14248.402, 
9928.03, 8489.895, 4800.567, 3598.98, 2048.337, 940.945, 326.4, 
80.117, 16.777, 0, 57424.3568738, 44475.4981681, 41751.7574114, 
39628.4338929, 34756.9473002, 30605.4764074, 27182.9274184, 23792.1416427, 
20723.6990619, 17056.1374644, 14058.6774471, 10585.1369045, 8102.9339783, 
5306.0245132, 3366.8557838, 2040.3836514, 963.1478214, 314.5586137, 
79.8410136, 15.502702, 0.6041464, 60272.2061248, 44780.1982856, 
41803.6541424, 40229.0292664, 35154.7682192, 30977.7459584, 27456.1890384, 
24096.7577352, 20872.8705064, 17545.6344304, 13989.6245496, 11145.817084, 
7840.7544168, 5737.6492352, 3184.2782128, 2061.6884304, 960.8241584, 
310.8441432, 79.5691376, 14.577848, 1.0774144, 62726.896388, 
45681.1355532, 42100.72506, 40473.8600572, 35598.545404, 31438.8560456, 
27758.1143968, 24396.368626, 21055.4431156, 17958.2723088, 14045.871538, 
11588.5228268, 7731.3586756, 6059.5514704, 3085.9480552, 2082.5787944, 
948.7749752, 312.0351852, 79.4346976, 13.97192, 1.4305504, 64815.5458002, 
47136.5341033, 42508.3026466, 40532.3096745, 36082.7490698, 31939.7069818, 
28091.9515352, 24693.2568243, 21274.4768803, 18299.2816212, 14223.4688239, 
11906.0609901, 7785.1704959, 6254.5961836, 3090.4472686, 2083.8345794, 
937.704655, 315.6524977, 79.5367336, 13.650846, 1.6768184), .Dim = c(21L, 
5L), .Dimnames = list(c("0", "5", "10", "15", "20", "25", "30", 
"35", "40", "45", "50", "55", "60", "65", "70", "75", "80", "85", 
"90", "95", "100"), c("1950", "1951", "1952", "1953", "1954")))
closed.out <- spragueCloseout(popmat)
colSums(closed.out) - colSums(popmat)
spragueCloseout(popmat, pivotAge = 85)
# giving a different single-age split to close out this way:
popg <- grabill(popmat)
grabill.closed.out <- spragueCloseout(popmat, popg)
# totals not necessarily preserved if mixed w Grabill
# I wouldn't recommend a rescale of the total, since the 
# only part we mess with here is the old age section. Ergo,
# one may wish to instead rescale results colSums() of 
# popg at age pivotAge and higher.
colSums(grabill.closed.out) - colSums(popmat)
}
