% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprague.R
\name{spragueExpand}
\alias{spragueExpand}
\title{create the Sprague coefficient matrix}
\usage{
spragueExpand(popmat)
}
\arguments{
\item{popmat}{numeric matrix of age-period population counts in 5-year age groups}
}
\description{
The resulting coefficient matrix is based on the number of rows in \code{popmat}
where we assume that each row of data is a 5-year age group and the final row is an open age group
to be preserved as such.
}
\details{
The \code{popmat} matrix is really just a placeholder in this case. This function is 
a utility called by the Sprague family of functions, where it is most convenient to just pass
in the same matrix being used in those calcs to determine the layout of the coefficient matrix.
}
