% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprague.R
\name{spragueSimple}
\alias{spragueSimple}
\title{the basic Sprague age-splitting method}
\usage{
spragueSimple(popmat)
}
\arguments{
\item{popmat}{a numeric matrix of population counts in 5-year age groups, with integer-labeled 
margins (age in rows and year in columns).}
}
\value{
an age-period matrix od split population counts with the same number of 
columns as \code{popmat}, and single ages in rows.
}
\description{
This method is based on the first stage of the Sprague R 
script prepared by Thomas Buettner and Patrick Gerland, itself based on the description
in Siegel and Swanson, 2004, p. 727.
}
\details{
Ages should refer to lower age bounds, ending in the open age group in the last row (not a closed terminal age). 
Dimension labelling is necessary. There must be at least six age groups (including the open group). One year of data will 
work as well, as long as it's given as a single-column matrix.
}
\examples{
p5 <- structure(c(54170.08, 44774.6, 42141.587, 38463.515, 34405.607, 
162369.816, 57424.3568738, 44475.4981681, 41751.7574114, 39628.4338929, 
34756.9473002, 164194.0485702, 60272.2061248, 44780.1982856, 
41803.6541424, 40229.0292664, 35154.7682192, 166275.9022992, 
62726.896388, 45681.1355532, 42100.72506, 40473.8600572, 35598.545404, 
168556.5331816, 64815.5458002, 47136.5341033, 42508.3026466, 
40532.3096745, 36082.7490698, 170990.473735, 66579.122, 49070.407, 
42953.604, 40534.586, 36596.844, 173545.633), .Dim = c(6L, 6L
), .Dimnames = list(seq(0,25,5), 1950:1955))
head(p5) # this is the entire matrix
p1 <- spragueSimple(p5)
head(p1); tail(p1)
colSums(p1) - colSums(p5) 
}
\references{
Shryock, H. S., Siegel, J. S., & Larmon, E. A. (1973). 
The methods and materials of demography. US Bureau of the Census.

Seigel, J. S., & Swanson, D. A. (2004). T
he methods and materials of demography. Elsevier Academic Press, London.
}
