% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sr_f_mat_five}
\alias{sr_f_mat_five}
\title{Survival rates matrix for females five year age groups between 1950 and 2045}
\format{
A matrix of dimensions 21 x 20
}
\source{
Migration residual PAS spreadhseet
}
\usage{
sr_f_mat_five
}
\description{
Survival rates matrix for females five year age groups between 1950 and 2045
for unknown country
}
\keyword{datasets}
