% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sr_m_mat}
\alias{sr_m_mat}
\title{Survival rates matrix for males five year age groups between 1950 and 2045
for unknown country}
\format{
A matrix of dimensions 21 x 20
}
\source{
Migration residual PAS spreadhseet
}
\usage{
sr_m_mat
}
\description{
Survival rates matrix for males five year age groups between 1950 and 2045
for unknown country
}
\keyword{datasets}
