% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sr_m_mat_single}
\alias{sr_m_mat_single}
\title{Survival rates matrix for males single ages between 1999 and 2019}
\format{
A matrix of dimensions 101 x 20
}
\source{
Migration residual PAS spreadsheet
}
\usage{
sr_m_mat_single
}
\description{
Survival rates matrix for males single ages between 1999 and 2019 for
Sweden
}
\keyword{datasets}
