% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srb_vec}
\alias{srb_vec}
\title{Sex ratio at birth between 1950 and 2045}
\format{
A vector of length 20
}
\source{
Migration residual PAS spreadhseet
}
\usage{
srb_vec
}
\description{
Sex ratio at birth between 1950 and 2045 for unknown country
}
\keyword{datasets}
