% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srb_vec_single}
\alias{srb_vec_single}
\title{Sex ratio at birth between 1999 and 2019}
\format{
A vector of length 20
}
\source{
Migration residual PAS spreadhseet
}
\usage{
srb_vec_single
}
\description{
Sex ratio at birth between 1999 and 2019 for Sweden
}
\keyword{datasets}
