% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENSUR.R
\name{surv5}
\alias{surv5}
\title{estimate survival curve from censuses spaced 5 years apart}
\usage{
surv5(pop1, pop2)
}
\arguments{
\item{pop1}{numeric vector population counts of census 1}

\item{pop2}{numeric vector population counts of census 2}
}
\value{
lx numeric vector of surivorship (radix 1)
}
\description{
This simple function reproduces calculations through column F of \code{CENSUR~3.XLS} 
by Griff Feeney. We assume censuses spaced 5 years apart and population counts for both censuses in 5-year age groups. 
The staggered ratio of these is like the lifetable px (1-qx). The cumulative product of this then spits back something 
we can use to approximate lx.
}
\details{
Checking for census spacing of 5 years must happen prrior to this function. 
Also, we assume the open age group has already been trimmed off. No checking done here at all.
}
\examples{
# 1965 vs 1970 pops
pop1 <- c(3983902,3854281,4513237,5373547,4572392,4206801,4110076,3751030,
3231736,2697217,2485095,2071540,1719370,1343444,955567,644043,341170)
pop2 <- c(4292503,3988292,3852101,4492096,5347327,4571868,4190340,4085338,
3674127,3198934,2648360,2382691,1970485,1584699,1172155,736258,408191)
surv5(pop1,pop2)
}
