% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGESMTH.R
\name{united_nations_smth}
\alias{united_nations_smth}
\title{The old United Nations method of population count smoothing}
\usage{
united_nations_smth(Value, Age, OAG = TRUE)
}
\arguments{
\item{Value}{numeric vector of counts in single, abridged, or 5-year age groups.}

\item{Age}{numeric vector of ages corresponding to the lower integer bound of the counts.}

\item{OAG}{logical. Whether or not the top age group is open. Default \code{TRUE}.}
}
\value{
numeric vector of smoothed counts in 5-year age groups.
}
\description{
Smooth population counts in 5-year age groups.
}
\details{
The open age group is aggregated down to be evenly divisible by 10. 
This method accounts for the youngest and oldest age groups. Age classes must be cleanly groupable to 5-year age groups.
}
\examples{
# from PASEX AGESMTH
MalePop      <- c(642367, 515520, 357831, 275542, 268336, 278601, 242515, 
		198231, 165937, 122756, 96775, 59307, 63467, 32377, 29796, 16183, 34729)
Ages         <- seq(0, 80, by = 5)
un_test <- c(NA,NA,364491,279123,268724,272228,243638,200923,162752,126304,
		91662,67432,54677,38833,NA,NA,NA)
un_result <- united_nations_smth(MalePop, Ages, TRUE)
all(round(un_result) - un_test == 0, na.rm = TRUE)
\dontrun{
plot(Ages, MalePop)
lines(as.integer(names(un_result)),un_result)
}
}
\references{
Carrier, Norman H., and A. M. Farrag. "The reduction of errors 
in census populations for statistically underdeveloped countries." 
Population Studies 12.3 (1959): 240-285.
}
