% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Whipple.R
\name{Whipple}
\alias{Whipple}
\title{calculate Whipple's index of age heaping}
\usage{
Whipple(Value, Age, ageMin = 25, ageMax = 65, digit = c(0, 5))
}
\arguments{
\item{Value}{a vector of demographic counts by single age}

\item{Age}{a vector of ages corresponding to the lower integer bound of the counts}

\item{ageMin}{the lowest age included in calcs. Default 25}

\item{ageMax}{the upper age bound used for calcs. Default 65}

\item{digit}{any digit 0-9. Default \code{c(0,5)}. Otherwise it needs to be a single digit.}
}
\description{
Implementation following the PASEX spreadsheet SINGAGE, with some extra options for more digit checking.
}
\details{
\code{ageMax} is the hard upper bound, treated as interval. If you want ages
20 to 89, then give \code{ageMin = 20} and \code{ageMax = 90}, not 89. You can get 
arbitrary W(i) indicies by specifying other digits. Note you can only do pairs of digits
if it's 0,5. Otherwise just one digit at a time.
}
