% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZELNIK.R
\name{zelnik}
\alias{zelnik}
\title{Zelnik 11-term moving average. Adjusting for digit preference}
\usage{
zelnik(Value, q, Age)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts in single age groups.}

\item{q}{to be defined}

\item{Age}{integer lower bound of age classes}
}
\value{
a named vector with the adjusted values
}
\description{
Zelnik method is used to adjust distributions by year of age for digit preference.This comes from 
Gray_1987_The Missing Ages: Adjusting for Digit Preference
}
\details{
Single year age groups are assumed.
}
\examples{
# data from gray1987missingages, Table 2, page 21: Aplication of Q1 and 
# Q2 linear operators, Bangladesh Census, 1 March 1974-Males.
Pop<-c(941307,1041335,1237034,1411359,1383853,1541942,1321576,1285877,1563448,886705,
      1623998,562924,1485173,543216,771219,903496,686431,370007,942999,250820,
      1023667,200131,688640,222011,281738,1239965,288363,263326,483143,78635,
      1349886,68438,415127,101596,100758,1392434,178633,126351,286520,50836,
      1331036,48995,251153,58393,54995,1033812,68792,72766,175943,28254,
      1038747,32894,136179,37667,38230,596049,52602,36493,74106,16759,
      790643,20596,70109,18044,19891,357491,15253,17489,31057,8481,
      429816,7951,35583,8612,6589,454645)
Age  <-c(0:75) 
zelnik(Pop,1,Age)
}
\references{
\insertRef{gray1987missingages}{DemoTools}
}
