% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGZAG.R
\name{zigzag_min}
\alias{zigzag_min}
\title{Objective function to minimize Feeney's zigzag method residual}
\usage{
zigzag_min(Value, Age, ageMin = 40, ageMax = 80, p)
}
\arguments{
\item{Value}{numeric vector of (presumably) counts in 5-year age groups.}

\item{Age}{integer vector of age group lower bounds.}

\item{ageMin}{integer. Lower age bound to adjust values.}

\item{ageMax}{integer. Upper age bound to adjust values.}

\item{p}{numeric vector of adjustment parameters.}
}
\value{
positive residual to minimize.
}
\description{
This function is auxiliary to \code{zigzag()}, see \code{?zigzag} for a description.
}
\details{
This function is not intended to be used at the top level, but just in case, make sure that \code{ageMax = ageMin + 10 * length(p)}. Age groups \code{ >= ageMin} AND \code{ <= ageMin} must be in 5-year age groups. This function does no checks.
}
\examples{
Value <- c(13331,4151,1746,1585,3859,8354,11146,12076,
	12216,12016,12473,11513,12899,11413,12710,11516,11408,6733,4031,2069)
Age <- c(0,1,seq(5,90,by=5))
zigzag_min(Value,Age,ageMin=40,ageMax = 80,p=rep(.05,4))
# it's used like this in zigzag()
(p <- optim(
		rep(.05,4), 
		zigzag_min, 
		Value = Value, 
		Age = Age, 
		ageMin = 40, 
		ageMax = 80)$par)
Smoothed <- zigzag_p(Value,Age,40,80,p)
# de facto unit test:
# check result using results frozen in Feeney spreadsheet
# after fixing probable cell range 'error'
p.feeney <- c(0.0235802695087692,0.0724286618207911,
		      0.0242327829742267,0.0883411499065237)
ans      <- 106.1147411629
stopifnot(abs(zigzag_min(Value, Age, 40,80,p.feeney) - ans) < 1e-6)
}
\references{
Feeney, G. 2013 "Removing "Zigzag" from Age Data," http://demographer.com/white-papers/2013-removing-zigzag-from-age-data/
}
