% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abacus.R
\name{AbacusLIFTB}
\alias{AbacusLIFTB}
\title{Literal translation of mortpak from Fortran.}
\usage{
AbacusLIFTB(NFIN, NUMOUT, NTYPE, NSEX, QXMX)
}
\arguments{
\item{NFIN}{integer. The open age.}

\item{NUMOUT}{integer. The desired open age group (under 100).}

\item{NTYPE}{integer. 1 means nqx. 2 means nMx. 3 means nMx but do not extrapolate to close out.}

\item{NSEX}{integer. 1 is male. 2 is female.}

\item{QXMX}{numeric. Vector of either nMx or nqx in abridged ages.}
}
\description{
This has not been modified in any meaningful way by the package maintainer, nor will it be documented in depth, 
as it is intended to be replaced in the future with a refactored version of the same.
}
\details{
Try \code{AbacusLIFTB_wrap()} if you just want to make quick use of this functionality. Standard abridged ages are assumed. 
The lifetable stops at 100+. Note \code{QXMX} needs to be an object in single ages
}
\references{
\insertRef{mortpak1988}{DemoTools}
}
