% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abacus.R
\name{AbacusLIFTB_wrap}
\alias{AbacusLIFTB_wrap}
\title{wrapper to a rote translation of the Abacus version LIFTB}
\usage{
AbacusLIFTB_wrap(Mx, qx, mx_ind = TRUE, OAnew = 100, Sex = "m")
}
\arguments{
\item{Mx}{numeric vector of Mx in standard abridged age classes}

\item{qx}{numeric vector of qx in standard abridged age classes}

\item{mx_ind}{logical indicator of whether Mx or qx is given/preferred to be used}

\item{OAnew}{the desired open age group, max 100}

\item{Sex}{\code{"m"} or \code{"f"}. Anything not \code{"m"} treated as \code{"f"}}
}
\description{
A minimal argument wrapper that calls a minimally modified \code{AbacusLIFTB()}, an R transaltion of by UN staff of the Mortpak Fortran code. A wrapper is necessary because \code{AbacusLIFTB()} abides by some non-conventional indexing. We also take care of dimension naming.
}
\details{
The wrapper itself does little argument checking, but \code{AbacusLIFTB()} itself has nuanced argument checking and error reporting. Such errors rarely stop the function. Instead they are caught and returned. The wrapper returns such messages as warnings, but does not pass them on as objects. This function is used internally by \code{LTabr()}. Although this is a wrapper, it is not intended to be used directly.
}
\references{
\insertRef{mortpak1988}{DemoTools}
}
