% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{age_abridge_force}
\alias{age_abridge_force}
\title{force a (count) vector to abridged ages}
\usage{
age_abridge_force(Value, AgeInt, Age)
}
\description{
This is a robustness utility, in place to avoid annoying hang-ups in \code{LTAbr()}. If data are given in non-standard ages, they are forced to standard abrdiged ages on the fly. Really this should happen prior to calling \code{LTAbr()}
}
\details{
This should be able to group up and group down as needed. \code{splitMono()} is used below the hood. \code{pclm()} or \code{splitUniform()} out to be flexible enough to do the same.
}
\examples{
V1        <- pop1m_ind
Age       <- c(0,1,3,seq(5,100,5))
AgeInt    <- c(1,2,2,rep(5,19),1)
Value     <- tapply(V1,rep(Age,times=AgeInt), sum)

is_abridged(Age)
age_abridge_force(Value, AgeInt, Age)
}
\seealso{
splitMono, LTAbr
}
