% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{AGEN}
\alias{AGEN}
\title{Trim two age vectors to matching N-year age groups.}
\usage{
AGEN(Age1, Age2, N = 5, consecutive = TRUE, ageMin = 0,
  ageMax = max(c(Age1, Age2)))
}
\arguments{
\item{Age1}{integer. Vector of first age groups (lower bounds).}

\item{Age2}{integer. Vector of second age groups (lower bounds).}

\item{N}{integer. Target age group interval (just one number).}

\item{consecutive}{logical.  Whether or not to throw error if resulting age groups not consecutive. Default \code{TRUE}.}

\item{ageMin}{integer. Optional lower age bound for output.}

\item{ageMax}{integer. Optional upper age bound for output.}
}
\value{
Integer vector of the N-year age groups present in both \code{Age1} and \code{Age2}.
}
\description{
Determining which N-year (e.g. 5-year) age groups two vectors have in common
is helpful for matching vector lengths, and questions of alignment. Used as a utility throughout.
}
\examples{
Age1 <- seq(0, 100, by = 5)
Age2 <- 0:80
AGEN(Age1, Age2, N = 5)
AGEN(Age1, Age2, N = 5, ageMax = 70)
}
