% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGESEX.R
\name{AHI}
\alias{AHI}
\title{Kannisto's age heaping index}
\usage{
AHI(Value, Age, Agei = 90)
}
\arguments{
\item{Value}{a vector of demographic counts or rates by single age}

\item{Age}{a vector of ages corresponding to the lower integer bound of the counts}

\item{Agei}{the age on which the index is centered.}
}
\value{
the index value.
}
\description{
This age heaping index is used for particular old ages, such as 90, 95, 100, 105, and so forth.
}
\details{
The index looks down two ages and up two ages, so the data must accomodate that range.
}
\examples{
Value <- c(80626,95823,104315,115813,100796,105086,97266,116328,
75984,89982,95525,56973,78767,65672,53438,85014,
47600,64363,42195,42262,73221,30080,34391,29072,
20531,66171,24029,44227,24128,23599,82088,16454,
22628,17108,12531,57325,17220,28425,16206,17532,
65976,11593,15828,13541,8133,44696,11165,18543,
12614,12041,55798,9324,10772,10453,6773,28358,
9916,13348,8039,7583,42470,5288,5317,6582,
3361,17949,3650,5873,3279,3336,27368,1965,
2495,2319,1335,12022,1401,1668,1360,1185,
9167,424,568,462,282,6206,343,409,333,291,4137,133,169,157,89,2068,68,81,66,57)
Age <- 0:99
AHI(Value, Age, 90)
AHI(Value, Age, 95)
}
\references{
Kannisto (1999) Assessing the Information on Age at Death of Old Persons in National Vital Statistics. 
Validation of Exceptional Longevity. [Monographs on Population Aging. Vol. 6]
}
