% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAx.R
\name{ax.greville.mortpak}
\alias{ax.greville.mortpak}
\title{UN version of the Greville formula for a(x) from M(x)}
\usage{
ax.greville.mortpak(nMx, nqx, lx, IMR = NA, Sex = "m", region = "W",
  mod = TRUE)
}
\arguments{
\item{nMx}{numeric vector. Mortality rates in standard abridged age groups.}

\item{nqx}{numeric vector. Age specific death probabilities in standard abridged age groups.}

\item{IMR}{numeric infant death probability. Optional if using \code{nMx}, not required otherwise.}

\item{Sex}{character. \code{"m"}, \code{"f"} or \code{"b"} for male, female, or both.}

\item{region}{character. \code{"n"}, \code{"e"}, \code{"s"} or \code{"w"} for North, East, South, or West.}

\item{mod}{}

\item{nlx}{numeric vector. Lifetable survivorship in standard abridged age groups.}
}
\value{
ax numeric vector of a(x) the average time spent in the interval of those that die in the interval.
}
\description{
The UN ax formula uses Coale-Demeny for ages 0, and 1-4, values of 2.5 
for ages 5-9 and 10-14, and the Greville formula thereafter. In the original sources 
these are referred to as separation factors.
}
\details{
ax for age 0 and age group 1-4 are based on Coale-Demeny q0-based lookup tables.
We use an approximation to get from m0 to q0 for the sake of generating a0 and 4a1. 
The final ax value is closed out as if the final Mx value were constant thereafter, 
which is a common lifetable closeout choice. Age groups must be standard abridged, 
and we do not check age groups here!
}
\references{
\insertRef{greville1977short}{DemoTools}
\insertRef{un1982model}{DemoTools}
\insertRef{arriaga1994population}{DemoTools}
\insertref{mortpak1988}{DemoTools}
}
