% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bachi.R
\name{Bachi}
\alias{Bachi}
\title{calculate Bachi's index of age heaping}
\usage{
Bachi(Value, Age, ageMin = 30, ageMax = 80, pasex = FALSE)
}
\arguments{
\item{Value}{a vector of demographic counts by single age}

\item{Age}{a vector of ages corresponding to the lower integer bound of the counts}

\item{ageMin}{the lowest age included in calcs. Default 30}

\item{ageMax}{the upper age bound used for calcs. Default 80}

\item{pasex}{logical default \code{FALSE}. Do we want to reproduce the specific age weightings in the PASEX spreadsheet?}
}
\description{
Two Implementations: one following the PASEX spreadsheet SINGAGE (\code{pasex = TRUE}), with ages hard-coded, and another with flexible upper and lower age bounds, but that does not match the PASEX implementation.
}
\details{
\code{ageMax} is the hard upper bound, treated as interval. If you want ages
20 to 89, then give \code{ageMin = 20} and \code{ageMax = 90}, not 89. These are only heeded if \code{pasex = FALSE}.
}
