% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepop.R
\name{basepop_five}
\alias{basepop_five}
\alias{basepop_single}
\title{BPA and BPE methods for adjusting age groups under 10}
\usage{
basepop_five(
  refYear,
  Females_five,
  SRB,
  nlxFemale,
  nlxDatesIn,
  asfrMat,
  AsfrDatesIn,
  ...,
  Sex = c("f", "m"),
  SmoothedFemales = NULL,
  Males_five = NULL,
  nlxMale = NULL
)

basepop_single(
  refYear,
  Females_single,
  SRB,
  nlxFemale,
  nlxDatesIn,
  asfrMat,
  AsfrDatesIn,
  ...,
  Sex = c("f", "m"),
  SmoothedFemales = NULL,
  Males_single = NULL,
  nlxMale = NULL
)
}
\arguments{
\item{refYear}{An integer. The decimal reference year for which the reported population pertain}

\item{Females_five}{A named numeric vector. The names of the vector should reflect the age groups. See examples. Reported population by abridged 5-year lower bound age groups for females pertaining to \code{refYear}. This means that age 0 must be kept as a single age, ages 1-4 must grouped together as abridged age 1, and thereafter 5-year age groups.}

\item{SRB}{A numeric. Sex ratio at birth (males / females)}

\item{nlxFemale}{A numeric matrix. The female nLx function of two life tables with ages in the rows and time in columns. The earlier date should be at least 7.5 years before the reference date of the "reported" population. The later date should be no earlier than one-half year before the reference date of the "reported" population}

\item{nlxDatesIn}{Vector of dates. The dates which pertain to the columns in \code{nlxFemale} and \code{nlxMale}. For details on the format of the dates, see the details of \code{\link{interp}}}

\item{asfrMat}{A numeric matrix. An age-period matrix of age specific fertility rates with age in rows, time in columns.}

\item{AsfrDatesIn}{A vector of dates. The dates which pertain to the columns in \code{asfrMat}. For details on the format of the dates, see the details of \code{\link{interp}}}

\item{...}{Arguments passed to \code{\link{interp}}. In particular, users might be interested in changing the interpolation method. By default, it's linearly interpolated}

\item{Sex}{A character string. Either 'f' or 'm', for adjusting either \code{Females} or \code{Males}. See the details section for the mandatory arguments for adjusting the male counts.}

\item{SmoothedFemales}{A numeric vector. This should be the result of passing the \code{Females} arguments through \code{smooth_age_5}. It is assumed that this argument is a numeric vector with the lower bound age group definitions as names. By default, this is what \code{smooth_age_5} returns. See the example section for an example.}

\item{Males_five}{A named numeric vector. The names of the vector should reflect the age groups. See examples. Reported population by abridged 5-year lower bound age groups for males pertaining to \code{refYear}. This means that age 0 must be kept as a single age, ages 1-4 must grouped together as abridged age 1, and thereafter 5-year age groups. This argument is only used when \code{Sex} is set to \code{"m"}.}

\item{nlxMale}{A numeric matrix. The male nLx function of two life tables with ages in the rows and time in columns. The male nLx functions should be for age groups 0, 1 to 4, and 5 to 9. The dates which are represented in the columns are assumed to be the same as \code{nlxDatesIn}. This argument is only used when \code{Sex} is set to \code{"m"}}

\item{Females_single}{A named numeric vector. Reported population by 1-year age groups for \code{refYear} for females. The names of the vector should reflect the age groups. See examples. The method assumes that the last age group is open (for example, the population ends at '80+' and '100+')}

\item{Males_single}{A named numeric vector. Reported population by 1-year age groups format \code{refYear} for males. The names of the vector should reflect the age groups. See examples. The method assumes that the last age group is open (for example, the population ends at '80+' and '100+')}
}
\value{
When using \code{basepop_five}, a numeric vector of adjusted counts
for the age groups 0, 1-4 and 5-9 for either \code{Females} or \code{Males},
depending on whether the user specified \code{"m"} or \code{"f"} in
\code{Sex}. The remaining age counts are left as is. When
\code{basepop_single} is used, the return value issues a numeric vector with
\strong{single year age groups} where the counts between 0 and 10 are adjusted.
}
\description{
Adjust population counts for the age groups 0, 1-4 and 5-9
}
\details{
\code{basepop_five} and \code{basepop_single} can estimate both the BPA and
BPE methods. If the user specifies \code{SmoothedFemales}, both \code{basepop_*}
functions will return the BPA method. If \code{SmoothedFemales} is left empty,
both \code{basepop_*} functions will adjust using the BPE method.

For \code{basepop_five}, adjusting the female population counts is the
default. For this, all arguments except \code{Males_five} and
\code{nlxMale} are needed. However, for adjusting the male population
counts, the user needs to specify all arguments as well as \code{Males_five},
\code{nlxMale} and \code{nlxFemale} and set \code{sex == "m"}.

For \code{basepop_single}, the same procedure applies. The only difference
is that the vector \code{Males_five} is named \code{Males_single} and accepts
a vector of single ages rather than five year abridged age groups. Similarly,
the vector for females is \code{Females_single} rather than
\code{Females_five} and accepts single age groups.

Currently, \code{basepop_five} works only with five year abridged age groups
while \code{basepop_single} works only with single year age groups.
}
\section{BPA}{


Description:

The method estimates a smoothed population ages 10 and over and adjusts
the population under age 10 using the smoothed population and estimates
of fertility and mortality.

Based on the smoothed female population counts, it rejuvenates the female
"reported" population 20 to 59 years of age for the two 5 year periods prior
to the census date to represent the female population in reproductive ages
5 and 10 years earlier.  Based on the rejuvenated population and fertility
and mortality levels, the method then estimates the male and female births
during the two 5 year periods prior to the census date. Next, it projects
the two 5-year birth cohorts to the census date. The projected figures
represent the adjusted population ages 0 to 4 years and 5 to 9 years
at the census date.

Advantages:

(1) The method adjusts under-10 population to be consistent with fertility
and mortality levels and adjusted adult female population.

Limitations:

(1) BPA assumes a linear change in fertility and mortality during the decade
prior to the reference year.

(2) The procedure ignores migration, which can lead to misleading results.
There are two issues. First, age groups 0-4 and 5-9 are subject to migration,
which would affect the comparability of estimated and reported populations
in the base year.  Second, the estimated size of age groups 0-4 and 5-9 are
calculated from numbers of women of reproductive age in the base year
rejuvenated to points in the past.  With migration, rejuvenated number of
women may exceed or be smaller than the number present, and giving
birth to children, in the decade prior to the base year.

(3) BPA’s smoothing calculations may mask unusual, but real, variations
in population age group size.  Smoothing irregularities in age structure
not attributable to age misreporting will distort estimated births and
survived children in the base year.

Assumptions:

(1) No significant international migration took place within the
reference periods for the population, mortality, and fertility input.

(2) The data input as the "reported" population is not affected by
underenumeration of persons in certain ages, nor by age misreporting.
}

\section{BPE}{


Description:

The method adjusts the population under age 10 using the reported population
ages 10 and above and estimates of fertility and mortality.

The method rejuvenates the reported female population 20 to 59 years of age
for the two 5 year periods prior to the census date to represent the female
population in reproductive ages 5 and 10 years earlier. Based on the
rejuvenated population and fertility and mortality levels, the method then
estimates the male and female births during the two 5 year periods prior to
the census date.  Next, it projects the two 5-year birth cohorts to the
census date.  The projected figures represent the adjusted population ages
0 to 4 years and 5 to 9 years at the census date.

Advantages:

(1) The method adjusts the under-10 population to be consistent with
fertility and mortality levels and adult female population.

Limitations:

(1) BPE assumes a linear change in fertility and mortality during the decade
prior to the reference year.

(2) The procedure ignores migration, which can lead to misleading results.
There are two issues.  First, age groups 0-4 and 5-9 are subject to
migration, which would affect the comparability of estimated and reported
populations in the base year.  Second, the estimated size of age groups
0-4 and 5-9 are calculated from numbers of women of reproductive age in
the base year rejuvenated to points in the past.  With migration, rejuvenated
number of women may exceed or be smaller than the number present, and
giving birth to children, in the decade prior to the base year.

(3) The method does not adjust for possible underenumeration and age
misreporting errors in the over-10 “reported” population. If the
reported population is subject to age-misreporting or age-sex-specific
underenumeration, the over-10 population should be smoothed or otherwise
corrected prior to use.

Assumptions:

(1) No significant international migration took place within the reference
periods for the population, mortality, and fertility input.

(2) The data input as the “reported” population is not affected by
underenumeration of persons in certain ages, nor by age misreporting.
}

\examples{

 \dontrun{
 # (1) RefYear
 refYear <- 1986.21

 # (2) Reported population by 5-year age groups and sex in the base year
 # (Include unknowns).

 pop_male_counts <- c(`0` = 11684,
                      `1` = 46738,
                      `5` = 55639,
                      `10` = 37514,
                      `15` = 29398,
                      `20` = 27187,
                      `25` = 27770,
                      `30` = 20920,
                      `35` = 16973,
                      `40` = 14999,
                      `45` = 11330,
                      `50` = 10415,
                      `55` = 6164,
                      `60` = 7330,
                      `65` = 3882,
                      `70` = 3882,
                      `75` = 1840,
                      `80` = 4200
                      )

 pop_female_counts <- c(`0` = 11673,
                        `1` = 46693,
                        `5` = 55812,
                        `10` = 35268,
                        `15` = 33672,
                        `20` = 31352,
                        `25` = 33038,
                        `30` = 24029,
                        `35` = 16120,
                        `40` = 14679,
                        `45` = 8831,
                        `50` = 9289,
                        `55` = 4172,
                        `60` = 6174,
                        `65` = 2715,
                        `70` = 3344,
                        `75` = 1455,
                        `80` = 4143)

 # (4) Sex ratio at birth (m/f)
 sex_ratio <- 1.0300

 # (6) The male and female nLx functions for ages under 1 year, 1 to 4 years, and 5 to 9
 # years, pertaining to an earlier and later date
 nlxDatesIn <- c(1977.31, 1986.50)

 nlxMale <- matrix(c(87732,
                     304435,
                     361064,
                     88451,
                     310605,
                     370362
                     ),
                   nrow = 3, ncol = 2)
 
 nlxFemale <- matrix(c(89842,
                       314521,
                       372681,
                       353053,
                       340650,
                       326588,
                       311481,
                       295396,
                       278646,
                       261260,
                       241395,
                       217419,
                       90478,
                       320755,
                       382531,
                       364776, 
                       353538, 
                       340687, 
                       326701, 
                       311573, 
                       295501, 
                       278494, 
                       258748,
                       234587),
                     nrow = 12,
                     ncol = 2)
 
 # (7) A set of age-specific fertility rates pertaining to an earlier and later
 # date
 
 asfrmat <- matrix(c("15-19" = 0.2000,
                     "20-24" = 0.3000,
                     "25-29" = 0.3000, 
                     "30-34" = 0.2500, 
                     "35-39" = 0.2000, 
                     "40-44" = 0.1500, 
                     "45-49" = 0.0500,
                     "15-19" = 0.1500,
                     "20-24" = 0.2000,
                     "25-29" = 0.2750, 
                     "30-34" = 0.2250, 
                     "35-39" = 0.1750, 
                     "40-44" = 0.1250, 
                     "45-49" = 0.0500),
                   nrow = 7,
                   ncol = 2)
 
 AsfrDatesIn <- c(1977.81, 1985.71)
 
 smoothed_females <- smooth_age_5(Value = pop_female_counts,
                                  Age = as.numeric(names(pop_female_counts)),
                                  method = "Arriaga",
                                  OAG = TRUE,
                                  young.tail = "Original")
 
 ## This is the only number that messes up the whole calculation.
 ## smooth_age_5 returns the same result as the PASS excel sheet
 ## except for the age groups 10-15 and 15-19. Here we only use
 ## age group 15-19. If we plug in manually the correct value,
 ## we get all results match exactly, otherwise there are
 ## some differences.
 smoothed_females[4] <- 34721

 # For adjusting using BPA for males, we need to specify
 # Sex = "m" with Males and nlxMale.
 bpa_male <-
   basepop_five(
     refYear = refYear,
     Males_five = pop_male_counts,
     Females_five = pop_female_counts,
     SmoothedFemales = smoothed_females,
     SRB = sex_ratio,
     nlxFemale = nlxFemale,
     nlxMale = nlxMale,
     nlxDatesIn = nlxDatesIn,
     asfrMat = asfrmat,
     AsfrDatesIn = AsfrDatesIn,
     Sex = "m"
     )

 # See adjustments?
 pop_male_counts[1:3]
 bpa_male[1:3]

 # Adjusting the BPA for females requires less arguments
 
 bpa_female <-
   basepop_five(
     refYear = refYear,
     Females_five = pop_female_counts,
     SmoothedFemales = smoothed_females,
     SRB = sex_ratio,
     nlxFemale = nlxFemale,
     nlxDatesIn = nlxDatesIn,
     asfrMat = asfrmat,
     AsfrDatesIn = AsfrDatesIn,
     Sex = "f"
   )

 pop_female_counts[1:3]
 bpa_female[1:3]

 # For adjustment using BPE, we use exactly the same definitions as above
 # but remove SmoothedFemales.
 bpe_male <-
   basepop_five(
     refYear = refYear,
     Males_five = pop_male_counts,
     Females_five = pop_female_counts,
     SRB = sex_ratio,
     nlxFemale = nlxFemale,
     nlxMale = nlxMale,
     nlxDatesIn = nlxDatesIn,
     asfrMat = asfrmat,
     AsfrDatesIn = AsfrDatesIn,
     Sex = "m"
     )

 # See adjustments?
 pop_male_counts[1:3]
 bpa_male[1:3]
 bpe_male[1:3]

 # Adjusting the BPA for females requires less arguments
 
 bpe_female <-
   basepop_five(
     refYear = refYear,
     Females_five = pop_female_counts,
     SRB = sex_ratio,
     nlxFemale = nlxFemale,
     nlxDatesIn = nlxDatesIn,
     asfrMat = asfrmat,
     AsfrDatesIn = AsfrDatesIn,
     Sex = "f"
   )

 pop_female_counts[1:3]
 bpa_female[1:3]
 bpe_female[1:3]

# basepop_single for single ages
# Single ages for males and females

pop_male_counts <-
  c(
    `0` = 11684,
    `1` = 11473,
    `2` = 11647,
    `3` = 11939,
    `4` = 11680,
    `5` = 10600,
    `6` = 11100,
    `7` = 11157,
    `8` = 11238,
    `9` = 11544,
    `10` = 7216,
    `11` = 7407,
    `12` = 7461,
    `13` = 7656,
    `14` = 7774,
    `15` = 5709,
    `16` = 5629,
    `17` = 5745,
    `18` = 6056,
    `19` = 6259,
    `20` = 5303,
    `21` = 5423,
    `22` = 5497,
    `23` = 5547,
    `24` = 5417,
    `25` = 5441,
    `26` = 5466,
    `27` = 5500,
    `28` = 5668,
    `29` = 5694,
    `30` = 4365,
    `31` = 4252,
    `32` = 4122,
    `33` = 4142,
    `34` = 4039,
    `35` = 3210,
    `36` = 3222,
    `37` = 3258,
    `38` = 3413,
    `39` = 3871,
    `40` = 2684,
    `41` = 2844,
    `42` = 3052,
    `43` = 3182,
    `44` = 3237,
    `45` = 2263,
    `46` = 2298,
    `47` = 2318,
    `48` = 2257,
    `49` = 2194,
    `50` = 2231,
    `51` = 2172,
    `52` = 2072,
    `53` = 2008,
    `54` = 1932,
    `55` = 1301,
    `56` = 1262,
    `57` = 1213,
    `58` = 1197,
    `59` = 1191,
    `60` = 1601,
    `61` = 1593,
    `62` = 1490,
    `63` = 1348,
    `64` = 1299,
    `65` = 568,
    `66` = 745,
    `67` = 843,
    `68` = 801,
    `69` = 925,
    `70` = 806,
    `71` = 883,
    `72` = 796,
    `73` = 725,
    `74` = 672,
    `75` = 470,
    `76` = 441,
    `77` = 340,
    `78` = 300,
    `79` = 289,
    `80` = 4200
  )
pop_female_counts <-
  c(
    `0` = 11673,
    `1` = 11474,
    `2` = 11670,
    `3` = 11934,
    `4` = 11614,
    `5` = 10603,
    `6` = 11144,
    `7` = 11179,
    `8` = 11269,
    `9` = 11617,
    `10` = 6772,
    `11` = 6948,
    `12` = 7030,
    `13` = 7211,
    `14` = 7306,
    `15` = 6531,
    `16` = 6443,
    `17` = 6535,
    `18` = 6951,
    `19` = 7213,
    `20` = 6096,
    `21` = 6234,
    `22` = 6327,
    `23` = 6410,
    `24` = 6285,
    `25` = 6464,
    `26` = 6492,
    `27` = 6549,
    `28` = 6739,
    `29` = 6795,
    `30` = 5013,
    `31` = 4888,
    `32` = 4735,
    `33` = 4747,
    `34` = 4646,
    `35` = 3040,
    `36` = 3068,
    `37` = 3107,
    `38` = 3246,
    `39` = 3658,
    `40` = 2650,
    `41` = 2788,
    `42` = 2977,
    `43` = 3108,
    `44` = 3156,
    `45` = 1756,
    `46` = 1784,
    `47` = 1802,
    `48` = 1764,
    `49` = 1724,
    `50` = 1982,
    `51` = 1935,
    `52` = 1846,
    `53` = 1795,
    `54` = 1731,
    `55` = 863,
    `56` = 850,
    `57` = 825,
    `58` = 819,
    `59` = 816,
    `60` = 1348,
    `61` = 1342,
    `62` = 1246,
    `63` = 1138,
    `64` = 1101,
    `65` = 391,
    `66` = 520,
    `67` = 585,
    `68` = 560,
    `69` = 659,
    `70` = 670,
    `71` = 750,
    `72` = 686,
    `73` = 634,
    `74` = 604,
    `75` = 353,
    `76` = 340,
    `77` = 270,
    `78` = 246,
    `79` = 247,
    `80` = 4143
  )

 smoothed_females <- smooth_age_5(Value = pop_female_counts,
                                  Age = as.numeric(names(pop_female_counts)),
                                  method = "Arriaga",
                                  OAG = TRUE,
                                  young.tail = "Original")


 # For adjusting using BPA for males, we need to specify
 # Sex = "m" with Males and nlxMale.
 bpa_male <-
   basepop_single(
     refYear = refYear,
     Males_single = pop_male_counts,
     Females_single = pop_female_counts,
     SmoothedFemales = smoothed_females,
     SRB = sex_ratio,
     nlxFemale = nlxFemale,
     nlxMale = nlxMale,
     nlxDatesIn = nlxDatesIn,
     asfrMat = asfrmat,
     AsfrDatesIn = AsfrDatesIn,
     Sex = "m"
     )

 # See adjustments?
 pop_male_counts[1:10]
 bpa_male[1:10]

 # Adjusting the BPA for females requires less arguments
 bpa_female <-
   basepop_single(
     refYear = refYear,
     Females_single = pop_female_counts,
     SmoothedFemales = smoothed_females,
     SRB = sex_ratio,
     nlxFemale = nlxFemale,
     nlxDatesIn = nlxDatesIn,
     asfrMat = asfrmat,
     AsfrDatesIn = AsfrDatesIn,
     Sex = "f"
   )

 pop_female_counts[1:10]
 bpa_female[1:10]

 # For adjustment using BPE, we use exactly the same definitions as above
 # but remove SmoothedFemales.
 bpe_male <-
   basepop_single(
     refYear = refYear,
     Males_single = pop_male_counts,
     Females_single = pop_female_counts,
     SRB = sex_ratio,
     nlxFemale = nlxFemale,
     nlxMale = nlxMale,
     nlxDatesIn = nlxDatesIn,
     asfrMat = asfrmat,
     AsfrDatesIn = AsfrDatesIn,
     Sex = "m"
     )

 # See adjustments?
 pop_male_counts[1:10]
 bpa_male[1:10]
 bpe_male[1:10]

 # Adjusting the BPA for females requires less arguments
 bpe_female <-
   basepop_single(
     refYear = refYear,
     Females_single = pop_female_counts,
     SRB = sex_ratio,
     nlxFemale = nlxFemale,
     nlxDatesIn = nlxDatesIn,
     asfrMat = asfrmat,
     AsfrDatesIn = AsfrDatesIn,
     Sex = "f"
   )

 pop_female_counts[1:10]
 bpa_female[1:10]
 bpe_female[1:10]

 }

}
\references{
\insertRef{arriaga1994population}{DemoTools}
\insertRef{PAS}{DemoTools}
}
