% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beers.R
\name{beers}
\alias{beers}
\title{ord1 <-  beers(Value, OAG = TRUE, method = "ord")
mod1 <- beers(Value, OAG = TRUE, method = "mod")
\dontrun{
plot(Age,Value)
lines(Age, ord1)
lines(Age, mod1)
}}
\usage{
beers(popmat, Age, OAG = FALSE, method = "mod", johnson = FALSE, Age0)
}
\description{
# notice this negative value. Yuck! 
tail(mod1)
# this replaces ages 90+, guaranteed no negatives.
monoCloseout(Value, pops = mod1, OAG = TRUE)
# Note: there are no kludges built into beers() to handle such cases.
# these ought to be handled by wrappers as appropriate.
}
\details{
# This replicates Johnson_2016_BEERSP.XLS, males
M <- c(752124,582662,463534,369976,286946,235867,
		199561,172133,151194,131502,113439,95614,
		78777,60157,40960,21318,25451)
dim(M)      <- c(length(M),1)
rownames(M) <- seq(0,80,by=5)
Age0        <- 184499
johnson     <- beers(
		         popmat = M, 
		         OAG = TRUE,
			     method = "ord", 
			     johnson = TRUE, 
			     Age0 = Age0)
# copied from spreadsheet output, a de facto unit test, 
output <- c(184499,158163,143895,135416,130151,
            126153,122028,117182,111566,105733,
			101088,96719,92551)
# allow for rounding differences, so maximum absolute 
# difference of 1
stopifnot(max(abs(round(johnson)[1:length(output)] - output)) == 1)
}
