% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beers.R
\name{beersExpand}
\alias{beersExpand}
\title{Create the Beers ordinary or modified coefficient matrix}
\usage{
beersExpand(popmat, OAG = FALSE, method = "Mod")
}
\arguments{
\item{popmat}{numeric. Matrix of age-period population counts in 5-year age groups.}

\item{OAG}{logical. Whether or not the final age group open. Default \code{TRUE}.}

\item{method}{character. Valid values are \code{"mod"} or \code{"ord"}. Default \code{"mod"}.}
}
\description{
The resulting coefficient matrix is based on the number of rows in \code{popmat}
which must be in 5-year age groups (not abridged). The final row may be an open 
or closed age group, as indicated by the \code{OAG} argument.
}
\details{
The \code{popmat} matrix is really just a placeholder in this case. This function is 
a utility called by the Beers family of functions, where it is most convenient to just pass
in the same matrix being used in those calculations to determine the layout of the coefficient matrix.
}
\examples{
popmat <- structure(c(54170, 44775, 42142, 38464, 34406, 30386, 26933, 
				23481, 20602, 16489, 14248, 9928, 8490, 4801, 3599, 2048, 941, 
				326, 80, 17, 0, 57424, 44475, 41752, 39628, 34757, 30605, 27183, 
				23792, 20724, 17056, 14059, 10585, 8103, 5306, 3367, 2040, 963, 
				315, 80, 16, 1, 60272, 44780, 41804, 40229, 35155, 30978, 27456, 
				24097, 20873, 17546, 13990, 11146, 7841, 5738, 3184, 2062, 961, 
				311, 80, 15, 1, 62727, 45681, 42101, 40474, 35599, 31439, 27758, 
				24396, 21055, 17958, 14046, 11589, 7731, 6060, 3086, 2083, 949, 
				312, 79, 14, 1, 64816, 47137, 42508, 40532, 36083, 31940, 28092, 
				24693, 21274, 18299, 14223, 11906, 7785, 6255, 3090, 2084, 938, 
				316, 80, 14, 2), 
		.Dim = c(21L, 5L), 
		.Dimnames = list(seq(0,100,by=5), 1950:1954))

coefsOA     <- beersExpand(popmat, OAG = TRUE, method = "mod")
coefsclosed <- beersExpand(popmat, OAG = FALSE, method = "mod")
dim(beersExpand(popmat, TRUE))
dim(beersExpand(popmat, FALSE))
coefso      <- beersExpand(popmat, OAG = TRUE, method = "ord")

# how to use (under the hood in beersSimple()

}
\references{
\insertRef{beers1945modified}{DemoTools}
\insertRef{siegel2004methods}{DemoTools}
}
