% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_coh.R
\name{census_cohort_adjust}
\alias{census_cohort_adjust}
\title{shift census populations to match single year cohorts}
\usage{
census_cohort_adjust(pop, age, date)
}
\arguments{
\item{pop}{numeric vector. Population counts in age groups, presumably from a census with an exact reference date.}

\item{age}{integer vector. Lower bound of single age groups}

\item{date}{Either a \code{Date} class object or an unambiguous character string in the format \code{"YYYY-MM-DD"}.}
}
\description{
Matches the (single) ages of a census to single cohorts. For use in intercensal interpolations. Ages are potentially blended to match single cohort line assuming that the population in each age is uniformly distributed over the age group.
}
\examples{
pop <- seq(10000,100,length.out = 101)
age <- 0:100
d1 <- "2020-01-01"
d2 <- "2020-07-01"
d3 <- "2020-12-21"

census_cohort_adjust(pop, age, d1)
census_cohort_adjust(pop, age, d2)
census_cohort_adjust(pop, age, d3)
census_cohort_adjust(pop, age, 2020.5)
}
