% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_heaping.R
\name{check_heaping_bachi}
\alias{check_heaping_bachi}
\title{calculate Bachi's index of age heaping}
\usage{
check_heaping_bachi(
  Value,
  Age,
  ageMin = 23,
  ageMax = 77,
  method = "balanced",
  details = FALSE
)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{integer. The lowest age included in calculations. Default 25.}

\item{ageMax}{integer. The upper age bound used for calculations. Default 65.}

\item{method}{either \code{"balanced"} or \code{"pasex"}}

\item{details}{logical. Should a list of output be given}
}
\value{
The value of the index.
}
\description{
Bachi's index involves applying the Whipple method repeatedly to determine the extent of preference for each final digit. Similarly to Myers', it equals the sum of the positive deviations from 10 percent. It has a theoretical range from 0 to 90, and 10 is the expected value for each digit. There are two implementations.
}
\details{
\code{ageMax} is an inclusive upper bound, treated as interval. If you want ages
23 to 77, then give \code{ageMin = 23} and \code{ageMax = 77}, not 80. The \code{ageMin} is respected strictly, whereas \code{ageMax} could be higher than the actual maximum age used. You can see the age ranges actually used by specifying \code{details = TRUE}.
}
\examples{
 Age <- 0:99
 check_heaping_bachi(pop1m_pasex, Age, ageMin = 23, ageMax = 77, method = "balanced")
 check_heaping_bachi(pop1m_ind, Age, ageMin = 23, ageMax = 77, method = "balanced")
 # default simpler
 check_heaping_bachi(pop1m_pasex, Age, ageMin = 23, ageMax = 77, method = "pasex")
# linear population, should give 0 for pasex
 check_heaping_bachi(seq(100000,1000,by=-1000),Age, ageMin = 23, ageMax = 77, method = "pasex")
# fully concentrated, should give 90 
 pop_concetrated <- rep(c(100,rep(0,9)),10)
 check_heaping_bachi(pop_concetrated,Age, ageMin = 23, ageMax = 77, method = "pasex")
 check_heaping_bachi(pop_concetrated,Age, ageMin = 23, ageMax = 77, method = "balanced")
}
\references{
\insertRef{PAS}{DemoTools}
\insertRef{bachi1951tendency}{DemoTools}
\insertRef{shryock1973methods}{DemoTools}
}
