% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGEHEAP.R
\name{CoaleLi}
\alias{CoaleLi}
\title{Coale-Li age heaping index}
\usage{
CoaleLi(Value, Age, ageMin = 60, ageMax = max(Age), terms = 5,
  digit = 0)
}
\arguments{
\item{Value}{numeric. A vector of demographic rates or counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{integer. The lowest age included in calculations. Default 20.}

\item{ageMax}{integer. The upper age bound used for calculations. Default 65.}

\item{terms}{integer. Length of the (centered) moving average be. Default 5.}

\item{digit}{integer. Any digit 0-9. Default 0.}
}
\value{
The value of the index.
}
\description{
This implementation is based largely on the sparse verbal description given in
Coale and Li (1991): calculate a two-stage 5-term moving average as a reference pattern, then
take ratios with respect to this. Ratios for a given terminal digit can then be averaged to produce
an index. This procedure was used in that paper for ages 65-100 for mortality rates. 
It is probably better suited to rates than counts, but that is not a hard rule.
}
\details{
\code{digit} could also be a vector of digits, but the more digits one includes (excepting 0 and 5) the closer the index will get to 1. It is therefore recommended for single digits, or else \code{c(0,5)}. \code{ageMax} is an inclusive upper bound, treated as interval. If you want ages 20 to 89, then give \code{ageMin = 20} and \code{ageMax = 89}, not 90. By default all available ages greater than or equal to \code{ageMin} are used.
}
\examples{
Value <- c(80626,95823,104315,115813,100796,105086,97266,116328,
75984,89982,95525,56973,78767,65672,53438,85014,
47600,64363,42195,42262,73221,30080,34391,29072,
20531,66171,24029,44227,24128,23599,82088,16454,
22628,17108,12531,57325,17220,28425,16206,17532,
65976,11593,15828,13541,8133,44696,11165,18543,
12614,12041,55798,9324,10772,10453,6773,28358,
9916,13348,8039,7583,42470,5288,5317,6582,
3361,17949,3650,5873,3279,3336,27368,1965,
2495,2319,1335,12022,1401,1668,1360,1185,
9167,424,568,462,282,6206,343,409,333,291,4137,133,169,157,89,2068,68,81,66,57)
Age <- 0:99
CoaleLi(Value, Age, 65, 95, 5, 0) # 3.7
CoaleLi(Value, Age, 65, 95, 5, 5) # 3.5 almost just as high
}
\references{
\insertRef{coale1991effect}{DemoTools}
}
