% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dec.date}
\alias{dec.date}
\title{Convert date to decimal year fraction.}
\usage{
dec.date(date)
}
\arguments{
\item{date}{Either a \code{Date} class object or an unambiguous character string in the format \code{"YYYY-MM-DD"}.}
}
\value{
Numeric expression of the date, year plus the fraction of the year passed as of the date.
}
\description{
Convert a character or date class to decimal, taking into account leap years.
}
\details{
This makes use of two HMD functions, \code{ypart()}, and \code{isLeapYear()} to compute. If the date is numeric, it is returned as such. 
If it is \code{"character"}, we try to coerce to \code{"Date"} class, ergo, it is best to specify a character string in an unambiguous \code{"YYYY-MM-DD"} format.
 If \code{date} is given in a \code{"Date"} class it is dealt with accordingly.
}
