% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoTools-package.R
\docType{package}
\name{DemoTools-package}
\alias{DemoTools-package}
\title{DemoTools: Standardize, Evaluate, and Adjust Demographic Data}
\description{
DemoTools: Standardize, Evaluate, and Adjust Demographic Data
}
\author{
\itemize{
\item {Tim Riffe}
\item {Jose Manuel Aburto}
\item {Sean Fennell}
\item {Marius D. Pascariu}
}

Maintainer: Tim Riffe  <tim.riffe@gmail.com>
}
