% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoTools-package.R
\docType{package}
\name{DemoTools-package}
\alias{DemoTools}
\alias{DemoTools-package}
\title{DemoTools: Standardize, Evaluate, and Adjust Demographic Data}
\description{
A collection of classic demographic methods used to evaluate, adjust, and standardize demographic data.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/timriffe/DemoTools/issues}
}

}
\author{
\strong{Maintainer}: Tim Riffe \email{tim.riffe@gmail.com}

Authors:
\itemize{
  \item Sean Fennell
  \item Jose Manuel Aburto
}

Other contributors:
\itemize{
  \item Juan Galeano [contributor]
  \item Marius D. Pascariu [contributor]
  \item Derek Burk [contributor]
  \item Greg Freedman Ellis [contributor]
}

}
