% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoTools-package.R
\docType{package}
\name{DemoTools-package}
\alias{DemoTools}
\alias{DemoTools-package}
\title{DemoTools: Standardize, Evaluate, and Adjust Demographic Data}
\description{
A collection of classic demographic methods used to evaluate, adjust, and standardize demographic data.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/timriffe/DemoTools/issues}
}

}
\author{
\strong{Maintainer}: Tim Riffe \email{tim.riffe@gmail.com} (\href{https://orcid.org/0000-0002-2673-4622}{ORCID})

Authors:
\itemize{
  \item José Manuel Aburto (\href{https://orcid.org/0000-0002-2926-6879}{ORCID})
  \item Ilya Kashnitsky (\href{https://orcid.org/0000-0003-1835-8687}{ORCID})
  \item Marius D. Pascariu (\href{https://orcid.org/0000-0002-2568-6489}{ORCID})
  \item Monica Alexander (\href{https://orcid.org/0000-0002-8135-3435}{ORCID})
  \item Sean Fennell
}

Other contributors:
\itemize{
  \item Jorge Cimentada (\href{https://orcid.org/0000-0001-5594-1156}{ORCID}) [contributor]
  \item Juan Galeano (\href{https://orcid.org/0000-0002-3682-1797}{ORCID}) [contributor]
  \item Derek Burk [contributor]
  \item Greg Freedman Ellis [contributor]
}

}
