% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoTools-package.R
\docType{package}
\name{DemoTools-package}
\alias{DemoTools}
\alias{DemoTools-package}
\title{DemoTools: Standardize, Evaluate, and Adjust Demographic Data}
\description{
A collection of classic demographic methods used to evaluate, adjust, and standardize demographic data.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/timriffe/DemoTools/issues}
}

}
\author{
\strong{Maintainer}: Tim Riffe \email{tim.riffe@gmail.com} (0000-0002-2673-4622)

Authors:
\itemize{
  \item José Manuel Aburto (0000-0002-2926-6879)
  \item Ilya Kashnitsky
  \item Marius D. Pascariu
  \item Monica Alexander
  \item Sean Fennell
}

Other contributors:
\itemize{
  \item Jorge Cimentada [contributor]
  \item Juan Galeano [contributor]
  \item Derek Burk [contributor]
  \item Greg Freedman Ellis [contributor]
}

}
