% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{DemoTools-renamed}
\alias{DemoTools-renamed}
\title{Renamed functions in DemoTools}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{monoCloseout}}: This function is now called \code{\link{graduate_mono_closeout}}.
 \item \code{\link{splitMono}}: This function is now called \code{\link{graduate_mono}}.
 \item \code{\link{grabill}}: This function is now called \code{\link{graduate_grabill}}.
 \item \code{\link{sprague}}: This function is now called \code{\link{graduate_sprague}}.
 \item \code{\link{beers}}: This function is now called \code{\link{graduate_beers}}.
 \item \code{\link{splitUniform}}: This function is now called \code{\link{graduate_uniform}}.
 
 \item \code{\link{aomegaMortalityLaws}}: This function is now called \code{\link{lt_a_closeout}}.
 \item \code{\link{lx2dx}}: This function is now called \code{\link{lt_id_l_d}}.  
 \item \code{\link{Lx2Tx}}: This function is now called \code{\link{lt_id_L_T}}.
 \item \code{\link{Lx2Tx}}: This function is now called \code{\link{lt_id_L_T}}.
 \item \code{\link{Lxlx2Sx}}: This function is now called \code{\link{lt_id_Ll_S}}.
 \item \code{\link{qx2lx}}: This function is now called \code{\link{lt_id_q_l}}.
 \item \code{\link{mxax2qx}}: This function is now called \code{\link{lt_id_ma_q}}.
 \item \code{\link{mx2qx}}: This function is now called \code{\link{lt_id_m_q}}.
 \item \code{\link{qxax2mx}}: This function is now called \code{\link{lt_id_qa_m}}.
 \item \code{\link{mxorqx2ax}}: This function is now called \code{\link{lt_id_morq_a}}.
 \item \code{\link{mxax2qx_Backstop}}: This function is now called \code{\link{lt_id_ma_q_robust}}.
 \item \code{\link{axUN}}: This function is now called \code{\link{lt_a_un}}.
 \item \code{\link{axPAS}}: This function is now called \code{\link{lt_a_pas}}.
 \item \code{\link{ax.greville.mortpak}}: This function is now called \code{\link{lt_id_morq_a_greville}}.
 \item \code{\link{LTabr}}: This function is now called \code{\link{lt_abridged}}.
 \item \code{\link{lt_single_simple}}: This function is now called \code{\link{lt_single_mx}}.
 
 \item \code{\link{extra.mortality}}: This function is now called \code{\link{lt_rule_m_extrapolate}}.
 \item \code{\link{geta0CD}}: This function is now called \code{\link{lt_rule_1a0_cd}}.
}
}
