% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepop.R
\name{downloadSRB}
\alias{downloadSRB}
\title{Extract SRB estimates from WPP2019}
\usage{
downloadSRB(SRB, country, DatesOut)
}
\arguments{
\item{SRB}{sex ratio at birth. Either \code{NULL}, a scalar to assume constant, or a vector of length 3, assumed.}

\item{country}{character country name available UN Pop Div \code{LocName} set}

\item{DatesOut}{numeric vector of three decimal dates produced by \code{basepop_ive()}}
}
\value{
numeric vector with three SRB estimates
}
\description{
We use the \code{WPP2019_births} dataset from \code{DemoToolsData} for the sex ratio at birth. Births from WPP 2019 were graduates to single year totals.
}
