% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_downloads.R
\name{fetch_wpp_births}
\alias{fetch_wpp_births}
\title{extract births from wpp2019}
\usage{
fetch_wpp_births(births, yrs_births, location, sex, verbose)
}
\arguments{
\item{births}{\code{NULL} or else a vector of births to simply return}

\item{yrs_births}{vector of years to extract}

\item{location}{UN Pop Dov \code{LocName} or \code{LocID}}

\item{sex}{\code{"male"}, \code{"female"}, or \code{"both"}}

\item{verbose}{logical, shall we send optional messages to the console?}
}
\value{
vector of births
}
\description{
extract births from wpp2019
}
