% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_model_lq.R
\name{find.my.case}
\alias{find.my.case}
\title{Function that determines the case/problem we have to solve
It also performs some checks}
\usage{
find.my.case(par_ind)
}
\arguments{
\item{par_ind}{logical vector of length 5}
}
\description{
Function that determines the case/problem we have to solve
It also performs some checks
}
\details{
\code{par_ind} should consist in logicals in the following order: \code{q0_5}, \code{q0_1}, \code{q15_45}, \code{q15_35}, \code{e0}. This is faithfully constructed in calling functions as required.
}
\keyword{internal}
