% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fitted_logquad_b}
\alias{fitted_logquad_b}
\title{LogQuad model for BOTH SEX fitted for all HMD life tables}
\format{
List of 6:
\describe{
\item{input}{List of parameters passed to \code{MortalityEstimate::wilmoth}.}
\item{call}{R code call line used for fitting.}
\item{coefficients}{Dataframe with 24 rows and 4 variables -- parameters \code{ax}, \code{bx}, \code{cx}, \code{vx} for each age group.}
\item{k}{\code{k} fitting parameter of the LogQuag model, a vector of 968 values.}
\item{fitted.values}{Fitted death rates, a matrix 968x24.}
\item{model.info}{Model formula.}
}
}
\source{
Human Mortality Database \link{Portal}. Retrieved 2019-11-28, from
\url{https://mortality.org}
}
\usage{
fitted_logquad_b
}
\description{
LogQuad model fitted using \code{MortalityEstimate::wilmoth} for all BOTH SEX period life tables present in Human Mortality Database (\url{https://mortality.org}) in December 2019 (968 life tables). Object of class \code{wilmoth}.
}
\keyword{datasets}
