% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getModelLifeTable}
\alias{getModelLifeTable}
\title{Wrapper to provide a single location to reference all model life tables.}
\usage{
getModelLifeTable(ModelName, InputSex)
}
\arguments{
\item{ModelName}{string naming the life table to return. Can be "coale-demeny west".}

\item{Sex}{string indicating which sex should be returned. Can be either "M" or "F".}
}
\value{
list of life tables
}
\description{

}
\details{

}
