% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grabill.R
\name{grabill}
\alias{grabill}
\title{The basic Grabill age-splitting method}
\usage{
grabill(popmat, Age, OAG = TRUE)
}
\arguments{
\item{popmat}{numeric. Matrix of age-period population counts in 5-year age groups with integer-labeled 
margins (age in rows and year in columns).}

\item{Age}{integer. A vector of ages corresponding to the lower integer bound of the counts. Detected from row names of \code{popmat} if missing.}

\item{OAG}{logical. Whether or not the final age group open. Default \code{TRUE}.}
}
\value{
An age-period matrix of split population counts with the same number of 
columns as \code{popmat}, and single ages in rows.
}
\description{
This method uses Grabill's redistribution of middle ages and blends into
Sprague estimated single-age population counts for the first and final ten ages. Open age groups
are preserved, as are annual totals.
}
\details{
Dimension labelling is necessary. There must be at least six age groups (including the open group). One year of data will 
work as well, as long as it's given as a single-column matrix. Data may be given in either single or grouped ages.
}
\examples{
p5 <- structure(c(54170, 44775, 42142, 38464, 34406, 30386, 26933, 
				23481, 20602, 16489, 14248, 9928, 8490, 4801, 3599, 2048, 941, 
				326, 80, 17, 0, 57424, 44475, 41752, 39628, 34757, 30605, 27183, 
				23792, 20724, 17056, 14059, 10585, 8103, 5306, 3367, 2040, 963, 
				315, 80, 16, 1, 60272, 44780, 41804, 40229, 35155, 30978, 27456, 
				24097, 20873, 17546, 13990, 11146, 7841, 5738, 3184, 2062, 961, 
				311, 80, 15, 1, 62727, 45681, 42101, 40474, 35599, 31439, 27758, 
				24396, 21055, 17958, 14046, 11589, 7731, 6060, 3086, 2083, 949, 
				312, 79, 14, 1, 64816, 47137, 42508, 40532, 36083, 31940, 28092, 
				24693, 21274, 18299, 14223, 11906, 7785, 6255, 3090, 2084, 938, 
				316, 80, 14, 2), 
		.Dim = c(21L, 5L), 
		.Dimnames = list(seq(0,100,by=5), 1950:1954))
head(p5) # this is the entire matrix
p1g <- grabill(p5)
head(p1g); tail(p1g)
colSums(p1g) - colSums(p5) 
#p1s <- sprague(p5)
\dontrun{
plot(seq(0,100,by=5),p5[,1]/5,type = "s", col = "gray", xlab = "Age", ylab = "Count")
lines(0:100, p1g[,1], col = "red", lwd = 2)
lines(0:100, p1s[,1], col = "blue", lty = 2)
legend("topright", 
	lty = c(1,1,2), 
	col = c("gray","red","blue"), 
	lwd = c(1,2,1), 
	legend = c("grouped","Grabill", "Sprague"))
}

# also works for single ages:
 popmat <- structure(
    c(9544406, 7471790, 11590109, 11881844, 11872503, 12968350, 
		11993151, 10033918, 14312222, 8111523, 15311047, 6861510, 13305117, 
		7454575, 9015381, 10325432, 9055588, 5519173, 12546779, 4784102, 
		13365429, 4630254, 9595545, 4727963, 5195032, 15061479, 5467392, 
		4011539, 8033850, 1972327, 17396266, 1647397, 6539557, 2233521, 
		2101024, 16768198, 3211834, 1923169, 4472854, 1182245, 15874081, 
		1017752, 3673865, 1247304, 1029243, 12619050, 1499847, 1250321, 
		2862148, 723195, 12396632, 733501, 2186678, 777379, 810700, 7298270, 
		1116032, 650402, 1465209, 411834, 9478824, 429296, 1190060, 446290, 
		362767, 4998209, 388753, 334629, 593906, 178133, 4560342, 179460, 
		481230, 159087, 155831, 1606147, 166763, 93569, 182238, 53567, 
		1715697, 127486, 150782, 52332, 48664, 456387, 46978, 34448, 
		44015, 19172, 329149, 48004, 28574, 9200, 7003, 75195, 13140, 
		5889, 18915, 21221, 72373), .Dim = c(101L, 1L), .Dimnames = list(
		0:100, NULL))
grab1 <- grabill(popmat)
\dontrun{
plot(0:100, c(popmat))
lines(0:100, c(grab1))
}
}
\references{
\insertRef{shryock1973methods}{DemoTools}
}
