% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_coh.R
\name{interp_coh}
\alias{interp_coh}
\title{component-free intercensal cohort interpolation}
\usage{
interp_coh(
  c1,
  c2,
  date1,
  date2,
  age1,
  age2 = age1,
  lxMat = NULL,
  age_lx = NULL,
  dates_lx = NULL,
  births = NULL,
  country = NULL,
  sex = "both",
  midyear = FALSE,
  ...
)
}
\arguments{
\item{c1}{numeric vector. The second (right) census in single age groups}

\item{date1}{reference date of c1`. Either a Date class object or an unambiguous character string in the format "YYYY-MM-DD".}

\item{date2}{reference date of c2`. Either a Date class object or an unambiguous character string in the format "YYYY-MM-DD".}

\item{age1}{integer vector. single ages of \code{c1}}

\item{age2}{integer vector. single ages of \code{c2}}

\item{births}{integer vector. Raw birth counts for the corresponding (sub)-population, one value per each year of the intercensal period including both census years}

\item{country}{text string. The country of the census}

\item{sex}{text string. Sex of the sub-population}

\item{midyear}{logical. \code{FALSE} means all Jan 1 dates between \code{date1} and \code{date2} are returned. \code{TRUE} means all July 1 intercensal dates are returned.}
}
\description{
Cohorts between two censuses are interpolated flexibly using linear, exponential, or power rules. The lower and upper intercensal triangles are filled using within-age interpolation. This function is experimental and still in development.
}
\examples{

\dontrun{
interp_coh(
country = "Russian Federation", 
sex = "male",
c1 = pop1m_rus2002, 
c2 = pop1m_rus2010, 
date1 = "2002-10-16", 
date2 = "2010-10-25", 
age1 = 0:100,
births = c(719511L, 760934L, 772973L, 749554L, 760831L, 828772L, 880543L, 905380L, 919639L)
) \%>\% 
  select(age, year, pop_jan1) \%>\% 
  pivot_wider(names_from = year, values_from = "pop_jan1") \%>\%
  arrange(age) 
}
}
\seealso{
interp
}
