% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_lc_lim.R
\name{interp_lc_lim}
\alias{interp_lc_lim}
\title{Lee-Carter method with limited data.}
\usage{
interp_lc_lim(
  data = NULL,
  dates_out = dates_in,
  Single = FALSE,
  dates_e0 = NULL,
  e0_Males = NULL,
  e0_Females = NULL,
  prev_divergence = FALSE,
  OAnew = max(data$Age),
  OAG = TRUE,
  extrapLaw = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame with cols: Date, Sex, Age, nMx (opt), nqx (opt), lx (opt)}

\item{dates_out}{numeric. Vector of decimal years to interpolate or extrapolate.}

\item{dates_e0}{numeric. Vector of decimal years where \code{"e_0"} should be fitted when apply method.}

\item{e0_Males}{numeric. Vector of life expectancy by year to be fitted. Same length than \code{"dates_e0"}.}

\item{e0_Females}{numeric. Vector of life expectancy by year to be fitted. Same length than \code{"dates_e0"}.}

\item{prev_divergence}{logical. Whether or not prevent divergence and sex crossover. Default \code{FALSE.}}

\item{OAG}{logical. Whether or not the last element of \code{nMx} (or \code{nqx} or \code{lx}) is an open age group. Default \code{TRUE.}}

\item{extrapLaw}{character. If extrapolating, which parametric mortality law should be invoked? Options include  \code{"Coale-Guo"} or other functions in \code{"lt_abridged"} function.}

\item{...}{Arguments passed to \verb{\link{lt_abridged}}.}
}
\value{
Lifetable in a data.frame with columns
\itemize{
\item{Year}{integer. Years included in dates_out},
\item{Age}{integer. Lower bound of abridged age class},
\item{AgeInt}{integer. Age class widths.}
\item{nMx}{numeric. Age-specific central death rates.}
\item{nAx}{numeric. Average time spent in interval by those deceased in interval. }
\item{nqx}{numeric. Age-specific conditional death probabilities.}
\item{lx}{numeric. Lifetable survivorship}
\item{ndx}{numeric. Lifetable deaths distribution.}
\item{nLx}{numeric. Lifetable exposure.}
\item{Sx}{numeric. Survivor ratios in uniform 5-year age groups.}
\item{Tx}{numeric. Lifetable total years left to live above age x.}
\item{ex}{numeric. Age-specific remaining life expectancy.}
}
}
\description{
Given a matrix of age (rows) by year (cols) for each sex, this function interpolate/extrapolate life tables
using the method for limited data suggested by  Li et. al (2004) (at least three observed years).
}
\details{
Based on spreedsheet "Li_2018_Limited_Lee-Carter-v4.xlsm" from UN. Only useful for abridged ages.
The main options are the use of non-divergent method for sex coherency (Li & Lee, 2005),
and the possibility of fitting \code{"k"} to replicate \code{"e_0"} at some given years.
Coale-Guo method allows to smooth rates at older ages.
}
\note{
Draft Version
}
\examples{
# Using Lee-Carter method for Sweden, assuming only is available mortality rates
library(dtplyr)
library(tidyverse)
library(HMDHFDplus)
library(ungroup)
library(DemoTools)

# LC with limited data
dates_in <- seq(1980,2010,10)
dates_out <- seq(1948,2018,5)
data <- readHMDweb("SWE", "Mx_5x1", user, pass, fixup = TRUE)\%>\% 
  select(Date = Year, Age, Male, Female) \%>\% 
  gather(Sex,nMx,-Date,-Age) \%>\% 
  mutate(Sex=ifelse(Sex=="Male","m","f")) \%>\% 
  filter(Date \%in\% dates_in, Age<=100)

# abr ages
lc_lim_data <- interp_lc_lim(data = data, dates_out = dates_out)

\dontrun{
  lc_lim_data \%>\% ggplot(aes(Age,nMx,col=factor(Date))) +
    geom_step() + scale_y_log10() + facet_wrap(~Sex)
}

# simple ages - takes like 2 minutes
start <- Sys.time()
lc_lim_data <- interp_lc_lim(data = data, dates_out = dates_out, OAnew = 100,
                             Single = T)
Sys.time() - start

\dontrun{
  lc_lim_data \%>\% ggplot(aes(Age,nMx,col=factor(Date))) +
    geom_step() + scale_y_log10() + facet_wrap(~Sex)
}

# Avoiding cross-over between sex
lc_lim_nondiv <- interp_lc_lim(data = data, dates_out = dates_out, OAnew = 100,
                               prev_divergence = T)
\dontrun{
  lc_lim_nondiv \%>\% ggplot(aes(Age,nMx,col=factor(Date))) + 
    geom_step() + scale_y_log10() + facet_wrap(~Sex)
}

# Using information about e0 in some years, replicate an estimate of e0 in dates_out 
# improve performance next step maybe
dates_e0 <- seq(1960,2015,5)
e0_Males <- readHMDweb("SWE", "E0per", user, pass, fixup = TRUE) \%>\% 
  filter(Year \%in\% dates_e0) \%>\% pull(Male)
e0_Females <- readHMDweb("SWE", "E0per", user, pass, fixup = TRUE) \%>\% 
  filter(Year \%in\% dates_e0) \%>\% pull(Female)
# needs that by the moment for interp: source("R/AGEINT.R")
lc_lim_fite0 <- interp_lc_lim(data = data, dates_out = dates_out, OAnew = 100,
                              dates_e0 = dates_e0,
                              e0_Males = e0_Males, 
                              e0_Females = e0_Females)
\dontrun{                           
  ggplot() + 
    geom_line(data = lc_lim_fite0 \%>\% filter(Age==0), aes(Date,ex,col=factor(Sex))) + 
    geom_point(data = data.frame(Sex = c(rep("m",length(e0_Males)), rep("f",length(e0_Males))),
                                 ex = c(e0_Males, e0_Females),
                                 Year = rep(dates_e0,2)),
               aes(Year,ex,col=factor(Sex)))
}

# smooth and/or extend open age group, in this case input is for 80+, and smoothing with Kannisto law
# Coale-Guo is coming ;)...
lc_lim_extOAg <- interp_lc_lim(data = data \%>\% filter(Age<=80), 
                               dates_out = dates_out, OAnew = 100,
                               OAG = F, extrapLaw = "kannisto")
\dontrun{ 
  lc_lim_extOAg \%>\% ggplot(aes(Age,nMx,col=factor(Date))) + 
    geom_step() + scale_y_log10() + facet_wrap(~Sex)
}
#End
}
\references{
\insertRef{Li2005}{DemoTools}
\insertRef{Li2004}{DemoTools}
}
