% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_lc_lim.R
\name{interp_lc_lim}
\alias{interp_lc_lim}
\title{Lee-Carter method with limited data.}
\usage{
interp_lc_lim(
  input = NULL,
  dates_out = dates_in,
  Single = FALSE,
  dates_e0 = NULL,
  e0_Males = NULL,
  e0_Females = NULL,
  prev_divergence = FALSE,
  OAG = TRUE,
  verbose = TRUE,
  error_sheet = FALSE,
  ...
)
}
\arguments{
\item{input}{data.frame with cols: Date, Sex, Age, nMx (opt), nqx (opt), lx (opt)}

\item{dates_out}{numeric. Vector of decimal years to interpolate or extrapolate.}

\item{Single}{logical. Wheter or not the lifetable output is by single ages.}

\item{dates_e0}{numeric. Vector of decimal years where \code{"e_0"} should be fitted when apply method.}

\item{e0_Males}{numeric. Vector of life expectancy by year to be fitted. Same length than \code{"dates_e0"}.}

\item{e0_Females}{numeric. Vector of life expectancy by year to be fitted. Same length than \code{"dates_e0"}.}

\item{prev_divergence}{logical. Whether or not prevent divergence and sex crossover. Default \code{FALSE.}}

\item{OAG}{logical. Whether or not the last element of \code{nMx} (or \code{nqx} or \code{lx}) is an open age group. Default \code{TRUE.}}

\item{verbose}{logical. Default \code{FALSE}.}

\item{error_sheet}{logical. Just to reproduce UN spreadsheets. Will be removed. Default \code{FALSE}.}

\item{...}{Other arguments to be passed on to the \code{\link[DemoTools]{lt_abridged}} function.}
}
\value{
Lifetable in a data.frame with columns
\itemize{
\item \code{Date} numeric. Dates included in dates_out,
\item \code{Sex} character. Male \code{"m"} or female \code{"f"},
\item \code{Age} integer. Lower bound of abridged age class,
\item `AgeInt`` integer. Age class widths.
\item \code{nMx} numeric. Age-specific central death rates.
\item \code{nAx} numeric. Average time spent in interval by those deceased in interval.
\item \code{nqx} numeric. Age-specific conditional death probabilities.
\item \code{lx} numeric. Lifetable survivorship
\item \code{ndx} numeric. Lifetable deaths distribution.
\item \code{nLx} numeric. Lifetable exposure.
\item \code{Sx} numeric. Survivor ratios in uniform 5-year age groups.
\item \code{Tx} numeric. Lifetable total years left to live above age x.
\item \code{ex} numeric. Age-specific remaining life expectancy.
}
}
\description{
Given a data frame with dates, sex and mortality data by age (rates, conditionated probabilities of death
or survival function), this function interpolate/extrapolate life tables
using the method for limited data suggested by  Li et. al (2004) (at least three observed years).
}
\details{
Based on spreedsheet "Li_2018_Limited_Lee-Carter-v4.xlsm" from UN. Useful for abridged or single ages, and allows
output in both formats also.
The main options are the use of non-divergent method for sex coherency (Li & Lee, 2005),
and the possibility of fitting \code{"k"} to replicate \code{"e_0"} at some given dates.
}
\note{
Draft Version
}
\examples{
# mortality rates from Sweden, for specific dates
data("mA_swe")

# needs mortality rates in this dates: 
dates_out <- as.Date(paste0(seq(1948,2018,5),"-07-01"))

# apply LC with limited data to extrap/interpolate
lc_lim_data <- interp_lc_lim(input = mA_swe, dates_out = dates_out, OAG = FALSE)

\dontrun{
lc_lim_data \%>\% ggplot(aes(Age,nMx,col=factor(round(Date,1)))) +
  geom_step() + scale_color_viridis_d() + 
  scale_y_log10() + theme_classic() + facet_wrap(~Sex)
}

# with simple ages as output
lc_lim_data_single <- interp_lc_lim(input = mA_swe, dates_out = dates_out, OAG = FALSE,
                                    Single = TRUE)

\dontrun{
lc_lim_data_single \%>\% ggplot(aes(Age,nMx,col=factor(round(Date,1)))) +
  geom_step() + scale_color_viridis_d() + 
  scale_y_log10() + theme_classic() + facet_wrap(~Sex)
}

# Avoiding cross-over between sex.
lc_lim_nondiv <- interp_lc_lim(input = mA_swe, dates_out = dates_out, OAG = FALSE,
                               prev_divergence = TRUE)
\dontrun{
lc_lim_nondiv \%>\% ggplot(aes(Age,nMx,col=factor(round(Date,1)))) +
  geom_step() + scale_color_viridis_d() + 
  scale_y_log10() + theme_classic() + facet_wrap(~Sex)
}

# Fitting information about e0 in Sweden for past years.
data("e0_swe")
lc_lim_fite0 <- interp_lc_lim(input = mA_swe, dates_out = dates_out, OAG = FALSE,
                              dates_e0 = unique(e0_swe$Date),
                              e0_Males = e0_swe$e0[e0_swe$Sex=="m"], 
                              e0_Females = e0_swe$e0[e0_swe$Sex=="f"])
\dontrun{                               
ggplot() + 
  geom_point(data = e0_swe, aes(Date,e0,col=factor(Sex)))+
  geom_line(data = lc_lim_fite0[lc_lim_fite0$Age==0,], aes(Date,ex,col=factor(Sex)))+
  labs(color = "Sex")+
  theme_classic()
}

# smooth and/or extend open age group, in this case input is for 80+, and chosen law is Makeham.
lc_lim_extOAg <- interp_lc_lim(input = mA_swe[mA_swe$Age<=80,], dates_out = dates_out,
                               OAG = FALSE,
                               OAnew=100,
                               extrapLaw = "makeham")
\dontrun{
ggplot() + 
  geom_step(data = lc_lim_extOAg, aes(Age,nMx,col=factor(Date))) +
  scale_y_log10() + facet_wrap(~Sex)
  }
#End
}
\references{
\insertRef{Li2005}{DemoTools}
\insertRef{Li2004}{DemoTools}
}
\seealso{
\code{\link[DemoTools]{lt_abridged}}
}
