% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_lc_lim.R
\name{interp_lc_lim_estimate}
\alias{interp_lc_lim_estimate}
\title{Estimate LC with limited data params}
\usage{
interp_lc_lim_estimate(M, dates_in, dates_out, SVD = F)
}
\arguments{
\item{M}{numeric. Matrix with many rows as ages and columns as dates_in.}

\item{dates_in}{numeric. Vector of dates with input rates.}

\item{dates_out}{numeric. Vector of dates for estimate a set of rates.}

\item{...}{Other arguments to be passed on to the \code{\link[DemoTools]{lt_abridged}} function.}
}
\description{
Estimate LC with limited data from a matrix of rates (age by dates).
}
\details{
SVD for ax and bx. Fit a simmple linear model for k and interp/extrapolate for objective dates.
}
\references{
\insertRef{Li2004}{DemoTools}
}
