% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beers.R
\name{johnsonAdjust}
\alias{johnsonAdjust}
\title{adjust ages under 10 using a modification of Beers}
\usage{
johnsonAdjust(Age0, pop5, pop1)
}
\arguments{
\item{Age0}{numeric an estimate of age 0}

\item{pop5}{numeric matrix population in 5-year age groups}

\item{pop1}{numeric matrix population split using Beers ordinary (or some other) method}
}
\value{
numeric matrix of single age population estimates
}
\description{
Assuming we have an external estimate of age 0, this method
refits to the ordinary Beers single age results, remaining constrained to the
original 5-year age groups and smoothly blending into ages greater than 10.
}
\details{
This has not been tested using \code{pop1} as generated from other methods, such as 
the Beers modified, Sprague, or Grabill methods. Called internally by \code{BeersSimple()}.
}
\references{
\insertRef{stover2008spectrum}{DemoTools}
}
