% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_regroup_age.R
\name{lt_abridged2single}
\alias{lt_abridged2single}
\title{create a life table by single year of age by graduating an abridged life table}
\usage{
lt_abridged2single(
  Deaths = NULL,
  Exposures = NULL,
  nMx = NULL,
  nqx = NULL,
  lx = NULL,
  Age,
  radix = 1e+05,
  axmethod = "un",
  a0rule = "ak",
  Sex = "m",
  region = "w",
  IMR = NA,
  mod = TRUE,
  SRB = 1.05,
  OAG = TRUE,
  OAnew = max(Age),
  extrapLaw = NULL,
  extrapFrom = max(Age),
  extrapFit = NULL,
  ...
)
}
\arguments{
\item{Deaths}{numeric. Vector of death counts in abridged age classes.}

\item{Exposures}{numeric. Vector of population exposures in abridged age classes.}

\item{nMx}{numeric. Vector of mortality rates in abridged age classes.}

\item{nqx}{numeric. Vector of conditional death probabilities in abridged age classes.}

\item{lx}{numeric. Vector of lifetable survivorship at abridged ages.}

\item{Age}{integer. A vector of ages of the lower integer bound of the age classes.}

\item{radix}{numeric. Lifetable radix, \ifelse{html}{\out{l<sub>0}}{\eqn{l_0}}. Default 100000.}

\item{axmethod}{character. Either \code{"pas"} or \code{"un"}.}

\item{a0rule}{character. Either \code{"ak"} (default) or \code{"cd"}.}

\item{Sex}{character. Either male \code{"m"}, female \code{"f"}, or both \code{"b"}.}

\item{region}{character. North, East, South, or West: code{"n"}, code{"e"}, code{"s"}, code{"w"}. Default code{"w"}.}

\item{IMR}{numeric. Infant mortality rate \ifelse{html}{\out{q<sub>0}}{\eqn{q_0}}, in case available and \code{nqx} is not specified. Default \code{NA}.}

\item{mod}{logical. If \code{"un"} specified for \code{axmethod}, whether or not to use Nan Li's modification for ages 5-14. Default \code{TRUE}.}

\item{SRB}{the sex ratio at birth (boys / girls), detault 1.05}

\item{OAG}{logical. Whether or not the last element of \code{nMx} (or \code{nqx} or \code{lx}) is an open age group. Default \code{TRUE}.}

\item{OAnew}{integer. Desired open age group (5-year ages only). Default \code{max(Age)}. If higher then rates are extrapolated.}

\item{extrapLaw}{character. If extrapolating, which parametric mortality law should be invoked? Options include
\code{"Kannisto", "Kannisto_Makeham", "Makeham", "Gompertz", "GGompertz", "Beard",    "Beard_Makeham", "Quadratic"}. Default \code{"Kannisto"} if the highest age is at least 90, otherwise \code{"makeham"}. See details.}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality.}

\item{extrapFit}{integer vector. Ages to include in model fitting. Defaults to all ages \code{>          =60}.}

\item{...}{optional arguments passed to \code{pclm()}. For example, if you pass an expicit \code{lambda} parameter via the \code{control} argument, you can speed up estimation}
}
\value{
Single-year lifetable in data.frame with columns
\itemize{
\item{Age}{integer. Lower bound of single year age class},
\item{AgeInt}{integer. Age class widths.}
\item{nMx}{numeric. Age-specific central death rates.}
\item{nAx}{numeric. Average time spent in interval by those deceased in interval. }
\item{nqx}{numeric. Age-specific conditional death probabilities.}
\item{lx}{numeric. Lifetable survivorship}
\item{ndx}{numeric. Lifetable deaths distribution.}
\item{nLx}{numeric. Lifetable exposure.}
\item{Sx}{numeric. Survivor ratios.}
\item{Tx}{numeric. Lifetable total years left to live above age x.}
\item{ex}{numeric. Age-specific remaining life expectancy.}
}
}
\description{
Computes single year of age life table by graduating the mortality schedule of an abridged life table, using the \code{ungroup::pclm()} to ungroup binned count data. Returns complete single-age lifetable.
}
\details{
Similar to \code{lt_abridged()} details, forthcoming.
}
\examples{
 Mx <- c(.23669,.04672,.00982,.00511,.00697,.01036,.01169,
         .01332,.01528,.01757,.02092,.02517,.03225,.04241,.06056,
         .08574,.11840,.16226,.23745)
 Age = c(0,1,seq(5,85,by=5))
 AgeInt <- inferAgeIntAbr(vec = Mx)
 LTabr <- lt_abridged(nMx = Mx,
                      Age = Age, 
                      axmethod = "un",
                      Sex = "m",
                      mod = TRUE)
 
 LT1 <- lt_abridged2single(nMx = Mx,
                    Age = Age, 
                    axmethod = "un",
                    Sex = "m",
                    mod = TRUE)
LTabr$ex[1]
LT1$ex[1]
\dontrun{
plot(Age, LTabr$nMx,type = 's', log = 'y')
lines(LT1$Age, LT1$nMx)

plot(Age, LTabr$lx,type='S')
lines(LT1$Age, LT1$lx)
}
}
