% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R, R/LTidentities.R
\name{Lx2Tx}
\alias{Lx2Tx}
\alias{lt_id_L_T}
\title{Derive lifetable total person years left to live from exposure.}
\usage{
Lx2Tx(...)

lt_id_L_T(Lx)
}
\arguments{
\item{Lx}{numeric. Vector of lifetable exposure.}
}
\value{
Tx total years left to live above age x.
}
\description{
A lifetable identity. Tx is interpreted as the total years
left to live above age x in the life table stationary population.
}
\details{
No \code{NA} or other error checking here. This is taken as the numerator in the classic lifetable method
 of calculation of life expectancy.
}
\references{
\insertRef{preston2000demography}{DemoTools}
}
