% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTPOPDTH.R
\name{LTabr}
\alias{LTabr}
\title{Calculate an abridged-age lifetable.}
\usage{
LTabr(Deaths, Exposures, nMx, nqx, lx, Age, AgeInt = inferAgeIntAbr(Age =
  Age), radix = 1e+05, axmethod = "pas", Sex = "m", region = "w",
  IMR = NA, mod = TRUE, OAnew = max(Age), extrapLaw = c("Kannisto",
  "Makeham", "Gompertz", "Gamma-Gompertz", "Beard", "Makeham-Beard",
  "Quadratic")[1], extrapFrom = max(Age), extrapFit = Age[Age >= 60])
}
\arguments{
\item{Deaths}{numeric. Vector of death counts in abridged age classes.}

\item{Exposures}{numeric. Vector of population exposures in abridged age classes.}

\item{nMx}{numeric. Vector of mortality rates in abridged age classes.}

\item{nqx}{numeric. Vector of conditional death probabilities in abridged age classes.}

\item{lx}{numeric. Vector of lifetable survivorship at abridged ages.}

\item{Age}{integer. A vector of ages of the lower integer bound of the age classes.}

\item{AgeInt}{integer. Vector of age class widths. Ddefault \code{inferAgeIntAbr(Age = Age)}.}

\item{radix}{numeric. Lifetable radix, \ifelse{html}{\out{l<sub>0}}{\eqn{l_0}}. Default 100000.}

\item{axmethod}{character. Either \code{"pas"} or \code{"un"}.}

\item{Sex}{character. Either male \code{"m"}, female \code{"f"}, or both \code{"b"}.}

\item{region}{character. North, East, South, or West: code{"n"}, code{"e"}, code{"s"}, code{"w"}. Default code{"w"}.}

\item{IMR}{numeric. Infant mortality rate \ifelse{html}{\out{q<sub>0}}{\eqn{q_0}}, in case available and \code{nqx} is not specified. Default \code{NA}.}

\item{mod}{logical. If \code{"un"} specified for \code{axmethod}, whether or not to use Patrick Gerland's modification for ages 5-14. Default \code{TRUE}.}

\item{OAnew}{integer. Desired open age group (5-year ages only). Default \code{max(Age)}. If higher then rates are extrapolated.}

\item{extrapLaw}{character. If extrapolating, which parametric mortality law should be invoked? Options include  \code{"Kannisto","Makeham", "Gompertz", "Gamma-Gompertz", "Beard",
"Makeham-Beard", "Quadratic"}. Default \code{"Kannisto"}. See details.}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality.}

\item{extrapFit}{integer vector. Ages to include in model fitting. Defaults to all ages \code{>=60}.}
}
\value{
Lifetable in data.frame with columns
\itemize{
  \item{Age}{integer. Lower bound of abridged age class},
  \item{AgeInt}{integer. Age class widths.}
  \item{nMx}{numeric. Age-specific central death rates.} 
  \item{nAx}{numeric. Average time spent in interval by those deceased in interval. } 
  \item{nqx}{numeric. Age-specific conditional death probabilities.} 
  \item{lx}{numeric. Lifetable survivorship} 
  \item{ndx}{numeric. Lifetable deaths distribution.} 
  \item{nLx}{numeric. Lifetable exposure.} 
  \item{Tx}{numeric. Lifetable total years left to live above age x.} 
  \item{ex}{numeric. Age-specific remaining life expectancy.}
}
}
\description{
Given vectors for Deaths and Exposures, or Mx, or qx, or lx, calculate a full abridged lifetable.
}
\details{
The main variations here are in the treatment of nAx. In all cases, the lifetable is extended and closed out 
using one from a selection of methods implemented in the \code{MortalityLaws} package. 
For this, a desired open age must be specified, defaulting to the present open age group, but which can not exceed 100 in the present implementation.
The ax method beyond and including the original open age group in the data follows the Mortpak (UN) method only.
}
\examples{
# trial code from PAS LTPOPDTH, North, Males, IMR = .1
Exposures <- c(100958,466275,624134,559559,446736,370653,301862,249409,
		247473,223014,172260,149338,127242,105715,79614,53660,
		31021,16805,8000,4000,2000,1000)

Deaths <- c(8674,1592,618,411,755,1098,1100,1357,
		1335,3257,2200,4023,2167,4578,2956,4212,
		2887,2351,1500,900,500,300)
# lower age bounds
Age    <- c(0, 1, seq(5, 100, by = 5))
AgeInt <- c(diff(Age), NA)

PASLT <- LTabr(Deaths = Deaths, 
		Exposures = Exposures, 
		Age = Age,
		AgeInt =AgeInt,
		axmethod = "pas",
		IMR = .1,
		region = "n",
		Sex = "m")

# examples based on UN 1982 (p. 34)
Mx <- c(.23669,.04672,.00982,.00511,.00697,.01036,.01169,
		.01332,.01528,.01757,.02092,.02517,.03225,.04241,.06056,
		.08574,.11840,.16226,.23745)
excheckUN <-  c(35.000,42.901,47.190,44.438,
		40.523,36.868,33.691,30.567,27.500,24.485,21.504,18.599,
		15.758,13.080,10.584,8.466,6.729,5.312,4.211)
AgeInt <- inferAgeIntAbr(vec = Mx)

# generate two variants: with and without PG's variants
# for ages 5-14
UNLT1 <- LTabr(nMx = Mx,
		Age = c(0,1,seq(5,85,by=5)),
		AgeInt = AgeInt,
		axmethod = "un",
		Sex = "m", 
		mod = FALSE)
UNLT2 <- LTabr(nMx = Mx,
		Age = c(0,1,seq(5,85,by=5)),
		AgeInt = AgeInt,
		axmethod = "un",
		Sex = "m", 
		mod = TRUE)
		
#  \dontrun{
# 	 plot(UNLT2$ex - UNLT1$ex)
#  }

# a Mortpak unit test:
# data from  p. 82 United Nations (1988) Mortpak - ...
MPnMx <- c(0.12846,0.02477,0.00603,0.0034,
		0.00417,0.00513,0.00581,0.00645,0.00725,
		0.00813,0.00913,0.01199,0.01647,
		0.0256,0.04047,0.06624,0.10638,0.19611)

MPexcheck <- c(49.997,55.675,57.245,53.921,
		49.803,45.799,41.922,38.084,34.249,
		30.420,26.578,22.701,18.945,
		15.349,12.095,9.240,6.903,5.099)

# First with lifetable extention to 100
MP_UNLT100 <- LTabr(
		nMx = MPnMx,
		Age = c(0,1,seq(5,80,by=5)),
		AgeInt = inferAgeIntAbr(vec = MPnMx),
		axmethod = "un",
		Sex = "f",
		mod = FALSE,
		OAnew = 100)

# lifetable to original open age group
MP_UNLT80 <- LTabr(
		nMx = MPnMx,
		Age = c(0,1,seq(5,80,by=5)),
		AgeInt = inferAgeIntAbr(vec = MPnMx),
		axmethod = "un",
		Sex = "f",
		mod = FALSE,
		OAnew = 80)

# same, but truncated at 60
MP_UNLT60 <- LTabr(
		nMx = MPnMx,
		Age = c(0,1,seq(5,80,by=5)),
		AgeInt = inferAgeIntAbr(vec = MPnMx),
		axmethod = "un",
		Sex = "f",
		mod = FALSE,
		OAnew = 60)
}
\references{
\insertRef{greville1977short}{DemoTools}
\insertRef{un1982model}{DemoTools}
\insertRef{arriaga1994population}{DemoTools}
\insertRef{mortpak1988}{DemoTools}
\insertRef{PAS}{DemoTools}
}
