% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_model_lq.R
\name{lthat.logquad}
\alias{lthat.logquad}
\title{Estimated life table using the log-quadratic model}
\usage{
lthat.logquad(coefs, x, q0_5, k, radix, Sex, ...)
}
\arguments{
\item{coefs}{Estimated coefficients}

\item{q0_5}{5q0. The probability that a new-born will die during the 
subsequent 5 years;}

\item{radix}{Life table radix. Default: 10^5;}

\item{Sex}{Choose the sex of the population. This choice defines the use
of a corresponding Log-Quadratic (\code{wilmoth})
 model fitted for the whole Human Mortality Database (as of Dec 2019, 
 there are 968 life tables for each sex).
The following options are available: \itemize{
  \item{\code{"b"}} -- Both sex;
  \item{\code{"f"}} -- Females;
  \item{\code{"m"}} -- Males.
  }}

\item{...}{Additional arguments affecting the predictions produced.}
}
\description{
Estimated life table using the log-quadratic model
}
\keyword{internal}
