% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{lxdxax2Lx}
\alias{lxdxax2Lx}
\title{Derive lifetable exposure from lx, ndx and nax.}
\usage{
lxdxax2Lx(lx, ndx, nax, AgeInt)
}
\arguments{
\item{lx}{numeric.  Vector of age-specific lifetable survivorship.}

\item{ndx}{numeric. Vector of lifetable deaths, summing to radix of \code{lx}.}

\item{nax}{numeric. Vector of average time spent in interval by those dying in interval.}

\item{AgeInt}{integer. Vector of age class widths.}
}
\value{
nLx numeric vector of lifetable exposure.
}
\description{
This is a common approximation of lifetable exposure: 
All persons surviving to the end of the interval time the interval width, plus all those that died 
in the interval multiplied by their average time spent in the interval.
}
\details{
There is no checking of equal vector lengths at this time.
}
\references{
\insertRef{preston2000demography}{DemoTools}
}
