% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ma}
\alias{ma}
\title{a simple centered moving average function}
\usage{
ma(x, n = 5)
}
\arguments{
\item{x}{numeric vector to produce a moving average of}

\item{n}{integer the width of the moving avarage. Default 5 time steps (years)}
}
\value{
numeric vector of same length as \code{x}.
}
\description{
This function is defined based on a code chunk found here \url{https://stackoverflow.com/questions/743812/calculating-moving-average}. This is a centered moving average of arbitrary width.
}
\details{
\code{NA} values
are used as padding on the left and right to make the returned vector of equal length.
}
\examples{
x  <- runif(100)
xx <- cumsum(x)
plot(x)
lines(ma(x))
lines(ma(x),9)
}
