% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAV.R
\name{mav}
\alias{mav}
\title{Calculate the moving average (mav) over 3 or 5 years.}
\usage{
mav(Value, Age, n = 3, OAG = TRUE)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts in single age groups.}

\item{Age}{integer. A vector of ages corresponding to the lower integer bound of the counts.}

\item{n}{integer. A single number, (often 3 or 5), indicating the number of years taken to smooth the population distribution by single ages.}

\item{OAG}{logical. Whether or not the top age group is open. Default \code{TRUE}.}
}
\value{
Vector with the smoothed demographic counts.
}
\description{
This arithmetic smoothing technique aims to eliminate irregularities of the population pyramid by averaging values in a moving window of user-defined width.
}
\details{
Ages may be single or grouped, but all age intervals are assumed equal.
}
\examples{
# data from GDA1981IREDA, Table 11, page 361-363: Morocco (Census of 1960)
Pop  <-c(303583,390782,523903,458546,517996,400630,485606,325423,471481,189710,
         385442,143205,270890,145105,138078,157444,153035,91566,247160,73115,
         384222,83551,198555,111347,129851,477510,149272,100814,178465,50684,
         577167,51878,97788,55544,58011,393200,85048,51131,80336,31246,
         454698,34864,51810,31146,26618,228718,38504,23616,40836,15589,
         339158,21349,26997,17590,17513,119763,22704,12336,17635,8485,
         323263,9535,13906,9063,8294,90459,9817,6376,8884,3773,160609)
Age  <-c(0:70)  
# final age group assumed open
mav(Pop, n = 3,Age=Age)      
}
\references{
\insertRef{GDA1981IREDA}{DemoTools}
}
\author{
Juan Galeano
}
