% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig_beta.R
\name{mig_beta}
\alias{mig_beta}
\title{Estimate intercensal migration by comparing census population, by age and
sex, to the results of a RUP projection.}
\usage{
mig_beta(
  c1,
  c2,
  date1,
  date2,
  age1 = 1:length(c1) - 1,
  age2 = 1:length(c2) - 1,
  dates_out = NULL,
  lxMat = NULL,
  age_lx = NULL,
  dates_lx = NULL,
  births = NULL,
  years_births = NULL,
  location = NULL,
  sex = "both",
  midyear = FALSE,
  verbose = TRUE,
  child_adjust = c("none", "cwr", "constant"),
  childage_max = NULL,
  oldage_adjust = c("none", "beers", "mav"),
  oldage_min = 65,
  ...
)
}
\arguments{
\item{c1}{numeric vector. The first (left) census in single age groups}

\item{c2}{numeric vector. The second (right) census in single age groups}

\item{date1}{reference date of c1`. Either a Date class object or an unambiguous character string in the format "YYYY-MM-DD".}

\item{date2}{reference date of c2`. Either a Date class object or an unambiguous character string in the format "YYYY-MM-DD".}

\item{age1}{integer vector. single ages of \code{c1}}

\item{age2}{integer vector. single ages of \code{c2}}

\item{dates_out}{vector of desired output dates coercible to numeric using \code{dec.date()}}

\item{lxMat}{numeric matrix containing lifetable survivorship, \code{l(x)}. Each row is an age group and each column a time point. At least two intercensal time points needed.}

\item{age_lx}{integer vector. Age classes in \code{lxMat}}

\item{dates_lx}{date, character, or numeric vector of the column time points for \code{lxMat}. If these are calendar-year estimates, then you can choose mid-year time points}

\item{births}{integer vector. Raw birth counts for the corresponding (sub)-population, one value per each year of the intercensal period including both census years. The first and last years should include all births in the given year; don't discount them in advance.}

\item{years_births}{numeric vector of calendar years of births.}

\item{location}{country name or LocID}

\item{sex}{character string, either \code{"male"}, \code{"female"}, or \code{"both"}}

\item{midyear}{logical. \code{FALSE} means all Jan 1 dates between \code{date1} and \code{date2} are returned. \code{TRUE} means all July 1 intercensal dates are returned.}

\item{verbose}{logical. Shall we send informative messages to the console?}

\item{child_adjust}{The method with which to adjust the youngest age groups.
If \code{"none"}, no adjustment is applied (default). If
child-woman ratio (\code{"cwr"}) is chosen, the first cohorts reflecting the
difference between \code{date2 - date1} are adjusted (plus age 0). If
child constant ratio (\code{"constant"}) is chosen, the first 15 age groups
are adjusted.}

\item{childage_max}{The maximum age from which to apply \code{child_adjust}.
By default, set to \code{NULL}, which gets translated into all the cohorts
between \code{date2} and \code{date1}. If \code{date2} is 2010 and
\code{date1} is 2002, the first 8 cohorts are adjusted. Otherwise, the user
can supply an integer.}

\item{oldage_adjust}{The type of adjustment to apply to ages at and above
\code{oldage_min}. \code{'beers'} applies a beers graduation method
while \code{'mav'} applies a moving average with cascading on the tails.
For more information see \code{?mav} and \code{?graduation_beers}.}

\item{oldage_min}{The minimum age from which to apply \code{oldage_adjust}.
By default, set to 65, so any adjustment from \code{oldage_adjust} will be
applied for 65+.}

\item{...}{optional arguments passed to \code{lt_single_qx}}
}
\value{
a numeric vector of the total migration in the intercensal period
for each age. Ages are set as names of each migration estimate.
}
\description{
This methods projects population from the first starting point to next census
without migration and computes the "Net Census Error" (NCE) which is
Census - Estimate by age from projection. It then distributes the NCE over
the cohort parallelogram assuming uniform distribution assuming it is all
migration. It finalizes by summing the estimate by age groups across the entire
intercensal period to have a total migration during the entire period.
Alternatively, a child adjustment and an old age adjustment can be applied.
}
\examples{

\dontrun{

mig_beta(
  location = "Russian Federation",
  sex = "male",
  c1 = pop1m_rus2002,
  c2 = pop1m_rus2010,
  date1 = "2002-10-16",
  date2 = "2010-10-25",
  age1 = 0:100,
  births = c(719511L, 760934L, 772973L, 749554L, 760831L, 828772L, 880543L, 905380L, 919639L)
)
}
}
