% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig_un_fam.R
\name{mig_un_fam}
\alias{mig_un_fam}
\title{Net migration by age for an UN family}
\usage{
mig_un_fam(NM, family, Single = TRUE)
}
\arguments{
\item{NM}{numeric. Total net migration to distribuite between ages and sex.}

\item{family}{character. Could be "Family", "Female Labor", "Male Labor".}

\item{Single}{logical. Results by simple age. Default \code{FALSE}.
Typically from pre-working age and working age parts of in Roger-Castro formula.}
}
\value{
List with
\itemize{
\item{params_RC} {data.frame. Roger-Castro parameters in a data.frame. Same as \code{mig_un_params} data.}
\item{net_migr} {data.frame. Net migrants by age and sex for the chosen family.}
}
}
\description{
Given a total net migration,
calculate the net migration age schedule based on the Rogers and Castro formula for UN families.
}
\examples{
# 10000 net migrants, comparing two possible families
nm1 <- mig_un_fam(NM = 10000, family = "Male Labor")
nm2 <- mig_un_fam(NM = 10000, family = "Family")
# See the female profile in for these models:
\dontrun{
plot(nm1$net_migr$age[nm1$net_migr$sex=="Female"],
     nm1$net_migr$nm[nm1$net_migr$sex=="Female"],
     xlab="Age",ylab="nm",ylim=c(0,300))
points(nm2$net_migr$age[nm2$net_migr$sex=="Female"],
       nm2$net_migr$nm[nm2$net_migr$sex=="Female"], col=2)
}
}
