% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mig_un_params}
\alias{mig_un_params}
\title{Parameters for considered migration profiles}
\format{
A data frame with:
\describe{
\item{family}{Types Family, Male Labor or Female Labor.}
\item{sex}{Male and Female.}
\item{mig_sign}{Inmigration or Emigration.}
\item{param}{Parameters from Roger-Castro.}
\item{median}{median of posterior distribution using Monte Carlo Markov Chains in \code{mig_estimate_rc}.}
}
}
\source{
UN spreadsheet "UNPD_Migration Age Profiles.xlsx"
}
\usage{
mig_un_params
}
\description{
Roger-Castro estimated parameters using \code{mig_estimate_rc} for Pre Working Age and Working Age profiles of migration.
}
\keyword{datasets}
