% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{mx2qx}
\alias{mx2qx}
\title{Derive nqx from nMx and nax.}
\usage{
mx2qx(nMx, nax, AgeInt = inferAgeIntAbr(vec = nMx))
}
\arguments{
\item{nMx}{numeric. Vector of age-specific death rates.}

\item{nax}{numeric. Vector of average time spent in interval by those dying in interval.}

\item{AgeInt}{integer. Vector of age class widths.}
}
\value{
nqx vector of age specific death probabilities derived via identity.
}
\description{
This is the standard identity to derive nqx from nax and nMx.
}
\details{
qx values calculated as greater than 1 are imputed with 1.
}
\references{
\insertRef{preston2000demography}{DemoTools}
}
