% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myers.R
\name{Myers}
\alias{Myers}
\title{calculate Myer's blended index of age heaping}
\usage{
Myers(Value, Age, ageMin = 10, ageMax = 90)
}
\arguments{
\item{Value}{a vector of demographic counts by single age}

\item{Age}{a vector of ages corresponding to the lower integer bound of the counts}

\item{ageMin}{the lowest age included in calcs. Default 10}

\item{ageMax}{the upper age bound used for calcs. Default 90}
}
\description{
Implementation following the PASEX spreadsheet SINGAGE
}
\details{
\code{ageMax} is the hard upper bound, treated as interval. If you want ages
20 to 89, then give \code{ageMin = 20} and \code{ageMax = 90}, not 89.
}
