% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsAge.R
\name{names2age}
\alias{names2age}
\title{detect ages from names of vector(s)}
\usage{
names2age(...)
}
\arguments{
\item{...}{one or more vectors of any class, which has/have a names attribute.}
}
\value{
integer vector of ages (presumably).
}
\description{
Often as a shorthand we pull lower age bounds from the names of a vector. This modulates that action, and allows for giving several vectors to check for names.
}
\details{
If more than one vector is given, names are pulled from the first available named vector. All vectors must be of the same length. If no names are available on any vector, then \code{NA}s are returned. This clearly won't work if the names on the vector are of something else.
}
