% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPAG.R
\name{OPAG_fit_stable_standard}
\alias{OPAG_fit_stable_standard}
\title{creates stable standard based on optimizing the growth rate}
\usage{
OPAG_fit_stable_standard(Pop_fit, Age_fit, AgeInt_fit, Lx1, Age_Lx1)
}
\arguments{
\item{Pop_fit}{numeric vector of at least two population counts to use for fitting}

\item{Age_fit}{integer vector of lower bounds for age groups of \code{Pop_fit}}

\item{AgeInt_fit}{integer vector of widths of age groups of \code{Pop_fit}}

\item{Lx1}{numeric vector of stable population standard by single ages}

\item{Age_Lx1}{integer vector of lower bounds for age groups of \code{Lx1}}
}
\value{
list constaining
\enumerate{
\item \code{Standard} numeric vector, the transformed \code{nLx} to be used for
redistribution in \code{OPAG()}
\item r_opt the output of \code{optimize()}, where \code{min} is the growth parameter, \code{r}
}
}
\description{
The stationary standard, \code{nLx} is transformed into a stable standard by optimizing a growth rate, \code{r} such that the stable standard matches observed population counts in selected age groups. Usually the ages used for fitting are wide age groups in older ages preceding the open age group. The standard output by this function is used by \code{OPAG} to create the standard used to redistribute counts over older age groups up to a specified open age group, such as 100.
}
\details{
The argument \code{method} don't have much leverage on the result. In short, the stable population transformation is done by ungrouping \code{nLx} to single ages (if it isn't already), and \code{method} controls which graduation method is used for this, where \code{"uniform"}, \code{"mono"}, \code{"pclm"} are the reasonable choices at this writing.
}
\examples{
Pop_fit    <- c(85000,37000)
Age_fit    <- c(70,80)
nLx        <- downloadnLx(NULL, "Spain","female",1971)
Age_nLx    <- names2age(nLx)
Lx1 <- graduate(nLx,Age=Age_nLx,method = "mono")
Age_Lx1 <- 0:100
# India Males, 1971
Pop        <- smooth_age_5(pop1m_ind,
                           Age = 0:100,
                           method = "Arriaga")
Pop80      <- groupOAG(Pop, names2age(Pop), 80)
Age        <- names2age(Pop80)

nLx        <- downloadnLx(NULL, "India","male",1971)
Age_nLx    <- names2age(nLx)


Pop_fit    <- groupAges(Pop80, Age, N = 10)[c("60","70")]
Age_fit    <- c(60,70)
AgeInt_fit <- c(10,10)

Standard <- OPAG_fit_stable_standard(
  Pop_fit = Pop_fit,
  Age_fit = Age_fit,
  AgeInt_fit = AgeInt_fit,
  Lx1=Lx1,
  Age_Lx1 = Age_Lx1
)

# A visual comparison:
nL60 <- rescale_vector(nLx[Age_nLx >= 60])
St60p <- rescale_vector( Standard$Standard[c(0:100) >= 60] )
ages_plot <- seq(60,100,by=5)
\dontrun{
  plot(ages_plot,nL60, type = 'l')
  lines(60:100, St60p, col = "blue")
}
}
