% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPAG.R
\name{OPAG_r_min}
\alias{OPAG_r_min}
\title{calculates residual for optimizing growth rate r for OPAG family}
\usage{
OPAG_r_min(r, Age_fit, Pop_fit, AgeInt_fit, Lx1, Age_Lx1)
}
\arguments{
\item{r}{given stable growth rate}

\item{Age_fit}{integer vector of lower bounds for age groups of \code{Pop_fit}}

\item{Pop_fit}{numeric vector of at least two population counts to use for fitting}

\item{AgeInt_fit}{integer vector of widths of age groups of \code{Pop_fit}}

\item{Lx1}{numeric vector of stable population standard by single ages}

\item{Age_Lx1}{integer vector of lower bounds for age groups of \code{Lx1}}
}
\value{
numeric. A residual that you're presumably trying to minimize.
}
\description{
For a given set of age groups to fit against, and a given stable growth rate, $r$,
what is the error implied given the current $r$ and stationary standard?
}
\details{
This is a utility function for \code{OPAG()}, which needs to optimize $r$ for a
given population vector and stationary standard.
}
\examples{
# Make up some population data to fit to:
Pop_fit    <- c(85000,37000)
Age_fit    <- c(70,80)
AgeInt_fit <- c(10,10)
nLx        <- downloadnLx(NULL, "Spain","female",1971)
# graduate(nLx, Age_nLx, method = method, constrain = TRUE)
Ageab    <- names2age(nLx)
Lx1 <- graduate(c(nLx), Ageab, method = "mono", constrain = TRUE)
Age_Lx1 <- 0:100
r          <- .01

OPAG_r_min(r,
           Pop_fit = Pop_fit,
           Age_fit = Age_fit,
           AgeInt_fit = AgeInt_fit,
           Lx1 = Lx1,
           Age_Lx1 = Age_Lx1)

(r_opt <- optimize(OPAG_r_min,
         Pop_fit = Pop_fit,
         Age_fit = Age_fit,
         AgeInt_fit = AgeInt_fit,
         Lx1 = Lx1,
         Age_Lx1 = Age_Lx1,
         interval = c(-0.05,.05))$min)

}
