% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop1m_ind}
\alias{pop1m_ind}
\title{#' Example data [this is the short explanation to appera in the list of fucntions]
#'
#' Some more explanation of the data [this will appear on the help page]
#'
#' @format
#'   A data frame with X rows and Y variables:
#'   \describe{
#'     \item{rank}{Rank of the country in a given year.}
#'     \item{country}{Country.}
#'   }
#'
#' @source
#'   The data comes from
#'   \url{https://URL}
"data_object"
Indian male population 1991}
\format{A numeric vector of length 101}
\source{
The data comes from
  \url{https://URL}
}
\usage{
pop1m_ind
}
\description{
Indian male population 1991
}
\keyword{datasets}
