% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{qx2lx}
\alias{qx2lx}
\title{Derive lifetable survivorship (lx) from death probabilities.}
\usage{
qx2lx(nqx, radix = 1e+05)
}
\arguments{
\item{nqx}{numeric. Vector of age specific death probabilities.}

\item{radix}{numeric. The lifetable starting population. Default 100000.}
}
\value{
lx vector of lifetable survivorship.
}
\description{
This lifetable identity is the same no matter what kind of lifetable is required.
 You can find it in any demography textbook.
}
\details{
set \code{radix = 1} for the probability of surviving until age x. The vector returned is
the same length as \code{nqx}, thereby throwing out the final value of qx, which is usually set to 1.
}
\references{
\insertRef{preston2000demography}{DemoTools}
}
