% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{qxmx2ax}
\alias{qxmx2ax}
\title{Derive nax from nqx and nMx.}
\usage{
qxmx2ax(nqx, nMx, AgeInt)
}
\arguments{
\item{nqx}{numeric. Vector of age specific death probabilities.}

\item{nMx}{numeric. Vector of age-specific death rates.}

\item{AgeInt}{integer. Vector of age class widths.}
}
\value{
nax numeric vector of average time spent in interval by those dying in interval via identity.
}
\description{
This is the standard identity to derive nax from nqx and nMx.
}
\references{
\insertRef{preston2000demography}{DemoTools}
}
