% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTidentities.R
\name{qxmx2ax}
\alias{qxmx2ax}
\title{derive nax from nqx and nMx}
\usage{
qxmx2ax(nqx, nMx, AgeInt)
}
\arguments{
\item{nMx}{numeric vector. Age specific death rates.}

\item{AgeInt}{integer width of each age interval}

\item{nax}{numeric vector. Average time spent in interval by those dying in interval}
}
\value{
nax numeric vector of average time spent in interval by those dying in interval
}
\description{
This is the common ax-qx-mx identity as found in any demography text.
}
