% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale.vector}
\alias{rescale.vector}
\title{Rescale a vector proportionally to a new sum.}
\usage{
rescale.vector(x, scale = 1)
}
\arguments{
\item{x}{numeric vector.}

\item{scale}{numeric. Value the vector should sum to. Default 1.}
}
\value{
The vector rescaled.
}
\description{
Function used to rescale a vector to a given value. This is a frequently needed operation.
}
\details{
For a distribution, use \code{scale = 1}. For percentages, use \code{scale = 100}, etc.
}
\examples{
x <- runif(10)
sum(x)
xx <- rescale.vector(x,100)
sum(xx)
}
