% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rlog}
\alias{rlog}
\title{Logging that does not cause jams.}
\usage{
rlog(x, base = exp(1))
}
\arguments{
\item{x}{numeric or complex vector.}

\item{base}{positive or complex number. The base with respect to which
logarithms are computed.  Defaults to \code{e=exp(1)}.}
}
\description{
Zeros happen. Logging zeros leads to errors in downstream code.
For the sake of robustness, we replace zeros with a small positive number to be 
able to continue with calculations.
}
\details{
Use judiciously, since errors are good too.
}
