% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprague.R
\name{splitMono}
\alias{splitMono}
\title{split age groups using a monotonic spline}
\usage{
splitMono(Value, Age5 = seq(0, length(Value) * 5 - 5, 5), keep.OAG = FALSE)
}
\arguments{
\item{Value}{numeric vector of counts in age groups}

\item{Age5}{integer vector of lower bound of age groups}

\item{keep.OAG}{logical (default \code{FALSE}). Would we like to re-impute the last 
element of \code{Value} as the open age group?}
}
\value{
numeric vector of single age counts
}
\description{
Take the cumulative sum of \code{Value} and then run a monotonic spline through it. The first 
differences split back single-age estimates of \code{Value}. Optionally keep the open age group untouched.
}
\details{
We use the \code{"monoH.FC"} method of \code{stats::splinefun()} to fit the spline because 1)
it passes exactly through the points, 2) it is monotonic and therefore guarantees positive counts, and 3) 
it seems to be a bit less wiggly (lower average first differences of split counts) than a pchip tends to do, 
at least in the tested data.
}
\examples{
Value <- structure(c(88623, 90842, 93439, 96325, 99281, 102051, 104351, 
			 106555, 109170, 112188, 113582, 112614, 108904, 102622, 95867, 
			 80874, 60196, 37523, 17927, 5642, 1110), .Names = c("0", "5", 
			 "10", "15", "20", "25", "30", "35", "40", "45", "50", "55", "60", 
			 "65", "70", "75", "80", "85", "90", "95", "100"))

splitMono(Value)
splitMono(Value, keep.OAG = TRUE)
}
\references{
\insertRef{fritsch1980monotone}{DemoTools}
}
