% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprague.R
\name{splitMono}
\alias{splitMono}
\title{split age groups using a monotonic spline}
\usage{
splitMono(Value, Age5 = seq(0, length(Value) * 5 - 5, 5), keep.OAG = FALSE)
}
\arguments{
\item{Value}{numeric vector of counts in age groups}

\item{Age5}{integer vector of lower bound of age groups}

\item{keep.OAG}{logical (default \code{FALSE}). Would we like to re-impute the last 
element of \code{Value} as the open age group?}
}
\value{
numeric vector of single age counts
}
\description{
Take the cumulative sum of \code{Value} and then run a monotonic spline through it. The first 
differences split back single-age estimates of \code{Value}. Optionally keep the open age group untouched.
}
\details{
We use the \code{"monoH.FC"} method of \code{stats::splinefun()} to fit the spline because 1)
it passes exactly through the points, 2) it is monotonic and therefore guarantees positive counts, and 3) 
it seems to be a bit less wiggly (lower average first differences of split counts) than a pchip tends to do, 
at least in the tested data.
}
\examples{
Value <- structure(c(88623.0176512, 90841.8228447, 93438.8052066, 96324.9863902, 
				99281.3083695, 102050.512557, 104351.4333985, 106554.7539415, 
				109169.72444, 112188.3656672, 113582.4293037, 112613.9259593, 
				108903.7255528, 102622.1778867, 95866.994701, 80874.3840953, 
				60195.7933955, 37523.1859903, 17927.2905862, 5642.0540798, 1110.2324251
		), .Names = c("0", "5", "10", "15", "20", "25", "30", "35", "40", 
				"45", "50", "55", "60", "65", "70", "75", "80", "85", "90", "95", 
				"100"))

splitMono(Value)
splitMono(Value, keep.OAG = TRUE)
}
