% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGEHEAP.R
\name{Spoorenberg}
\alias{Spoorenberg}
\title{Spoorenberg's total modified Whipple index}
\usage{
Spoorenberg(Value, Age, ageMin = 20, ageMax = 64)
}
\arguments{
\item{Value}{numeric. A vector of demographic counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{integer. The lowest age included in calculations. Default 20.}

\item{ageMax}{integer. The upper age bound used for calculations. Default 64.}
}
\value{
The value of the index.
}
\description{
Using the digit-specific modified Whipple's index, this index summarizes
all age preference and avoidance effects by taking the sum of the absolute differences between digit-specific Whipple's index and 1 (counting all
differences as positive).
}
\details{
\code{ageMin} and \code{ageMax} are applied to numerator ages, not denominators.
 Denominators are always 5-year age groups centered on the digit in question,
and these therefore stretch into ages a bit higher or lower than the numerator ages. \code{ageMax} is an inclusive upper bound, treated as interval.
If you want ages 20 to 64, then give \code{ageMin = 20} and \code{ageMax = 64}, not 65.
}
\examples{
Age <- 0:99
Spoorenberg(pop1m_pasex, Age)
}
\references{
\insertRef{spoorenberg2007quality}{DemoTools}
}
