% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprague.R
\name{spragueSimple}
\alias{spragueSimple}
\title{the basic Sprague age-splitting method}
\usage{
spragueSimple(popmat)
}
\arguments{
\item{popmat}{a numeric matrix of population counts in 5-year age groups, with integer-labeled 
margins (age in rows and year in columns).}
}
\value{
an age-period matrix od split population counts with the same number of 
columns as \code{popmat}, and single ages in rows.
}
\description{
This method is based on the first stage of the Sprague R 
script prepared by Thomas Buettner and Patrick Gerland, itself based on the description
in Siegel and Swanson, 2004, p. 727.
}
\details{
Ages should refer to lower age bounds, ending in the open age group in the last row (not a closed terminal age). 
Dimension labelling is necessary. There must be at least six age groups (including the open group). One year of data will 
work as well, as long as it's given as a single-column matrix.
}
\examples{
p5 <- structure(c(54170.08, 44774.6, 42141.587, 38463.515, 34405.607, 
162369.816, 57424.3568738, 44475.4981681, 41751.7574114, 39628.4338929, 
34756.9473002, 164194.0485702, 60272.2061248, 44780.1982856, 
41803.6541424, 40229.0292664, 35154.7682192, 166275.9022992, 
62726.896388, 45681.1355532, 42100.72506, 40473.8600572, 35598.545404, 
168556.5331816, 64815.5458002, 47136.5341033, 42508.3026466, 
40532.3096745, 36082.7490698, 170990.473735, 66579.122, 49070.407, 
42953.604, 40534.586, 36596.844, 173545.633), .Dim = c(6L, 6L
), .Dimnames = list(seq(0,25,5), 1950:1955))
head(p5) # this is the entire matrix
p1 <- spragueSimple(p5)
head(p1); tail(p1)
colSums(p1) - colSums(p5) 

# another case, starting with single ages
# note \\spragueSimple() does not group ages. You need to do it 
# first.
Value <- c(9544406,7471790,11590109,11881844,11872503,12968350,11993151,10033918,
		14312222,8111523,15311047,6861510,13305117,7454575,9015381,10325432,
		9055588,5519173,12546779,4784102,13365429,4630254,9595545,4727963,
		5195032,15061479,5467392,4011539,8033850,1972327,17396266,1647397,
		6539557,2233521,2101024,16768198,3211834,1923169,4472854,
		1182245,15874081,1017752,3673865,1247304,1029243,12619050,1499847,
		1250321,2862148,723195,12396632,733501,2186678,777379,810700,
		7298270,1116032,650402,1465209,411834,9478824,429296,1190060,
		446290,362767,4998209,388753,334629,593906,178133,
		4560342,179460,481230,159087,155831,1606147,166763,93569,182238,
		53567,1715697,127486,150782,52332,48664,456387,46978,34448,
		44015,19172,329149,48004,28574,9200,7003,75195,13140,5889,
		18915,21221,72373)
Age         <- 0:100
# group ages
Val5        <- groupAges(Value, Age)
# name the vector (or dims of matrix if you end up
# producing a matrix)
names(Val5) <- Age[Age \%\% 5 == 0]
# notice how this particular case produces a negative value in the last age
# before OAG:
(pops <- spragueSimple(Val5))
# this replaces ages 90+, guaranteed no negatives.
spragueCloseout(Val5, pops = pops)
# Note: there are no kludges built into spragueSimple() to handle such cases.
# these ought to be handled by wrappers as appropriate.
}
\references{
Shryock, H. S., Siegel, J. S., & Larmon, E. A. (1973). 
The methods and materials of demography. US Bureau of the Census.

Seigel, J. S., & Swanson, D. A. (2004). T
he methods and materials of demography. Elsevier Academic Press, London.
}
