% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Whipple.R
\name{Whipple}
\alias{Whipple}
\title{calculate Whipple's index of age heaping}
\usage{
Whipple(Value, Age, ageMin = 25, ageMax = 65, fives = TRUE)
}
\arguments{
\item{Value}{a vector of demographic counts by single age}

\item{Age}{a vector of ages corresponding to the lower integer bound of the counts}

\item{ageMin}{the lowest age included in calcs. Default 25}

\item{ageMax}{the upper age bound used for calcs. Default 65}

\item{fives}{logical default \code{TRUE}. Do we count terminal digit 0's and 5' (\code{TRUE}) or just terminal digit 0's (\code{FALSE})?}
}
\description{
Implementation following the PASEX spreadsheet SINGAGE
}
\details{
\code{ageMax} is the hard upper bound, treated as interval. If you want ages
20 to 89, then give \code{ageMin = 20} and \code{ageMax = 90}, not 89.
}
