% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGEHEAP.R
\name{Whipple}
\alias{Whipple}
\title{Calculate Whipple's index of age heaping}
\usage{
Whipple(Value, Age, ageMin = 25, ageMax = 65, digit = c(0, 5))
}
\arguments{
\item{Value}{numeric. A vector of demographic counts by single age.}

\item{Age}{numeric. A vector of ages corresponding to the lower integer bound of the counts.}

\item{ageMin}{integer. The lowest age included in calculations. Default 25.}

\item{ageMax}{integer. The upper age bound used for calculations. Default 65.}

\item{digit}{integer. Any digit between 0 and 9. Default \code{c(0,5)}. Otherwise it needs to be a single digit.}
}
\value{
The value of the index.
}
\description{
Implementation following the PASEX spreadsheet SINGAGE, with some extra options for more digit checking. 
Whipple's index is a summary measure of age heaping on ages ending usually in the digits 0 or 5 used to determine variability in the
quality of age reporting between regions or countries and its evolution over time. It assumes a
linear distribution of ages in each five-year age range. Digits refer to terminal age digits. For example, 9, 19, 29, etc are all of digit 9.
}
\details{
\code{ageMin} and \code{ageMax} refer to the bounds of the numerator, where \code{ageMax} is inclusive. 
The denominator looks 7 ages lower and 2 ages higher, so these ages must be available. You can get 
arbitrary W(i) indices by specifying other digits. Note you can only do pairs of digits
they are 0 and 5. Otherwise just one digit at a time.
}
\examples{
 Value <- c(80626,95823,104315,115813,100796,105086,97266,116328,
 		75984,89982,95525,56973,78767,65672,53438,85014,
 		47600,64363,42195,42262,73221,30080,34391,29072,
 		20531,66171,24029,44227,24128,23599,82088,16454,
 		22628,17108,12531,57325,17220,28425,16206,17532,
 		65976,11593,15828,13541,8133,44696,11165,18543,
 		12614,12041,55798,9324,10772,10453,6773,28358,
 		9916,13348,8039,7583,42470,5288,5317,6582,
 		3361,17949,3650,5873,3279,3336,27368,1965,
 		2495,2319,1335,12022,1401,1668,1360,1185,
 		9167,424,568,462,282,6206,343,409,333,291,4137,133,169,157,89,2068,68,81,66,57)
 Age <- 0:99
 
 (w05 <- Whipple(Value, Age, 25, 60, digit = c(0,5))) # 2.34,replicates SINGAGE males
 
 # implements formula from Roger et al. (1981, p. 148)
 (w0 <- Whipple(Value, Age, 25, 60, digit = 0))
 (w5 <- Whipple(Value, Age, 25, 60, digit = 5)) 
 
 # Whipple types
 Whipple(Value, Age, 25, 60, digit = 3) 
}
\references{
\insertRef{spoorenberg2007quality}{DemoTools}
\insertRef{GDA1981IREDA}{DemoTools}
\insertRef{PAS}{DemoTools}
}
