% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ypart}
\alias{ypart}
\title{Determine the proportion of a year passed as of a particular date.}
\usage{
ypart(Year, Month, Day)
}
\arguments{
\item{Year}{string or integer. 4-digit year.}

\item{Month}{string or integer. Month digits, 1 or 2 characters.}

\item{Day}{string or integer. Day of month digits, 1 or 2 characters.}
}
\description{
The fraction returned by this is used e.g. for intercensal estimates.
}
\examples{
ypart(2001,2,14) # general use
ypart(2001,6,30) # mid year options, default detection
ypart(2001,7,1)  # also
ypart(2000,6,30) # no change for leap year, still detected as mid year
ypart(2000,7,1)  # July 1 leap year
ypart(2001,7,1)  # July 1 not leap year
ypart(2002,12,31) # assumes end of day by default.
ypart(2002,1,1) # end of day year fraction
}
