% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGZAG.R
\name{zigzag_p}
\alias{zigzag_p}
\title{Smooth population counts using Feeney's zigzag method and smoothing parameters.}
\usage{
zigzag_p(Value, Age, ageMin = 40, ageMax = 80, p = rep(0.05,
  floor((ageMax - ageMin)/10)))
}
\arguments{
\item{Value}{numeric vector of (presumably) counts in 5-year age groups.}

\item{Age}{integer vector of age group lower bounds.}

\item{ageMin}{integer. Lower age bound to adjust values.}

\item{ageMax}{integer. Upper age bound to adjust values.}

\item{p}{numeric vector of adjustment parameters.}
}
\value{
numeric vector of smoothed population counts.
}
\description{
This function is auxiliary to \code{zigzag()}, see \code{?zigzag} for a description.
}
\details{
This function is not intended to be used at the top level, but just in case, make sure that \code{ageMax = ageMin + 10 * length(p)}. Age groups \code{ >= ageMin} AND \code{ <= ageMin} must be in 5-year age groups. This function does no checks.
}
\examples{
Age <- c(0,1,seq(5,90,by=5))
zigzag_min(dth5_zigzag,Age,ageMin=40,ageMax = 80,p=rep(.05,4))
# it's used like this in zigzag()
(p <- optim(
		rep(.05,4), 
		zigzag_min, 
		Value = dth5_zigzag, 
		Age = Age, 
		ageMin = 40, 
		ageMax = 80)$par)
Smoothed <- zigzag_p(dth5_zigzag,Age,40,80,p)
}
\references{
Feeney, G. 2013 "Removing "Zigzag" from Age Data," http://demographer.com/white-papers/2013-removing-zigzag-from-age-data/
}
